Normal climatology values

Get normal climatology values for a station (or all the stations with aemet_normal_clim_all(). Standard climatology from 1981 to 2010.

Usage

aemet_normal_clim(
  station = NULL,
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)

aemet_normal_clim_all(
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)

Arguments

station

Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf

Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

extract_metadata

Logical TRUE/FALSE. On TRUE the output is a tibble with the description of the fields. See also get_metadata_aemet().

progress

Logical, display a cli::cli_progress_bar() object. If verbose = TRUE won't be displayed.

Value

A tibble or a sf object.

Note

Code modified from project https://github.com/SevillaR/aemet.

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_normal_clim(c("9434", "3195"))
glimpse(obs)
#> Rows: 26
#> Columns: 475
#> $ indicativo  <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "9…
#> $ w_racha_max <dbl> 28.9, 28.9, 30.3, 26.7, 28.3, 30.8, 37.5, 27.5, 25.8, 23.3…
#> $ np_010_n    <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ np_010_s    <dbl> 3.32, 2.31, 2.79, 2.43, 3.02, 2.48, 1.61, 1.54, 1.79, 3.11…
#> $ q_max_s     <dbl> 4.75, 3.69, 3.71, 3.08, 2.69, 1.88, 2.09, 1.55, 2.11, 2.49…
#> $ n_tor_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 29, 30, 29, 30, 30…
#> $ n_tor_s     <dbl> 0.00, 0.25, 0.82, 1.53, 2.62, 2.33, 2.10, 2.20, 1.73, 0.99…
#> $ q_max_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ tm_min_q4   <dbl> 4.2, 4.5, 6.7, 9.1, 13.5, 17.2, 19.8, 19.5, 16.5, 12.4, 7.…
#> $ tm_min_q1   <dbl> 2.0, 2.6, 5.4, 7.7, 11.2, 15.5, 17.7, 18.1, 14.5, 10.3, 5.…
#> $ tm_min_q3   <dbl> 3.2, 4.1, 6.2, 8.7, 12.6, 16.3, 19.1, 19.0, 15.6, 11.6, 6.…
#> $ tm_min_q2   <dbl> 2.6, 3.0, 5.6, 8.3, 12.1, 15.9, 18.0, 18.7, 15.1, 10.9, 6.…
#> $ q_mar_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ n_des_min   <dbl> 0, 0, 0, 1, 0, 2, 9, 5, 3, 1, 0, 0, 55, 0, 2, 0, 0, 1, 1, …
#> $ q_mar_s     <dbl> 5.00, 4.79, 4.51, 2.57, 1.74, 1.58, 1.11, 1.19, 1.94, 2.47…
#> $ q_med_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ ts_20_q1    <dbl> 6.4, 7.7, 12.1, NA, NA, 27.5, NA, 28.7, 24.4, 17.6, NA, NA…
#> $ ts_20_q2    <dbl> 7.3, 8.4, 13.1, NA, NA, 28.4, NA, 30.7, 25.3, 18.3, NA, NA…
#> $ e_cv        <dbl> 0.11, 0.13, 0.10, 0.11, 0.09, 0.08, 0.08, 0.10, 0.10, 0.08…
#> $ ts_20_q4    <dbl> 8.1, 9.7, 14.9, NA, NA, 30.6, NA, 32.0, 26.8, 20.7, NA, NA…
#> $ e_min       <dbl> 62, 48, 70, 75, 95, 115, 132, 132, 119, 110, 69, 60, 105, …
#> $ np_300_q4   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 0, 0, 0, 0, 0, 0, 0…
#> $ np_300_q1   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ np_300_q3   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0…
#> $ hr_max      <dbl> 83, 78, 70, 65, 64, 57, 57, 57, 65, 77, 84, 93, 66, 81, 73…
#> $ np_300_cv   <dbl> 5.48, NA, 3.81, 2.59, 3.05, 2.59, 5.48, 3.05, 2.59, 3.05, …
#> $ n_nub_max   <dbl> 24, 22, 26, 24, 28, 25, 21, 25, 25, 27, 26, 24, 255, 20, 2…
#> $ tm_min_cv   <dbl> 0.50, 0.38, 0.17, 0.14, 0.10, 0.08, 0.06, 0.05, 0.08, 0.11…
#> $ n_des_mn    <dbl> 5.0, 5.5, 4.5, 4.5, 4.5, 7.0, 14.0, 12.0, 7.0, 5.5, 3.0, 4…
#> $ n_des_md    <dbl> 4.7, 5.9, 6.3, 4.9, 5.1, 8.1, 13.7, 11.2, 7.3, 5.7, 3.7, 4…
#> $ ts_20_s     <dbl> 1.29, 1.39, 1.54, NA, NA, 2.38, NA, 2.11, 1.91, 1.87, NA, …
#> $ q_med_mn    <dbl> 990.1, 988.9, 987.5, 984.4, 984.9, 986.0, 985.7, 985.5, 98…
#> $ evap_cv     <dbl> 0.37, 0.31, 0.23, 0.23, 0.20, 0.21, 0.17, 0.18, 0.22, 0.26…
#> $ q_med_md    <dbl> 990.5, 989.2, 986.8, 984.0, 984.9, 985.6, 985.6, 985.5, 98…
#> $ nt_30_cv    <dbl> NA, NA, NA, 2.14, 0.72, 0.33, 0.20, 0.18, 0.67, 2.02, NA, …
#> $ mes         <chr> "01", "02", "03", "04", "05", "06", "07", "08", "09", "10"…
#> $ ts_20_cv    <dbl> 0.18, 0.16, 0.12, NA, NA, 0.08, NA, 0.07, 0.07, 0.10, NA, …
#> $ inso_max    <dbl> 6.9, 8.7, 10.1, 11.0, 11.1, 12.0, 13.0, 11.8, 10.0, 8.8, 7…
#> $ ts_20_max   <dbl> 10.1, 11.8, 15.6, NA, NA, 32.7, NA, 33.4, 29.3, 23.4, NA, …
#> $ np_001_max  <dbl> 20, 14, 20, 16, 20, 15, 8, 9, 12, 16, 17, 20, 108, 20, 18,…
#> $ n_llu_md    <dbl> 9.3, 7.7, 8.9, 10.1, 10.4, 7.9, 5.1, 5.4, 7.6, 10.0, 10.9,…
#> $ nv_1000_s   <dbl> 2.97, 2.20, 0.46, 0.00, 0.38, 0.00, 0.00, 0.00, 0.31, 0.97…
#> $ ta_min_mn   <dbl> -2.9, -1.6, 0.8, 3.2, 6.8, 11.1, 14.2, 14.2, 9.9, 4.7, -0.…
#> $ tm_mes_max  <dbl> 9.7, 11.5, 14.6, 17.4, 20.8, 26.6, 28.2, 27.9, 24.0, 19.5,…
#> $ ta_max_mn   <dbl> 18.0, 19.8, 24.7, 28.0, 33.0, 37.3, 39.0, 38.4, 34.0, 28.7…
#> $ nv_1000_n   <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 28, 29…
#> $ ta_max_md   <dbl> 17.7, 19.6, 24.7, 28.1, 32.7, 37.3, 39.2, 38.3, 33.6, 28.2…
#> $ nw_91_min   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ ta_max_n    <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ ta_max_s    <dbl> 1.85, 2.16, 1.55, 2.06, 2.27, 2.44, 2.10, 1.84, 2.34, 2.36…
#> $ ts_10_md    <dbl> 7.2, 9.1, 14.0, NA, NA, 30.4, NA, 31.8, 26.3, 19.1, NA, NA…
#> $ nw_55_mn    <dbl> 7.0, 9.0, 8.0, 7.0, 6.0, 7.0, 6.5, 5.5, 4.0, 4.0, 6.0, 6.0…
#> $ ta_min_s    <dbl> 2.07, 1.89, 2.10, 1.85, 1.96, 1.58, 1.78, 1.40, 1.80, 2.35…
#> $ nw_55_md    <dbl> 8.0, 9.2, 8.9, 8.2, 7.6, 6.6, 7.0, 5.8, 4.1, 4.8, 6.6, 5.7…
#> $ np_300_mn   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0…
#> $ evap_q4     <dbl> 1092, 1527, 2115, 2363, 3008, 3599, 4248, 3926, 2787, 2043…
#> $ evap_q1     <dbl> 628, 843, 1463, 1684, 2153, 2852, 3331, 2992, 1902, 1276, …
#> $ np_300_md   <dbl> 0.0, 0.0, 0.1, 0.1, 0.1, 0.1, 0.0, 0.1, 0.1, 0.1, 0.2, 0.0…
#> $ evap_q3     <dbl> 993, 1307, 1931, 2068, 2751, 3281, 3828, 3714, 2569, 1741,…
#> $ evap_q2     <dbl> 797, 1186, 1750, 1923, 2424, 3016, 3695, 3281, 2345, 1504,…
#> $ p_mes_min   <dbl> 0.5, 0.0, 0.0, 3.6, 4.3, 0.0, 0.3, 0.9, 1.4, 0.6, 0.4, 0.6…
#> $ ts_min_min  <dbl> 5.2, 5.5, 8.5, 10.5, 14.2, 18.7, 20.4, 20.8, 17.1, 13.5, 9…
#> $ n_des_s     <dbl> 3.32, 3.30, 5.08, 3.27, 2.97, 3.96, 2.91, 3.14, 2.28, 3.86…
#> $ nv_0100_mn  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0…
#> $ p_mes_md    <dbl> 23.7, 19.8, 28.0, 40.0, 40.2, 28.5, 16.5, 17.8, 27.3, 34.0…
#> $ n_tor_cv    <dbl> NA, 3.81, 1.44, 0.90, 0.59, 0.50, 0.54, 0.56, 0.58, 0.90, …
#> $ nv_0100_md  <dbl> 0.2, 0.2, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.1, 0.2, 0.4…
#> $ q_min_max   <dbl> 986.3, 988.6, 980.9, 976.6, 978.7, 979.0, 979.7, 979.0, 98…
#> $ n_des_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 27, 28…
#> $ p_sol_s     <dbl> 11.84, 10.09, 10.27, 7.68, 6.32, 5.93, 5.04, 4.49, 7.27, 9…
#> $ ts_20_n     <dbl> 15, 15, 15, 14, 14, 15, 14, 16, 15, 15, 12, 14, 11, 0, 0, …
#> $ ts_10_cv    <dbl> 0.17, 0.17, 0.10, NA, NA, 0.08, NA, 0.07, 0.07, 0.09, NA, …
#> $ nw_91_s     <dbl> 0.42, 0.51, 0.48, 0.26, 0.19, 0.33, 0.48, 0.19, 0.20, 0.00…
#> $ nw_91_n     <dbl> 28, 29, 30, 28, 27, 25, 28, 28, 26, 28, 28, 28, 17, 27, 27…
#> $ nt_00_min   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 3, 0, 0, 0, 0, 0, 0, 0…
#> $ p_sol_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 29, 29, 30, 30, 30, 29, 13, 14…
#> $ n_nie_mn    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0…
#> $ n_nub_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 27, 28…
#> $ n_nie_md    <dbl> 0.5, 0.6, 0.3, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.1, 0.4…
#> $ tm_mes_md   <dbl> 7.0, 8.5, 11.8, 14.4, 18.6, 23.1, 25.7, 25.6, 21.4, 16.6, …
#> $ p_sol_cv    <dbl> 0.25, 0.17, 0.17, 0.12, 0.10, 0.08, 0.06, 0.06, 0.11, 0.16…
#> $ n_nub_s     <dbl> 3.58, 2.88, 3.58, 3.41, 3.25, 3.53, 2.85, 2.88, 2.43, 3.08…
#> $ w_racha_min <dbl> 15.8, 16.9, 15.0, 16.4, 13.6, 12.5, 15.0, 14.4, 16.1, 12.2…
#> $ np_300_max  <dbl> 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 4, 0, 0, 1, 0, 2, 0, 0…
#> $ nv_0050_md  <dbl> 0.0, 0.1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.1…
#> $ ta_min_min  <dbl> -5.8, -7.2, -6.0, -0.8, 1.7, 8.0, 10.8, 10.8, 7.0, 1.0, -5…
#> $ glo_mn      <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ inso_s      <dbl> 1.12, 1.08, 1.21, 1.02, 0.92, 0.91, 0.77, 0.61, 0.90, 1.07…
#> $ glo_md      <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ n_nie_q1    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ w_med_q4    <dbl> 19.2, 22.0, 19.2, 19.0, 20.2, 19.0, 19.2, 18.0, 17.0, 16.2…
#> $ n_tor_max   <dbl> 0, 1, 3, 6, 10, 9, 10, 9, 8, 4, 2, 2, 41, 1, 1, 3, 4, 11, …
#> $ w_med_q2    <dbl> 14.0, 16.0, 16.6, 18.0, 16.6, 16.6, 18.0, 15.0, 14.0, 13.0…
#> $ w_med_q3    <dbl> 17.0, 18.4, 17.4, 18.4, 18.4, 18.0, 19.0, 17.4, 15.0, 14.0…
#> $ n_gra_md    <dbl> 0.1, 0.0, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.0, 0.0, 0.0…
#> $ inso_q4     <dbl> 5.7, 7.2, 8.2, 9.0, 10.1, 11.4, 12.4, 11.1, 9.4, 7.5, 5.8,…
#> $ q_max_min   <dbl> 994.7, 994.4, 988.8, 989.4, 990.8, 989.5, 991.2, 989.8, 99…
#> $ ts_10_max   <dbl> 8.9, 12.7, 16.3, NA, NA, 34.6, NA, 35.4, 29.7, 21.8, NA, N…
#> $ nt_30_min   <dbl> 0, 0, 0, 0, 0, 3, 14, 12, 0, 0, 0, 0, 49, 0, 0, 0, 0, 0, 1…
#> $ n_nub_cv    <dbl> 0.20, 0.16, 0.19, 0.17, 0.15, 0.18, 0.17, 0.16, 0.12, 0.15…
#> $ p_mes_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ p_mes_q1    <dbl> 7.2, 4.7, 9.2, 22.2, 16.3, 7.4, 3.6, 3.3, 7.5, 11.8, 15.7,…
#> $ p_mes_q2    <dbl> 14.6, 14.2, 15.5, 32.4, 31.3, 13.3, 6.8, 8.1, 14.7, 21.0, …
#> $ p_mes_q3    <dbl> 20.2, 21.8, 29.4, 41.4, 34.8, 30.4, 16.3, 19.3, 23.2, 32.0…
#> $ p_mes_q4    <dbl> 36.0, 29.1, 49.4, 49.4, 62.6, 47.4, 27.9, 29.4, 42.1, 49.8…
#> $ hr_cv       <dbl> 0.07, 0.09, 0.09, 0.08, 0.09, 0.09, 0.08, 0.09, 0.07, 0.08…
#> $ n_tor_md    <dbl> 0.0, 0.1, 0.6, 1.7, 4.5, 4.6, 3.9, 3.9, 3.0, 1.1, 0.2, 0.1…
#> $ p_mes_s     <dbl> 21.10, 17.29, 21.13, 27.75, 29.51, 25.48, 15.10, 17.29, 26…
#> $ nw_91_cv    <dbl> 1.95, 2.12, 2.42, 3.67, 5.20, 2.76, 2.66, 5.29, 5.10, NA, …
#> $ n_tor_mn    <dbl> 0.0, 0.0, 0.0, 1.0, 4.0, 5.0, 3.0, 4.0, 3.0, 1.0, 0.0, 0.0…
#> $ nv_1000_max <dbl> 12, 9, 2, 0, 1, 0, 0, 0, 1, 4, 9, 13, 22, 6, 4, 2, 0, 0, 0…
#> $ n_gra_cv    <dbl> 3.74, NA, 3.26, 3.81, 3.26, 2.59, 2.59, 3.26, 2.59, NA, 5.…
#> $ w_racha_n   <dbl> 28, 29, 30, 28, 27, 26, 28, 28, 26, 28, 29, 28, 19, 27, 27…
#> $ w_med_min   <dbl> 9, 8, 11, 9, 13, 13, 12, 9, 11, 9, 9, 6, 14, 4, 5, 6, 5, 7…
#> $ w_racha_s   <dbl> 3.47, 3.01, 3.58, 2.67, 2.68, 4.33, 4.55, 3.03, 2.48, 2.89…
#> $ np_100_max  <dbl> 2, 3, 3, 4, 5, 4, 2, 2, 4, 4, 3, 1, 17, 5, 5, 6, 4, 5, 3, …
#> $ e_s         <dbl> 7.98, 9.93, 8.79, 10.04, 10.61, 10.70, 12.98, 15.59, 14.81…
#> $ w_med_s     <dbl> 4.46, 4.83, 2.59, 3.44, 3.56, 2.28, 2.77, 2.89, 2.37, 2.71…
#> $ p_sol_md    <dbl> 48, 60, 61, 62, 65, 71, 79, 77, 69, 60, 52, 44, 63, NA, NA…
#> $ n_cub_q4    <dbl> 11.0, 7.0, 8.2, 8.0, 6.0, 3.0, 1.0, 2.0, 4.0, 7.0, 10.0, 1…
#> $ n_cub_q3    <dbl> 9.4, 5.4, 6.0, 6.0, 5.0, 2.0, 1.0, 1.0, 3.0, 5.4, 8.4, 8.4…
#> $ n_cub_q2    <dbl> 7.0, 3.6, 4.6, 5.0, 4.0, 2.0, 0.0, 1.0, 2.0, 3.0, 6.0, 8.0…
#> $ n_cub_q1    <dbl> 5.0, 2.8, 3.0, 3.8, 3.0, 1.0, 0.0, 0.0, 1.8, 2.0, 3.0, 6.8…
#> $ nw_55_max   <dbl> 20, 19, 19, 15, 19, 12, 14, 13, 8, 11, 20, 16, 120, 7, 7, …
#> $ p_sol_mn    <dbl> 48, 59, 62, 63, 67, 72, 79, 77, 71, 61, 53, 45, 62, NA, NA…
#> $ n_cub_s     <dbl> 3.55, 2.61, 3.09, 2.46, 1.79, 1.55, 0.73, 1.15, 1.76, 3.13…
#> $ hr_mn       <dbl> 75, 65, 60, 56, 53, 47, 46, 47, 54, 65, 73, 77, 60, 72, 66…
#> $ q_med_max   <dbl> 998.9, 996.5, 993.4, 988.9, 988.0, 987.9, 987.7, 988.5, 98…
#> $ hr_md       <dbl> 74, 66, 59, 56, 52, 48, 46, 48, 55, 65, 72, 76, 60, 72, 64…
#> $ np_010_md   <dbl> 4.4, 3.7, 4.8, 5.6, 6.2, 4.0, 2.6, 2.2, 3.2, 5.3, 5.6, 4.5…
#> $ n_cub_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 27, 28…
#> $ np_010_mn   <dbl> 4.0, 4.0, 4.0, 6.0, 6.5, 4.0, 2.0, 2.0, 3.0, 4.5, 5.5, 4.0…
#> $ nw_91_q4    <dbl> 0.6, 0.4, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0…
#> $ p_max_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ nw_91_q2    <dbl> 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0…
#> $ nw_91_q3    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0…
#> $ nw_91_q1    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ n_gra_q4    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 0, 0, 1, 1, 1, 0, 0…
#> $ n_gra_q1    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ n_gra_q3    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0…
#> $ nw_55_s     <dbl> 4.98, 4.59, 4.49, 3.73, 4.52, 2.95, 4.01, 3.69, 1.99, 2.90…
#> $ nw_55_n     <dbl> 28, 29, 30, 28, 27, 25, 28, 28, 26, 28, 28, 28, 17, 27, 27…
#> $ w_med_mn    <dbl> 15.0, 17.5, 17.0, 18.0, 17.0, 18.0, 18.0, 16.0, 14.0, 13.5…
#> $ n_nub_min   <dbl> 11, 9, 11, 10, 15, 10, 10, 13, 14, 13, 10, 8, 203, 9, 9, 7…
#> $ p_max_s     <dbl> 9.06, 7.24, 9.79, 12.39, 11.16, 11.83, 8.61, 12.03, 15.83,…
#> $ ts_20_min   <dbl> 4.9, 6.8, 10.6, NA, NA, 24.1, NA, 26.9, 22.0, 16.9, NA, NA…
#> $ inso_min    <dbl> 2.2, 4.2, 5.0, 6.2, 6.8, 7.6, 9.0, 9.5, 6.8, 4.7, 3.2, 2.2…
#> $ n_gra_max   <dbl> 1, 0, 2, 1, 2, 1, 1, 2, 1, 0, 1, 0, 5, 1, 0, 1, 3, 2, 2, 1…
#> $ p_sol_q1    <dbl> 38, 51, 52, 56, 61, 68, 76, 72, 63, 52, 43, 36, 60, NA, NA…
#> $ p_sol_q3    <dbl> 51, 62, 64, 64, 68, 73, 81, 79, 73, 64, 54, 48, 63, NA, NA…
#> $ p_sol_q2    <dbl> 44, 56, 56, 61, 64, 71, 78, 76, 67, 59, 51, 44, 61, NA, NA…
#> $ p_sol_q4    <dbl> 60, 68, 69, 67, 69, 75, 83, 80, 75, 67, 59, 53, 65, NA, NA…
#> $ n_fog_q1    <dbl> 2.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.6, 4.0…
#> $ ta_min_md   <dbl> -2.9, -2.0, 0.3, 3.0, 6.8, 11.3, 14.2, 14.1, 10.0, 4.8, -0…
#> $ n_fog_q3    <dbl> 6.4, 1.4, 0.4, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 4.0, 8.0…
#> $ n_fog_q2    <dbl> 5.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.6, 2.2, 5.0…
#> $ n_fog_q4    <dbl> 10.2, 5.0, 1.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 2.0, 5.0, 10…
#> $ w_med_cv    <dbl> 0.28, 0.28, 0.15, 0.19, 0.20, 0.13, 0.15, 0.18, 0.16, 0.20…
#> $ p_max_max   <dbl> 41.0, 29.0, 32.5, 57.9, 47.6, 41.1, 35.2, 51.9, 70.8, 49.8…
#> $ e_n         <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ q_min_md    <dbl> 971.5, 971.0, 970.0, 969.8, 973.1, 974.7, 976.6, 976.2, 97…
#> $ n_fog_cv    <dbl> 0.69, 1.29, 1.35, 2.77, 1.99, 3.05, 5.48, 5.48, 2.77, 1.12…
#> $ p_mes_cv    <dbl> 0.89, 0.87, 0.75, 0.69, 0.73, 0.89, 0.92, 0.97, 0.97, 0.79…
#> $ nv_0100_max <dbl> 2, 2, 0, 0, 0, 0, 0, 0, 0, 1, 1, 3, 6, 1, 1, 0, 0, 0, 0, 0…
#> $ np_300_q2   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0…
#> $ q_min_mn    <dbl> 971.7, 970.4, 971.0, 970.5, 973.0, 974.7, 976.6, 976.5, 97…
#> $ ti_max_min  <dbl> -0.6, 0.8, 4.0, 6.6, 10.6, 15.8, 19.6, 21.8, 15.7, 9.4, 3.…
#> $ np_100_cv   <dbl> 1.55, 1.82, 1.55, 0.89, 1.13, 1.26, 1.26, 1.35, 1.28, 1.26…
#> $ hr_q3       <dbl> 75, 67, 61, 56, 53, 48, 46, 48, 55, 68, 75, 78, 61, 74, 68…
#> $ hr_q2       <dbl> 74, 64, 59, 55, 51, 46, 45, 47, 54, 63, 70, 75, 59, 71, 63…
#> $ hr_q1       <dbl> 69, 61, 55, 53, 49, 45, 43, 45, 53, 61, 68, 71, 58, 66, 60…
#> $ nv_0050_s   <dbl> 0.18, 0.43, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.18…
#> $ hr_q4       <dbl> 79, 69, 63, 59, 55, 51, 48, 52, 58, 70, 78, 80, 61, 77, 69…
#> $ evap_s      <dbl> 334.97, 378.66, 407.50, 468.44, 499.63, 675.26, 633.10, 62…
#> $ w_med_q1    <dbl> 12.8, 13.0, 15.8, 16.0, 15.8, 16.0, 15.8, 14.0, 12.0, 11.0…
#> $ evap_n      <dbl> 30, 30, 28, 29, 29, 30, 29, 30, 30, 30, 30, 28, 26, 27, 28…
#> $ nv_0050_n   <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 28, 29…
#> $ nv_0050_q2  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ nv_0050_q3  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ q_max_md    <dbl> 1003.5, 1002.1, 999.5, 996.5, 995.6, 993.9, 994.3, 993.7, …
#> $ nv_0050_q1  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ nv_0050_q4  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0…
#> $ q_max_mn    <dbl> 1003.6, 1003.1, 1000.6, 996.7, 996.4, 994.0, 994.1, 993.3,…
#> $ n_llu_min   <dbl> 0, 0, 0, 0, 3, 3, 2, 1, 3, 2, 3, 1, 65, 0, 1, 0, 5, 1, 1, …
#> $ ts_min_max  <dbl> 13.0, 11.6, 15.3, 15.6, 20.2, 24.7, 23.8, 24.5, 24.0, 20.1…
#> $ evap_max    <dbl> 1860, 2140, 2555, 2978, 3507, 4881, 4766, 4664, 3071, 2446…
#> $ ta_min_q1   <dbl> -4.8, -3.5, -1.1, 1.8, 5.4, 10.4, 12.8, 12.9, 8.3, 2.8, -2…
#> $ ta_min_q3   <dbl> -2.6, -1.3, 1.4, 3.5, 7.1, 11.4, 14.4, 14.3, 10.3, 5.5, 0.…
#> $ n_cub_cv    <dbl> 0.44, 0.56, 0.56, 0.44, 0.39, 0.68, 1.28, 0.98, 0.63, 0.63…
#> $ ta_min_q4   <dbl> -1.0, -0.7, 1.6, 3.9, 8.7, 12.7, 15.6, 15.4, 11.6, 6.5, 1.…
#> $ tm_mes_n    <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ tm_max_s    <dbl> 1.71, 2.12, 1.58, 1.42, 1.78, 2.10, 1.66, 1.50, 1.56, 2.07…
#> $ nt_30_md    <dbl> 0.0, 0.0, 0.0, 0.4, 4.6, 16.0, 23.4, 22.6, 7.9, 0.5, 0.0, …
#> $ tm_max_n    <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ tm_mes_s    <dbl> 1.53, 1.59, 1.14, 1.22, 1.47, 1.65, 1.40, 1.15, 1.32, 1.58…
#> $ nt_30_mn    <dbl> 0.0, 0.0, 0.0, 0.0, 4.5, 16.5, 23.0, 23.5, 8.5, 0.0, 0.0, …
#> $ ts_50_q4    <dbl> 8.3, 9.1, 13.2, NA, NA, 27.6, NA, 30.4, NA, 21.2, NA, NA, …
#> $ ts_50_q3    <dbl> 7.9, 8.7, 12.3, NA, NA, 26.8, NA, 29.2, NA, 19.5, NA, NA, …
#> $ ts_50_q2    <dbl> 7.3, 8.1, 11.6, NA, NA, 26.3, NA, 28.9, NA, 18.7, NA, NA, …
#> $ ts_50_q1    <dbl> 6.7, 7.5, 11.0, NA, NA, 25.9, NA, 27.8, NA, 18.3, NA, NA, …
#> $ q_mar_mn    <dbl> 1021.5, 1020.2, 1018.4, 1014.7, 1014.8, 1015.4, 1014.8, 10…
#> $ np_100_q4   <dbl> 1.0, 1.0, 1.0, 2.0, 2.0, 2.0, 1.0, 1.0, 2.0, 2.0, 2.0, 1.0…
#> $ np_100_q3   <dbl> 0.0, 0.0, 0.0, 1.0, 1.0, 1.0, 1.0, 0.4, 1.0, 1.0, 1.0, 0.0…
#> $ np_100_q2   <dbl> 0.0, 0.0, 0.0, 1.0, 0.6, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0…
#> $ np_100_q1   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 5, 0, 0, 0, 1, 0, 0, 0…
#> $ n_des_cv    <dbl> 0.70, 0.56, 0.81, 0.66, 0.59, 0.49, 0.21, 0.28, 0.31, 0.68…
#> $ q_mar_md    <dbl> 1021.9, 1020.5, 1017.6, 1014.3, 1014.7, 1015.0, 1014.7, 10…
#> $ tm_mes_q4   <dbl> 8.3, 10.0, 12.8, 15.2, 20.3, 24.8, 27.1, 26.4, 22.5, 18.1,…
#> $ tm_mes_q2   <dbl> 6.5, 7.7, 11.2, 14.2, 18.3, 22.7, 25.1, 25.4, 21.0, 16.0, …
#> $ tm_mes_q3   <dbl> 7.3, 9.3, 12.0, 14.6, 18.9, 23.4, 26.4, 25.8, 21.8, 16.8, …
#> $ np_300_min  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ p_sol_min   <dbl> 23, 40, 41, 46, 47, 50, 61, 69, 54, 42, 32, 23, 55, NA, NA…
#> $ n_cub_min   <dbl> 1, 0, 0, 1, 2, 0, 0, 0, 0, 1, 2, 5, 37, 0, 0, 1, 3, 1, 0, …
#> $ n_gra_s     <dbl> 0.26, 0.00, 0.43, 0.25, 0.43, 0.35, 0.35, 0.43, 0.35, 0.00…
#> $ w_racha_mn  <dbl> 21.3, 22.5, 21.6, 20.9, 20.0, 20.8, 20.0, 19.6, 18.9, 19.7…
#> $ p_max_cv    <dbl> 0.91, 0.83, 0.86, 0.73, 0.69, 0.85, 0.92, 1.07, 1.04, 0.91…
#> $ n_des_q4    <dbl> 8.0, 10.0, 11.0, 7.2, 8.2, 11.2, 15.4, 14.0, 9.0, 8.2, 6.0…
#> $ n_des_q1    <dbl> 1.0, 3.0, 2.0, 2.0, 2.8, 5.0, 10.8, 8.8, 5.8, 2.0, 1.0, 1.…
#> $ n_des_q3    <dbl> 6.0, 6.0, 7.0, 5.0, 5.4, 8.0, 14.0, 12.0, 8.0, 6.0, 4.0, 4…
#> $ n_des_q2    <dbl> 4.2, 5.0, 4.0, 4.0, 3.6, 6.0, 13.0, 11.0, 7.0, 4.0, 2.0, 3…
#> $ ti_max_max  <dbl> 8.6, 13.4, 16.5, 19.7, 21.5, 26.4, 31.0, 28.3, 23.6, 20.2,…
#> $ n_gra_n     <dbl> 29, 30, 30, 30, 30, 30, 30, 30, 30, 30, 29, 30, 28, 30, 30…
#> $ tm_min_n    <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ nt_30_max   <dbl> 0, 0, 0, 3, 12, 24, 31, 29, 20, 4, 0, 0, 97, 0, 0, 0, 0, 1…
#> $ q_max_max   <dbl> 1012.8, 1007.3, 1004.4, 1004.3, 1002.5, 997.0, 998.0, 996.…
#> $ tm_min_s    <dbl> 1.48, 1.30, 1.05, 1.14, 1.26, 1.27, 1.18, 0.86, 1.16, 1.22…
#> $ nt_00_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ nv_0050_min <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ ts_10_s     <dbl> 1.19, 1.52, 1.41, NA, NA, 2.40, NA, 2.18, 1.85, 1.69, NA, …
#> $ n_nub_q4    <dbl> 21.2, 20.0, 21.2, 23.0, 24.0, 22.0, 20.0, 21.0, 22.0, 22.0…
#> $ n_nub_q1    <dbl> 15.0, 16.0, 16.0, 16.8, 18.0, 16.0, 14.8, 16.0, 18.0, 18.0…
#> $ n_nub_q2    <dbl> 18.0, 17.6, 18.0, 20.0, 20.6, 20.6, 16.0, 18.0, 19.0, 19.6…
#> $ n_nub_q3    <dbl> 18.4, 19.0, 20.4, 21.0, 23.0, 21.0, 17.0, 20.0, 21.0, 22.0…
#> $ ts_50_s     <dbl> 0.98, 1.03, 1.20, NA, NA, 1.79, NA, 1.53, NA, 1.46, NA, NA…
#> $ ts_10_n     <dbl> 15, 15, 15, 14, 14, 15, 14, 16, 15, 15, 12, 14, 11, 0, 0, …
#> $ nt_00_s     <dbl> 4.96, 3.51, 1.53, 0.25, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00…
#> $ ts_10_min   <dbl> 5.0, 7.3, 11.5, NA, NA, 24.9, NA, 27.4, 22.3, 16.9, NA, NA…
#> $ tm_mes_mn   <dbl> 7.2, 8.4, 11.7, 14.4, 18.5, 23.0, 25.5, 25.7, 21.4, 16.6, …
#> $ np_100_min  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 0, 0, 0, 0, 0, 0, 0…
#> $ nv_1000_min <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 0, 0, 0, 0, 0, 0, 0…
#> $ nt_30_s     <dbl> 0.00, 0.00, 0.00, 0.86, 3.28, 5.34, 4.60, 4.11, 5.34, 1.01…
#> $ nt_00_q3    <dbl> 7.0, 5.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 2.0, 5.4…
#> $ nt_00_q2    <dbl> 5.6, 3.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 2.0…
#> $ nt_00_q1    <dbl> 2.0, 1.8, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0…
#> $ n_tor_min   <dbl> 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 13, 0, 0, 0, 0, 0, 0, …
#> $ nt_00_q4    <dbl> 10.2, 8.0, 1.2, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 3.0, 10…
#> $ ts_20_q3    <dbl> 7.7, 8.8, 14.1, NA, NA, 29.7, NA, 31.3, 26.3, 20.2, NA, NA…
#> $ nt_00_max   <dbl> 20, 13, 7, 1, 0, 0, 0, 0, 0, 0, 5, 21, 49, 13, 12, 7, 1, 0…
#> $ nw_55_min   <dbl> 1, 2, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 26, 0, 0, 0, 0, 0, 0, …
#> $ n_llu_mn    <dbl> 10.0, 9.0, 8.0, 10.0, 10.0, 8.0, 4.0, 5.0, 7.0, 10.0, 11.0…
#> $ ts_50_mn    <dbl> 7.5, 8.3, 11.9, NA, NA, 26.7, NA, 29.0, NA, 19.0, NA, NA, …
#> $ glo_min     <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ ts_50_md    <dbl> 7.5, 8.4, 12.0, NA, NA, 26.5, NA, 29.1, NA, 19.4, NA, NA, …
#> $ q_med_min   <dbl> 978.6, 977.3, 976.4, 977.8, 981.3, 981.2, 983.7, 982.2, 98…
#> $ nv_0050_cv  <dbl> 5.48, 3.26, NA, NA, NA, NA, NA, NA, NA, 5.48, 5.48, 3.05, …
#> $ inso_q2     <dbl> 4.1, 6.0, 6.7, 8.1, 9.4, 10.8, 11.5, 10.6, 8.3, 6.6, 5.0, …
#> $ inso_q3     <dbl> 4.9, 6.5, 7.6, 8.5, 9.9, 11.1, 12.0, 10.9, 9.0, 7.1, 5.3, …
#> $ inso_q1     <dbl> 3.6, 5.5, 6.2, 7.4, 8.9, 10.3, 11.2, 10.0, 7.8, 5.8, 4.2, …
#> $ n_fog_max   <dbl> 17, 12, 2, 2, 2, 1, 1, 1, 2, 5, 12, 17, 36, 7, 5, 2, 0, 1,…
#> $ np_100_s    <dbl> 0.78, 0.97, 0.93, 1.01, 1.24, 1.17, 0.63, 0.63, 1.16, 1.17…
#> $ np_300_s    <dbl> 0.18, 0.00, 0.25, 0.35, 0.31, 0.35, 0.18, 0.31, 0.35, 0.31…
#> $ ta_min_cv   <dbl> -0.70, -0.96, 6.37, 0.61, 0.29, 0.14, 0.13, 0.10, 0.18, 0.…
#> $ tm_mes_cv   <dbl> 0.22, 0.19, 0.10, 0.08, 0.08, 0.07, 0.05, 0.05, 0.06, 0.10…
#> $ ts_min_md   <dbl> 9.3, 9.3, 11.3, 13.3, 17.2, 21.1, 22.4, 22.5, 20.0, 16.6, …
#> $ ts_min_mn   <dbl> 9.3, 9.8, 11.2, 13.6, 17.0, 21.1, 22.4, 22.6, 19.9, 16.6, …
#> $ inso_cv     <dbl> 0.24, 0.17, 0.17, 0.12, 0.10, 0.08, 0.07, 0.06, 0.10, 0.16…
#> $ np_300_n    <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ np_100_n    <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ ts_min_n    <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ tm_max_md   <dbl> 11.0, 13.4, 17.5, 20.4, 24.9, 30.0, 32.8, 32.4, 27.4, 21.8…
#> $ tm_max_mn   <dbl> 11.1, 13.4, 17.4, 20.4, 24.8, 29.8, 32.5, 32.3, 27.4, 21.5…
#> $ ts_min_s    <dbl> 2.10, 1.68, 1.48, 1.35, 1.41, 1.39, 0.91, 1.06, 1.60, 1.60…
#> $ tm_mes_q1   <dbl> 6.1, 7.0, 11.0, 13.6, 17.6, 22.1, 24.5, 24.9, 20.3, 15.4, …
#> $ p_max_q1    <dbl> 3.0, 2.4, 4.0, 8.1, 6.6, 3.9, 1.8, 3.2, 4.3, 5.3, 6.5, 3.3…
#> $ p_max_q3    <dbl> 8.4, 8.9, 8.8, 16.4, 16.2, 13.8, 10.3, 9.4, 12.6, 12.2, 12…
#> $ p_max_q2    <dbl> 6.7, 5.7, 6.0, 12.2, 10.4, 8.1, 4.5, 4.5, 7.3, 8.5, 8.8, 4…
#> $ p_max_q4    <dbl> 13.8, 15.2, 19.9, 22.5, 24.9, 21.3, 15.2, 17.2, 24.0, 16.1…
#> $ n_nie_s     <dbl> 1.02, 0.96, 0.52, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00…
#> $ n_gra_min   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ nt_30_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ n_nie_n     <dbl> 29, 30, 30, 30, 30, 30, 30, 30, 30, 30, 29, 30, 28, 30, 30…
#> $ ts_50_n     <dbl> 15, 15, 15, 14, 14, 15, 13, 15, 13, 15, 12, 14, 9, 0, 0, 0…
#> $ nv_0100_min <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ q_mar_q1    <dbl> 1018.3, 1015.9, 1014.0, 1012.6, 1013.5, 1013.7, 1013.6, 10…
#> $ q_mar_q2    <dbl> 1020.9, 1019.1, 1017.3, 1013.9, 1014.5, 1014.4, 1014.6, 10…
#> $ q_mar_q3    <dbl> 1023.1, 1021.5, 1019.0, 1015.1, 1015.1, 1015.7, 1015.0, 10…
#> $ q_mar_q4    <dbl> 1026.8, 1025.6, 1021.2, 1016.0, 1015.9, 1016.3, 1015.4, 10…
#> $ p_max_min   <dbl> 0.2, 0.0, 0.0, 1.4, 1.6, 0.0, 0.2, 0.7, 0.6, 0.5, 0.4, 0.3…
#> $ n_llu_cv    <dbl> 0.52, 0.55, 0.55, 0.47, 0.34, 0.42, 0.47, 0.47, 0.35, 0.43…
#> $ n_nub_mn    <dbl> 18.0, 18.5, 19.0, 20.0, 21.5, 21.0, 17.0, 19.0, 19.0, 21.0…
#> $ w_racha_q1  <dbl> 19.5, 20.6, 19.1, 18.7, 18.7, 18.1, 18.4, 17.4, 18.1, 18.3…
#> $ w_racha_q3  <dbl> 22.5, 23.2, 23.0, 21.7, 20.6, 21.4, 21.2, 20.0, 20.3, 20.3…
#> $ w_racha_q2  <dbl> 20.6, 22.2, 21.0, 20.0, 19.4, 20.6, 19.4, 18.5, 18.3, 19.4…
#> $ n_nub_md    <dbl> 18.1, 17.7, 19.2, 19.5, 21.3, 19.6, 16.7, 18.6, 19.9, 20.3…
#> $ w_racha_q4  <dbl> 25.1, 25.3, 24.5, 23.0, 22.1, 24.7, 23.0, 21.8, 21.1, 21.9…
#> $ nt_00_cv    <dbl> 0.74, 0.75, 1.84, 3.81, NA, NA, NA, NA, NA, NA, 1.17, 0.94…
#> $ np_010_q4   <dbl> 6.2, 6.0, 7.0, 7.2, 8.0, 6.0, 4.0, 3.2, 4.2, 8.0, 8.0, 7.0…
#> $ np_010_q1   <dbl> 2.0, 1.8, 3.0, 4.0, 4.0, 1.8, 1.0, 1.0, 2.0, 3.0, 2.8, 2.0…
#> $ np_010_q3   <dbl> 4.4, 5.0, 5.0, 6.0, 7.0, 5.0, 2.4, 2.0, 4.0, 5.4, 7.0, 4.0…
#> $ np_010_q2   <dbl> 3.0, 2.6, 4.0, 5.0, 5.0, 3.0, 2.0, 1.6, 3.0, 4.0, 5.0, 3.0…
#> $ nv_1000_q3  <dbl> 4.4, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 2.0, 4.0…
#> $ ts_50_cv    <dbl> 0.13, 0.12, 0.10, NA, NA, 0.07, NA, 0.05, NA, 0.08, NA, NA…
#> $ nv_0100_s   <dbl> 0.57, 0.48, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.25…
#> $ n_llu_max   <dbl> 17, 15, 21, 20, 20, 18, 10, 11, 13, 21, 21, 19, 137, 19, 1…
#> $ tm_max_min  <dbl> 6.8, 10.0, 14.4, 17.5, 20.4, 24.2, 29.0, 29.4, 24.7, 17.3,…
#> $ tm_min_md   <dbl> 2.9, 3.5, 6.0, 8.5, 12.3, 16.2, 18.6, 18.8, 15.4, 11.3, 6.…
#> $ q_mar_cv    <dbl> 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00…
#> $ nv_0100_n   <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 28, 29…
#> $ tm_min_mn   <dbl> 2.9, 3.5, 5.9, 8.4, 12.2, 16.0, 18.6, 18.9, 15.1, 11.3, 6.…
#> $ ts_10_mn    <dbl> 7.5, 9.0, 14.1, NA, NA, 30.8, NA, 32.1, 26.2, 18.5, NA, NA…
#> $ w_racha_cv  <dbl> 0.16, 0.13, 0.16, 0.13, 0.13, 0.20, 0.22, 0.15, 0.13, 0.15…
#> $ evap_min    <dbl> 358, 527, 947, 1004, 1370, 1374, 2318, 2059, 1146, 813, 41…
#> $ ta_max_cv   <dbl> 0.10, 0.11, 0.06, 0.07, 0.07, 0.07, 0.05, 0.05, 0.07, 0.08…
#> $ n_llu_q4    <dbl> 12.8, 11.0, 12.0, 14.0, 12.2, 9.4, 8.0, 7.0, 10.0, 12.2, 1…
#> $ n_llu_q2    <dbl> 9.0, 5.6, 8.0, 9.0, 10.0, 7.6, 4.0, 5.0, 7.0, 9.6, 10.0, 9…
#> $ n_llu_q3    <dbl> 10.0, 9.4, 9.4, 11.0, 11.0, 8.0, 5.0, 6.0, 8.0, 11.0, 11.8…
#> $ n_llu_q1    <dbl> 4.8, 3.8, 5.0, 5.8, 7.8, 5.8, 3.0, 3.0, 5.0, 6.8, 7.2, 6.8…
#> $ q_min_q4    <dbl> 980.0, 977.2, 976.6, 974.1, 976.3, 976.9, 978.1, 978.2, 97…
#> $ hr_min      <dbl> 63, 48, 48, 44, 44, 38, 40, 41, 49, 57, 55, 67, 55, 59, 51…
#> $ w_med_max   <dbl> 27, 27, 23, 26, 29, 23, 25, 21, 19, 19, 26, 24, 18, 9, 9, …
#> $ nw_91_mn    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0…
#> $ q_min_q1    <dbl> 964.3, 963.6, 964.6, 965.5, 970.6, 972.3, 975.4, 974.3, 97…
#> $ q_min_q2    <dbl> 967.6, 968.9, 969.4, 969.1, 972.5, 974.1, 976.0, 975.8, 97…
#> $ q_min_q3    <dbl> 974.3, 972.4, 973.1, 972.7, 973.6, 975.6, 977.0, 976.8, 97…
#> $ q_min_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ nw_91_md    <dbl> 0.2, 0.2, 0.2, 0.1, 0.0, 0.1, 0.2, 0.0, 0.0, 0.0, 0.1, 0.1…
#> $ q_min_s     <dbl> 9.13, 7.78, 7.39, 4.78, 3.38, 2.45, 1.86, 2.04, 4.03, 6.31…
#> $ n_tor_q4    <dbl> 0.0, 0.0, 1.0, 3.0, 6.0, 6.2, 6.0, 5.0, 4.0, 2.0, 0.0, 0.0…
#> $ n_tor_q1    <dbl> 0.0, 0.0, 0.0, 0.0, 2.0, 3.0, 2.0, 1.8, 1.8, 0.0, 0.0, 0.0…
#> $ n_tor_q2    <dbl> 0.0, 0.0, 0.0, 1.0, 4.0, 4.0, 3.0, 3.0, 2.0, 1.0, 0.0, 0.0…
#> $ n_tor_q3    <dbl> 0.0, 0.0, 0.4, 2.0, 4.4, 5.0, 4.0, 5.0, 3.0, 1.0, 0.0, 0.0…
#> $ tm_mes_min  <dbl> 3.2, 5.5, 9.3, 11.7, 14.8, 18.5, 22.7, 23.3, 19.1, 13.6, 8…
#> $ n_cub_max   <dbl> 15, 11, 12, 12, 10, 6, 2, 4, 8, 14, 13, 22, 74, 16, 11, 14…
#> $ nw_55_cv    <dbl> 0.62, 0.50, 0.50, 0.46, 0.59, 0.45, 0.57, 0.64, 0.48, 0.61…
#> $ tm_min_min  <dbl> -0.5, 0.8, 4.1, 5.8, 9.1, 12.8, 16.4, 17.1, 13.6, 8.6, 4.3…
#> $ q_mar_max   <dbl> 1030.7, 1027.7, 1024.3, 1019.2, 1018.3, 1017.4, 1017.0, 10…
#> $ nv_0100_q4  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 2, 0, 0, 0, 0, 0, 0, 0…
#> $ nv_0100_q1  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ nv_0100_q2  <dbl> 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0…
#> $ nv_0100_q3  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0…
#> $ q_max_q2    <dbl> 1002.7, 1001.8, 998.5, 995.6, 995.1, 993.7, 993.5, 993.0, …
#> $ q_max_q3    <dbl> 1004.7, 1004.0, 1001.3, 997.4, 996.5, 994.1, 994.5, 994.2,…
#> $ q_max_q1    <dbl> 1000.4, 998.8, 996.5, 993.8, 992.9, 993.1, 992.3, 992.7, 9…
#> $ q_max_q4    <dbl> 1007.9, 1005.2, 1002.9, 998.7, 997.4, 995.3, 996.7, 994.9,…
#> $ nv_0100_cv  <dbl> 2.44, 2.42, NA, NA, NA, NA, NA, NA, NA, 3.81, 2.27, 1.57, …
#> $ ts_10_q3    <dbl> 7.5, 9.2, 14.7, NA, NA, 30.9, NA, 32.1, 26.4, 19.5, NA, NA…
#> $ ts_10_q2    <dbl> 7.1, 8.3, 13.5, NA, NA, 30.4, NA, 31.8, 25.9, 18.4, NA, NA…
#> $ ts_10_q1    <dbl> 6.2, 7.9, 12.8, NA, NA, 28.9, NA, 30.1, 25.0, 17.7, NA, NA…
#> $ ts_10_q4    <dbl> 8.0, 9.9, 15.1, NA, NA, 32.1, NA, 34.0, 27.7, 20.8, NA, NA…
#> $ q_max_cv    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ ta_max_max  <dbl> 20.6, 25.5, 27.3, 32.4, 36.5, 43.2, 44.5, 42.8, 38.4, 32.0…
#> $ np_010_cv   <dbl> 0.75, 0.62, 0.58, 0.43, 0.49, 0.63, 0.63, 0.70, 0.55, 0.59…
#> $ ta_max_q4   <dbl> 19.4, 21.1, 25.7, 30.0, 34.9, 39.4, 40.7, 39.6, 35.5, 30.1…
#> $ ta_max_q3   <dbl> 18.1, 20.4, 25.1, 28.4, 33.5, 37.9, 39.2, 38.9, 34.4, 29.3…
#> $ ta_max_q2   <dbl> 17.5, 19.2, 24.5, 27.3, 32.1, 36.1, 38.4, 38.0, 33.7, 28.0…
#> $ ta_max_q1   <dbl> 16.2, 17.8, 23.4, 26.1, 30.8, 35.3, 37.7, 36.8, 31.0, 26.1…
#> $ np_010_min  <dbl> 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 36, 0, 0, 0, 2, 0, 0, …
#> $ w_med_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 28, 28…
#> $ nv_1000_q4  <dbl> 5.2, 3.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 3.0, 7.0…
#> $ nv_1000_q2  <dbl> 2.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 3.0…
#> $ n_nie_max   <dbl> 4, 3, 2, 0, 0, 0, 0, 0, 0, 0, 1, 4, 7, 5, 4, 2, 2, 0, 0, 0…
#> $ nv_1000_q1  <dbl> 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 7, 0, 0, 0, 0, 0, 0, 0…
#> $ nv_1000_cv  <dbl> 0.83, 1.69, 2.77, NA, 2.27, NA, NA, NA, 3.05, 1.71, 1.03, …
#> $ glo_max     <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ n_fog_s     <dbl> 4.37, 3.15, 0.63, 0.46, 0.60, 0.31, 0.18, 0.18, 0.46, 1.35…
#> $ ti_max_q2   <dbl> 4.1, 6.3, 9.0, 11.6, 16.2, 20.9, 24.4, 24.5, 19.3, 14.6, 8…
#> $ n_fog_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 29, 30, 29, 30, 30…
#> $ n_nie_cv    <dbl> 1.98, 1.52, 1.95, NA, NA, NA, NA, NA, NA, NA, 3.74, 2.42, …
#> $ np_001_min  <dbl> 2, 0, 0, 2, 2, 0, 1, 1, 2, 2, 1, 2, 61, 1, 0, 0, 4, 1, 1, …
#> $ nw_55_q4    <dbl> 13.0, 12.4, 13.0, 11.6, 12.0, 10.0, 11.0, 8.6, 6.0, 7.6, 1…
#> $ n_fog_min   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 3, 12, 0, 0, 0, 0, 0, 0, …
#> $ nw_55_q2    <dbl> 5.8, 8.0, 7.0, 7.0, 5.4, 6.0, 5.0, 4.8, 4.0, 4.0, 5.0, 5.0…
#> $ nw_55_q3    <dbl> 8.4, 10.6, 8.8, 9.0, 7.6, 7.0, 7.2, 7.0, 5.0, 5.2, 7.0, 6.…
#> $ nw_55_q1    <dbl> 4.0, 5.0, 5.0, 5.0, 4.0, 4.8, 4.0, 2.0, 2.0, 2.4, 3.4, 2.4…
#> $ hr_n        <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 23, 22…
#> $ hr_s        <dbl> 5.35, 6.14, 5.07, 4.72, 4.67, 4.44, 3.80, 4.28, 4.00, 5.11…
#> $ q_med_s     <dbl> 4.77, 4.70, 4.39, 2.53, 1.73, 1.53, 1.06, 1.17, 1.94, 2.44…
#> $ nv_0050_max <dbl> 1, 2, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 2, 0, 0, 0, 0, 0, 0, 0…
#> $ ta_min_q2   <dbl> -3.7, -2.4, 0.3, 3.0, 6.4, 10.8, 13.7, 13.6, 9.5, 3.8, -0.…
#> $ n_des_max   <dbl> 12, 13, 20, 16, 10, 18, 21, 17, 14, 16, 16, 12, 107, 16, 1…
#> $ n_fog_mn    <dbl> 5, 1, 0, 0, 0, 0, 0, 0, 0, 1, 3, 7, 20, 2, 0, 0, 0, 0, 0, …
#> $ p_sol_max   <dbl> 72, 82, 84, 83, 76, 79, 87, 86, 80, 79, 79, 59, 73, NA, NA…
#> $ np_001_s    <dbl> 4.80, 3.53, 4.11, 3.62, 3.42, 2.79, 1.95, 1.90, 2.24, 3.75…
#> $ n_fog_md    <dbl> 6.3, 2.4, 0.5, 0.2, 0.3, 0.1, 0.0, 0.0, 0.2, 1.2, 3.6, 7.3…
#> $ q_min_cv    <dbl> 0.01, 0.01, 0.01, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.01…
#> $ np_001_n    <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ e_max       <dbl> 97, 92, 104, 116, 143, 164, 192, 219, 180, 150, 123, 93, 1…
#> $ nt_00_md    <dbl> 6.7, 4.7, 0.8, 0.1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.5, 5.5…
#> $ np_100_mn   <dbl> 0.0, 0.0, 0.0, 1.0, 1.0, 0.5, 0.0, 0.0, 0.5, 0.5, 0.0, 0.0…
#> $ nt_30_q1    <dbl> 0.0, 0.0, 0.0, 0.0, 1.8, 11.0, 18.8, 19.8, 3.6, 0.0, 0.0, …
#> $ nt_00_mn    <dbl> 6.0, 4.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.5, 3.5…
#> $ glo_q4      <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ glo_q2      <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ glo_q3      <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ np_100_md   <dbl> 0.5, 0.5, 0.6, 1.1, 1.1, 0.9, 0.5, 0.5, 0.9, 0.9, 0.8, 0.3…
#> $ inso_n      <dbl> 30, 30, 30, 30, 30, 30, 30, 29, 29, 30, 30, 30, 29, 13, 14…
#> $ q_med_q1    <dbl> 987.1, 984.8, 983.5, 982.4, 983.7, 984.3, 984.6, 984.9, 98…
#> $ np_001_md   <dbl> 8.5, 6.7, 7.7, 8.7, 8.6, 6.2, 4.2, 3.6, 5.6, 7.9, 8.9, 8.5…
#> $ np_001_mn   <dbl> 7.5, 6.5, 7.0, 9.5, 8.0, 6.0, 4.0, 3.0, 5.0, 8.0, 9.0, 8.0…
#> $ ti_max_cv   <dbl> 0.60, 0.40, 0.32, 0.20, 0.18, 0.14, 0.10, 0.07, 0.11, 0.21…
#> $ nv_1000_mn  <dbl> 4.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 3.0…
#> $ nv_1000_md  <dbl> 3.6, 1.3, 0.2, 0.0, 0.2, 0.0, 0.0, 0.0, 0.1, 0.6, 1.8, 4.3…
#> $ evap_md     <dbl> 910, 1233, 1799, 2005, 2557, 3176, 3712, 3430, 2357, 1622,…
#> $ ta_min_max  <dbl> 1.8, 2.0, 4.8, 8.4, 10.2, 15.1, 17.7, 16.5, 12.7, 9.4, 3.5…
#> $ e_md        <dbl> 76, 74, 84, 96, 117, 140, 155, 161, 143, 125, 96, 80, 112,…
#> $ e_mn        <dbl> 75, 74, 84, 94, 117, 141, 153, 161, 140, 124, 96, 80, 112,…
#> $ nt_30_q3    <dbl> 0.0, 0.0, 0.0, 0.0, 5.0, 18.4, 26.0, 24.4, 10.0, 0.0, 0.0,…
#> $ ti_max_md   <dbl> 4.2, 7.0, 9.7, 12.1, 16.5, 21.2, 25.2, 25.0, 19.7, 14.7, 8…
#> $ n_llu_n     <dbl> 29, 30, 30, 30, 30, 30, 30, 30, 30, 30, 29, 30, 28, 30, 30…
#> $ ti_max_mn   <dbl> 4.5, 7.1, 10.1, 11.8, 17.0, 21.4, 24.9, 25.1, 19.7, 15.2, …
#> $ glo_s       <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ glo_n       <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ ts_min_cv   <dbl> 0.23, 0.18, 0.13, 0.10, 0.08, 0.07, 0.04, 0.05, 0.08, 0.10…
#> $ n_llu_s     <dbl> 4.89, 4.24, 4.90, 4.75, 3.52, 3.29, 2.36, 2.51, 2.65, 4.25…
#> $ nv_0050_mn  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ n_nie_q3    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 1, 1, 0, 0, 0, 0, 0…
#> $ n_nie_q2    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0…
#> $ n_nie_q4    <dbl> 1.0, 1.2, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.2…
#> $ glo_cv      <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ tm_max_max  <dbl> 13.5, 17.4, 20.7, 23.7, 27.8, 33.8, 35.8, 35.5, 30.4, 25.6…
#> $ ts_20_mn    <dbl> 7.5, 8.6, 13.7, NA, NA, 28.6, NA, 31.0, 25.7, 18.5, NA, NA…
#> $ tm_max_cv   <dbl> 0.16, 0.16, 0.09, 0.07, 0.07, 0.07, 0.05, 0.05, 0.06, 0.09…
#> $ ts_20_md    <dbl> 7.3, 8.8, 13.4, NA, NA, 28.8, NA, 30.6, 25.7, 19.3, NA, NA…
#> $ np_001_cv   <dbl> 0.57, 0.52, 0.54, 0.41, 0.40, 0.45, 0.47, 0.53, 0.40, 0.48…
#> $ p_max_mn    <dbl> 7.5, 6.1, 7.0, 14.3, 14.0, 10.7, 7.0, 7.0, 9.4, 10.3, 9.5,…
#> $ nw_91_max   <dbl> 1, 2, 2, 1, 1, 1, 2, 1, 1, 0, 1, 2, 3, 0, 1, 0, 0, 0, 0, 1…
#> $ p_max_md    <dbl> 9.9, 8.7, 11.4, 17.0, 16.1, 13.9, 9.4, 11.3, 15.2, 13.9, 1…
#> $ q_med_cv    <dbl> 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00…
#> $ ts_50_max   <dbl> 9.0, 10.6, 13.7, NA, NA, 29.4, NA, 31.5, NA, 21.8, NA, NA,…
#> $ ts_50_min   <dbl> 5.6, 7.1, 10.0, NA, NA, 23.0, NA, 26.4, NA, 17.4, NA, NA, …
#> $ ta_min_n    <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ e_q4        <dbl> 82, 84, 91, 103, 126, 147, 168, 169, 155, 135, 110, 88, 11…
#> $ e_q3        <dbl> 78, 75, 85, 97, 117, 142, 157, 161, 145, 127, 98, 83, 113,…
#> $ e_q2        <dbl> 74, 72, 83, 92, 114, 139, 150, 157, 137, 120, 94, 79, 111,…
#> $ e_q1        <dbl> 70, 69, 76, 88, 109, 133, 147, 150, 131, 116, 83, 71, 109,…
#> $ glo_q1      <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ w_racha_md  <dbl> 21.9, 22.8, 22.2, 20.9, 20.4, 21.4, 21.1, 19.7, 19.7, 19.5…
#> $ q_min_min   <dbl> 950.9, 958.9, 953.7, 959.5, 966.4, 970.3, 972.1, 971.1, 96…
#> $ ts_min_q3   <dbl> 9.7, 10.0, 11.4, 13.8, 17.4, 21.4, 22.5, 22.9, 20.3, 16.9,…
#> $ ts_min_q2   <dbl> 8.6, 9.2, 11.0, 13.0, 16.8, 20.6, 22.3, 22.3, 19.5, 16.5, …
#> $ ts_min_q1   <dbl> 7.1, 8.0, 10.4, 12.1, 16.0, 19.7, 21.5, 21.5, 18.9, 15.6, …
#> $ p_mes_max   <dbl> 81.0, 70.5, 71.2, 126.6, 141.9, 100.1, 50.6, 65.8, 101.4, …
#> $ ts_min_q4   <dbl> 11.1, 10.7, 12.3, 14.4, 18.3, 22.1, 23.2, 23.3, 21.2, 17.8…
#> $ q_mar_min   <dbl> 1009.4, 1008.3, 1006.8, 1008.0, 1011.2, 1010.6, 1012.6, 10…
#> $ tm_max_q3   <dbl> 11.6, 14.3, 17.8, 20.7, 25.3, 30.5, 33.5, 32.6, 27.9, 21.9…
#> $ tm_max_q2   <dbl> 10.8, 12.3, 16.8, 20.1, 24.4, 29.4, 32.3, 32.1, 26.9, 21.2…
#> $ tm_max_q1   <dbl> 9.8, 11.5, 16.2, 19.1, 23.5, 28.7, 31.3, 31.3, 25.9, 20.5,…
#> $ p_mes_mn    <dbl> 17.2, 16.6, 21.3, 35.6, 33.0, 22.7, 13.1, 9.5, 20.1, 26.8,…
#> $ tm_max_q4   <dbl> 12.7, 15.2, 18.8, 21.3, 26.7, 32.0, 34.4, 33.6, 28.5, 23.8…
#> $ n_gra_mn    <dbl> 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0…
#> $ tm_min_max  <dbl> 5.7, 5.6, 8.9, 11.2, 14.4, 19.4, 20.8, 20.2, 18.1, 13.4, 9…
#> $ np_001_q4   <dbl> 13.0, 9.2, 10.2, 11.0, 11.0, 8.2, 6.0, 5.0, 7.0, 10.0, 12.…
#> $ np_001_q3   <dbl> 9.0, 8.0, 8.4, 10.0, 8.4, 7.0, 4.0, 4.0, 6.0, 8.0, 10.0, 9…
#> $ np_001_q2   <dbl> 7.0, 6.0, 6.6, 7.0, 8.0, 5.0, 3.0, 3.0, 5.0, 7.6, 8.0, 6.6…
#> $ np_001_q1   <dbl> 4.0, 3.8, 5.0, 5.8, 6.0, 4.0, 3.0, 2.0, 4.0, 5.0, 6.6, 5.0…
#> $ ti_max_n    <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ w_med_md    <dbl> 16.1, 17.5, 17.4, 18.1, 18.0, 17.4, 18.0, 16.1, 14.5, 13.7…
#> $ ti_max_q4   <dbl> 6.2, 8.5, 11.7, 13.5, 18.5, 23.5, 27.2, 26.5, 21.4, 17.0, …
#> $ ti_max_q3   <dbl> 4.8, 7.7, 10.4, 11.9, 17.5, 21.9, 25.4, 25.5, 20.2, 16.1, …
#> $ n_gra_q2    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ ti_max_q1   <dbl> 1.8, 4.6, 6.5, 10.8, 13.8, 19.0, 23.5, 23.2, 17.4, 11.5, 6…
#> $ ti_max_s    <dbl> 2.50, 2.76, 3.08, 2.45, 2.93, 2.90, 2.47, 1.71, 2.10, 3.12…
#> $ q_med_q2    <dbl> 989.3, 987.9, 986.6, 983.7, 984.6, 985.2, 985.5, 985.2, 98…
#> $ q_med_q3    <dbl> 991.6, 990.1, 988.2, 984.8, 985.2, 986.3, 985.8, 985.9, 98…
#> $ nt_30_q2    <dbl> 0.0, 0.0, 0.0, 0.0, 3.0, 14.0, 22.6, 22.6, 6.6, 0.0, 0.0, …
#> $ n_nie_min   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ nt_30_q4    <dbl> 0.0, 0.0, 0.0, 1.0, 7.0, 21.0, 28.0, 26.0, 12.0, 1.2, 0.0,…
#> $ q_med_q4    <dbl> 995.2, 994.4, 990.3, 985.8, 986.2, 986.9, 986.4, 986.4, 98…
#> $ n_cub_md    <dbl> 8.2, 4.6, 5.6, 5.5, 4.6, 2.3, 0.6, 1.2, 2.8, 5.0, 6.9, 9.4…
#> $ n_cub_mn    <dbl> 8.0, 4.0, 6.0, 5.0, 4.5, 2.0, 0.0, 1.0, 2.5, 4.0, 7.5, 8.0…
#> $ ta_max_min  <dbl> 13.4, 15.9, 21.0, 24.8, 28.4, 32.3, 34.4, 34.9, 29.6, 22.8…
#> $ np_010_max  <dbl> 15, 7, 13, 11, 16, 10, 6, 5, 7, 13, 12, 16, 73, 14, 13, 18…
#> $ inso_md     <dbl> 4.6, 6.4, 7.3, 8.2, 9.5, 10.8, 11.7, 10.6, 8.6, 6.7, 5.2, …
#> $ inso_mn     <dbl> 4.6, 6.2, 7.4, 8.4, 9.7, 10.9, 11.8, 10.7, 8.8, 6.8, 5.2, …
#> $ evap_mn     <dbl> 876, 1256, 1801, 1972, 2559, 3156, 3775, 3475, 2370, 1634,…

Site built with pkgdown 2.1.1.

Template rostemplate by dieghernan, based on bootstraptemple.com.