Wind conditions for Zaragoza Airport ("9434") period 2000-2020

Daily observations of wind speed and directions for Zaragoza Airport (2000-2020). This is an example dataset.

Format

A tibble with columns:

fecha

Date of observation.

dir

Wind directions (0-360).

velmedia

Average wind speed (km/h)

Source

AEMET.

See also

Examples

data(climaemet_9434_wind)

Site built with pkgdown 2.2.0.

Template rostemplate by dieghernan, based on Universal Theme for Hugo.