Install an AEMET API Key

This function will store your AEMET API key on your local machine so it can be called securely without being stored in your code. After you have installed your key, it can be called any time by typing Sys.getenv("AEMET_API_KEY") and can be used in package functions by simply typing AEMET_API_KEY.

Alternatively, you can install the API Key manually:

Run Sys.setenv(AEMET_API_KEY = "Your_Key"). You would need to run this command on each session (Similar to install = FALSE).

Write this line on your .Renviron file: AEMET_API_KEY = "Your_Key" ( same behavior than install = TRUE). This would store your API key permanently.

Usage

aemet_api_key(apikey, overwrite = FALSE, install = FALSE)

Arguments

apikey

The API key provided to you from the AEMET formatted in quotes. A key can be acquired at https://opendata.aemet.es/centrodedescargas/inicio.

overwrite

If this is set to TRUE, it will overwrite an existing AEMET_API_KEY that you already have in local machine.

install

if TRUE, will install the key in your local machine for use in future sessions. Defaults to FALSE.

Value

None

Details

You can pass several apikey values as a vector c(api1, api2), in this case several AEMET_API_KEY values would be generated. In each subsequent api call climaemet would randomly choose one of the provided API keys.

This is useful when performing batch queries to avoid API throttling.

Note

To locate your API Key on your local machine, run rappdirs::user_cache_dir("climaemet", "R").

See also

Examples

# Don't run these examples!

if (FALSE) {
  aemet_api_key("111111abc", install = TRUE)

  # You can check it with:
  Sys.getenv("AEMET_API_KEY")
}

if (FALSE) {
  # If you need to overwrite an existing key:
  aemet_api_key("222222abc", overwrite = TRUE, install = TRUE)

  # You can check it with:
  Sys.getenv("AEMET_API_KEY")
}

Site built with pkgdown 2.0.9.

Template rostemplate by dieghernan, based on Bootstrapious.