Install a AEMET API Key in Your .Renviron File for Repeated Use

This function will add your AEMET API key to your .Renviron file so it can be called securely without being stored in your code. After you have installed your key, it can be called any time by typing Sys.getenv("AEMET_API_KEY") and can be used in package functions by simply typing AEMET_API_KEY If you do not have an .Renviron file, the function will create on for you. If you already have an .Renviron file, the function will append the key to your existing file, while making a backup of your original file for disaster recovery purposes.

aemet_api_key(apikey, overwrite = FALSE, install = FALSE)

Arguments

apikey
The API key provided to you from the AEMET formatted in quotes. A key can be acquired at https://opendata.aemet.es/centrodedescargas/inicio.

overwrite
If this is set to TRUE, it will overwrite an existing AEMET_API_KEY that you already have in your .Renviron file.

install
if TRUE, will install the key in your .Renviron file for use in future sessions. Defaults to FALSE.

Note

Code adapted from https://walker-data.com/tidycensus/reference/census_api_key.html

Examples

# Don't run these examples!

if (FALSE) {
  aemet_api_key("111111abc", install = TRUE)
  # First time, reload your environment o restart your session.
  readRenviron("~/.Renviron")
  # You can check it with:
  Sys.getenv("AEMET_API_KEY")
}

if (FALSE) {
  # If you need to overwrite an existing key:
  aemet_api_key("111111abc", overwrite = TRUE, install = TRUE)
  readRenviron("~/.Renviron")
  # You can check it with:
  Sys.getenv("AEMET_API_KEY")
}

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.