Daily/annual climatology values

Get climatology values for a station or for all the available stations. Note that aemet_daily_period() and aemet_daily_period_all() are shortcuts of aemet_daily_clim().

Usage,

aemet_daily_clim(
  station = "all",
  start = Sys.Date() - 7,
  end = Sys.Date(),
  verbose = FALSE,
  return_sf = FALSE
)

aemet_daily_period(
  station,
  start = 2020,
  end = 2020,
  verbose = FALSE,
  return_sf = FALSE
)

aemet_daily_period_all(
  start = 2020,
  end = 2020,
  verbose = FALSE,
  return_sf = FALSE
)

Arguments

station

Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

start, end

Character string with start and end date. See Details.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf

Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

A tibble or a sf object

Details

start and end parameters should be:

For aemet_daily_clim(): A Date object or a string with format: YYYY-MM-DD (2020-12-31) coercible with as.Date().

For aemet_daily_period() and aemet_daily_period_all(): A string representing the year(s) to be extracted: "2020", "2018".

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_daily_clim(c("9434", "3195"))
glimpse(obs)
#> Rows: 8
#> Columns: 20
#> $ fecha       <date> 2021-11-25, 2021-11-26, 2021-11-27, 2021-11-28, 2021-11-2…
#> $ indicativo  <chr> "9434", "9434", "9434", "9434", "3195", "3195", "3195", "3…
#> $ nombre      <chr> "ZARAGOZA, AEROPUERTO", "ZARAGOZA, AEROPUERTO", "ZARAGOZA,…
#> $ provincia   <chr> "ZARAGOZA", "ZARAGOZA", "ZARAGOZA", "ZARAGOZA", "MADRID", …
#> $ altitud     <dbl> 249, 249, 249, 249, 667, 667, 667, 667
#> $ tmed        <dbl> 7.6, 8.4, 6.5, 5.6, 6.2, 6.5, 5.4, 4.7
#> $ prec        <dbl> 4.0, 0.0, 0.6, 4.0, 0.0, 0.0, 0.5, 0.0
#> $ tmin        <dbl> 6.5, 6.5, 4.1, 3.6, 3.2, 2.8, 2.3, 1.1
#> $ horatmin    <time> 07:00:00, 21:50:00, 22:30:00, 03:10:00, 03:20:00, 07:20:00…
#> $ tmax        <dbl> 8.6, 10.4, 8.9, 7.6, 9.3, 10.2, 8.6, 8.3
#> $ horatmax    <time> 13:10:00, 12:30:00, 01:40:00, 12:20:00, 13:40:00, 14:30:00…
#> $ dir         <dbl> 31, 30, 31, 32, NA, NA, NA, NA
#> $ velmedia    <dbl> 6.7, 8.3, 8.6, 7.2, NA, NA, NA, NA
#> $ racha       <dbl> 13.9, 17.2, 18.6, 16.1, NA, NA, NA, NA
#> $ horaracha   <time> 21:40:00, 11:30:00, 11:30:00, 12:30:00,       NA,       NA…
#> $ sol         <dbl> 0.0, 7.1, 3.2, 0.6, NA, NA, NA, NA
#> $ presMax     <dbl> 979.0, 980.1, 979.3, 983.1, 933.0, 934.1, 934.4, 938.5
#> $ horaPresMax <chr> "21", "10", "24", "10", "24", "Varias", "24", "10"
#> $ presMin     <dbl> 975.3, 975.3, 970.7, 979.0, 929.5, 931.7, 927.9, 934.2
#> $ horaPresMin <chr> "Varias", "24", "05", "01", "06", "24", "07", "00"

Site built with pkgdown 2.0.0.

Template by Bootstrapious . Ported to pkgdown by dieghernan.