Daily/annual climatology values

Get climatology values for a station or for all the available stations. Note that aemet_daily_period() and aemet_daily_period_all() are shortcuts of aemet_daily_clim().

Usage

aemet_daily_clim(
  station = "all",
  start = Sys.Date() - 7,
  end = Sys.Date(),
  verbose = FALSE,
  return_sf = FALSE
)

aemet_daily_period(
  station,
  start = 2020,
  end = 2020,
  verbose = FALSE,
  return_sf = FALSE
)

aemet_daily_period_all(
  start = 2020,
  end = 2020,
  verbose = FALSE,
  return_sf = FALSE
)

Arguments

station

Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

start, end

Character string with start and end date. See Details.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf

Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

A tibble or a sf object

Details

start and end parameters should be:

For aemet_daily_clim(): A Date object or a string with format: YYYY-MM-DD (2020-12-31) coercible with as.Date().

For aemet_daily_period() and aemet_daily_period_all(): A string representing the year(s) to be extracted: "2020", "2018".

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_daily_clim(c("9434", "3195"))
glimpse(obs)
#> Rows: 8
#> Columns: 20
#> $ fecha       <date> 2023-02-09, 2023-02-10, 2023-02-11, 2023-02-12, 2023-02-0…
#> $ indicativo  <chr> "9434", "9434", "9434", "9434", "3195", "3195", "3195", "3…
#> $ nombre      <chr> "ZARAGOZA, AEROPUERTO", "ZARAGOZA, AEROPUERTO", "ZARAGOZA,…
#> $ provincia   <chr> "ZARAGOZA", "ZARAGOZA", "ZARAGOZA", "ZARAGOZA", "MADRID", …
#> $ altitud     <dbl> 249, 249, 249, 249, 667, 667, 667, 667
#> $ tmed        <dbl> 5.6, 4.8, 4.6, 4.6, 7.2, 5.5, 5.8, 6.6
#> $ prec        <dbl> 0.2, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0
#> $ tmin        <dbl> 1.8, -1.3, -1.8, -2.2, 4.2, 2.0, 1.3, 1.9
#> $ horatmin    <time> 23:59:00, 07:00:00, 06:30:00, 06:50:00, 05:20:00, 06:20:00…
#> $ tmax        <dbl> 9.5, 10.9, 11.0, 11.3, 10.3, 9.0, 10.4, 11.2
#> $ horatmax    <time> 13:00:00, 15:10:00, 14:50:00, 15:30:00, 16:20:00, 15:30:00…
#> $ dir         <chr> "99", "11", "12", "12", "16", "15", "08", "05"
#> $ velmedia    <dbl> 1.9, 1.1, 2.2, 1.9, 0.8, 1.9, 3.1, 2.5
#> $ racha       <dbl> 4.4, 4.2, 6.4, 6.7, 6.1, 6.4, 8.1, 6.9
#> $ horaracha   <chr> "Varias", "14:40", "14:10", "13:20", "21:00", "14:20", "12…
#> $ sol         <dbl> 3.1, 9.7, 10.0, 10.0, NA, NA, NA, NA
#> $ presMax     <dbl> 1005.6, 1007.8, 1006.0, 1004.3, 956.1, 958.2, 955.7, 953.8
#> $ horaPresMax <chr> "24", "10", "00", "09", "24", "10", "00", "24"
#> $ presMin     <dbl> 998.6, 1004.9, 1001.7, 1001.0, 948.9, 954.5, 951.1, 951.1
#> $ horaPresMin <chr> "01", "16", "17", "16", "Varias", "17", "16", "Varias"

Site built with pkgdown 2.0.7.

Template rostemplate by dieghernan, based on Bootstrapious.