Daily/annual climatology values

Get climatology values for a station or for all the available stations. Note that aemet_daily_period() and aemet_daily_period_all() are shortcuts of aemet_daily_clim().

Usage

aemet_daily_clim(
  station = "all",
  start = Sys.Date() - 7,
  end = Sys.Date(),
  verbose = FALSE,
  return_sf = FALSE
)

aemet_daily_period(
  station,
  start = 2020,
  end = 2020,
  verbose = FALSE,
  return_sf = FALSE
)

aemet_daily_period_all(
  start = 2020,
  end = 2020,
  verbose = FALSE,
  return_sf = FALSE
)

Arguments

station

Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

start, end

Character string with start and end date. See Details.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf

Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

A tibble or a sf object

Details

start and end parameters should be:

For aemet_daily_clim(): A Date object or a string with format: YYYY-MM-DD (2020-12-31) coercible with as.Date().

For aemet_daily_period() and aemet_daily_period_all(): A string representing the year(s) to be extracted: "2020", "2018".

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_daily_clim(c("9434", "3195"))
glimpse(obs)
#> Rows: 8
#> Columns: 20
#> $ fecha       <date> 2022-08-31, 2022-09-01, 2022-09-02, 2022-09-03, 2022-08-3…
#> $ indicativo  <chr> "9434", "9434", "9434", "9434", "3195", "3195", "3195", "3…
#> $ nombre      <chr> "ZARAGOZA, AEROPUERTO", "ZARAGOZA, AEROPUERTO", "ZARAGOZA,…
#> $ provincia   <chr> "ZARAGOZA", "ZARAGOZA", "ZARAGOZA", "ZARAGOZA", "MADRID", …
#> $ altitud     <dbl> 249, 249, 249, 249, 667, 667, 667, 667
#> $ tmed        <dbl> 27.7, 26.6, 26.8, 24.8, 26.0, 24.8, 23.9, 23.1
#> $ prec        <dbl> 0, 0, 0, 0, 0, 0, 0, 0
#> $ tmin        <dbl> 22.1, 19.9, 20.9, 19.1, 20.3, 17.6, 17.5, 17.4
#> $ horatmin    <time> 23:59:00, 05:30:00, 06:40:00, 05:40:00, 23:40:00, 06:30:00…
#> $ tmax        <dbl> 33.3, 33.2, 32.7, 30.6, 31.8, 31.9, 30.3, 28.8
#> $ horatmax    <time> 16:19:00, 16:50:00, 16:00:00, 16:00:00, 14:30:00, 13:30:00…
#> $ dir         <dbl> 31, 29, 31, 30, 25, 23, 23, 99
#> $ velmedia    <dbl> 4.7, 1.7, 4.2, 3.3, 2.2, 2.2, 2.5, 1.7
#> $ racha       <dbl> 11.1, 8.3, 14.4, 12.2, 8.6, 8.6, 9.2, 7.5
#> $ horaracha   <chr> "22:20", "04:30", "20:20", "Varias", "10:00", "17:20", "13…
#> $ sol         <dbl> 10.5, 12.1, 9.5, 10.7, NA, NA, NA, NA
#> $ presMax     <dbl> 986.8, 986.7, 983.9, 984.2, 942.4, 940.0, 938.6, 938.5
#> $ horaPresMax <chr> "09", "00", "Varias", "09", "08", "10", "09", "24"
#> $ presMin     <dbl> 983.4, 980.6, 978.9, 980.2, 938.9, 936.7, 935.6, 935.9
#> $ horaPresMin <chr> "17", "Varias", "17", "16", "18", "18", "18", "16"

Site built with pkgdown 2.0.6.

Template rostemplate by dieghernan, based on Bootstrapious.