Daily/annual climatology values

Get climatology values for a station or for all the available stations. Note that aemet_daily_period() and aemet_daily_period_all() are shortcuts of aemet_daily_clim().

aemet_daily_clim(
  station = "all",
  start = Sys.Date() - 7,
  end = Sys.Date(),
  verbose = FALSE,
  return_sf = FALSE
)

aemet_daily_period(
  station,
  start = 2020,
  end = 2020,
  verbose = FALSE,
  return_sf = FALSE
)

aemet_daily_period_all(
  start = 2020,
  end = 2020,
  verbose = FALSE,
  return_sf = FALSE
)

Arguments

station
Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

start, end
Character string with start and end date. See Details.

verbose
Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf
Logical. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

a tibble or a sf object

Details

start and end parameters should be:

For aemet_daily_clim(): A Date object or a string with format: %Y%m%d (2020-12-31).

For aemet_daily_period() and aemet_daily_period_all(): A string representing the year(s) to be extracted: "2020", "2018".

Examples

# Run this example only if AEMET_API_KEY is detected

if (aemet_detect_api_key()) {
  library(tibble)
  obs <- aemet_daily_clim(c("9434", "3195"))
  glimpse(obs)
}

#> Rows: 8
#> Columns: 20
#> $ fecha       <date> 2021-06-23, 2021-06-24, 2021-06-25, 2021-06-26, 2021-06-2…
#> $ indicativo  <chr> "9434", "9434", "9434", "9434", "3195", "3195", "3195", "3…
#> $ nombre      <chr> "ZARAGOZA, AEROPUERTO", "ZARAGOZA, AEROPUERTO", "ZARAGOZA,…
#> $ provincia   <chr> "ZARAGOZA", "ZARAGOZA", "ZARAGOZA", "ZARAGOZA", "MADRID", …
#> $ altitud     <dbl> 249, 249, 249, 249, 667, 667, 667, 667
#> $ tmed        <dbl> 19.2, 20.6, 23.8, 25.6, 19.0, 20.7, 23.8, 25.7
#> $ prec        <dbl> 0.2, 0.0, 0.0, 2.8, 0.0, 0.0, 0.0, 0.0
#> $ tmin        <dbl> 15.3, 14.1, 15.6, 18.8, 14.7, 13.6, 17.0, 19.5
#> $ horatmin    <time> 06:10:00, 04:57:00, 04:50:00, 05:00:00, 01:40:00, 04:40:00…
#> $ tmax        <dbl> 23.2, 27.2, 31.9, 32.3, 23.4, 27.8, 30.6, 31.9
#> $ horatmax    <chr> "16:10", "17:29", "18:00", "13:30", "15:40", "14:10", "Var…
#> $ dir         <dbl> 31, 30, 30, 14, NA, NA, NA, NA
#> $ velmedia    <dbl> 6.9, 6.4, 2.8, 4.2, NA, NA, NA, NA
#> $ racha       <dbl> 14.4, 12.2, 9.7, 13.1, NA, NA, NA, NA
#> $ horaracha   <time> 10:40:00, 06:40:00, 02:30:00, 18:40:00,       NA,       NA…
#> $ sol         <dbl> 9.4, 14.6, 13.9, 9.0, NA, NA, NA, NA
#> $ presMax     <dbl> 992.4, 992.8, 992.4, 986.4, 943.3, 944.3, 943.7, 939.8
#> $ horaPresMax <chr> "23", "08", "00", "Varias", "24", "08", "08", "06"
#> $ presMin     <dbl> 987.5, 989.5, 984.7, 980.7, 939.9, 941.6, 938.5, 935.6
#> $ horaPresMin <chr> "03", "17", "19", "19", "01", "Varias", "20", "18"

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.