Daily/annual climatology values

Get climatology values for a station or for all the available stations. Note that aemet_daily_period() and aemet_daily_period_all() are shortcuts of aemet_daily_clim().

aemet_daily_clim(
  station = "all",
  start = Sys.Date() - 7,
  end = Sys.Date(),
  verbose = FALSE,
  return_sf = FALSE
)

aemet_daily_period(
  station,
  start = 2020,
  end = 2020,
  verbose = FALSE,
  return_sf = FALSE
)

aemet_daily_period_all(
  start = 2020,
  end = 2020,
  verbose = FALSE,
  return_sf = FALSE
)

Arguments

station
Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

start, end
Character string with start and end date. See Details.

verbose
Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf
Logical. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

a tibble or a sf object

Details

start and end parameters should be:

For aemet_daily_clim(): A Date object or a string with format: %Y%m%d (2020-12-31).

For aemet_daily_period() and aemet_daily_period_all(): A string representing the year(s) to be extracted: "2020", "2018".

Examples

# Run this example only if AEMET_API_KEY is detected

if (aemet_detect_api_key()) {
  library(tibble)
  obs <- aemet_daily_clim(c("9434", "3195"))
  glimpse(obs)
}

#> Rows: 7
#> Columns: 20
#> $ fecha       <date> 2021-03-28, 2021-03-29, 2021-03-31, 2021-03-28, 2021-03-2…
#> $ indicativo  <chr> "9434", "9434", "9434", "3195", "3195", "3195", "3195"
#> $ nombre      <chr> "ZARAGOZA, AEROPUERTO", "ZARAGOZA, AEROPUERTO", "ZARAGOZA,…
#> $ provincia   <chr> "ZARAGOZA", "ZARAGOZA", "ZARAGOZA", "MADRID", "MADRID", "M…
#> $ altitud     <dbl> 249, 249, 249, 667, 667, 667, 667
#> $ tmed        <dbl> 14.2, 15.4, 16.4, 14.6, 16.2, 16.2, 16.8
#> $ prec        <dbl> 0, NA, 0, 0, 0, 0, 0
#> $ tmin        <dbl> 6.1, 7.4, 8.3, 8.3, 9.8, 11.8, 10.3
#> $ horatmin    <time> 06:30:00, 05:20:00, 06:00:00, 06:10:00, 05:20:00, 06:30:00…
#> $ tmax        <dbl> 22.3, 23.5, 24.6, 20.9, 22.5, 20.5, 23.4
#> $ horatmax    <time> 16:00:00, 14:50:00, 15:20:00, 15:30:00, 13:40:00, 15:20:00…
#> $ dir         <dbl> 12, 12, 12, NA, NA, NA, NA
#> $ velmedia    <dbl> 1.9, 5.8, 4.7, NA, NA, NA, NA
#> $ racha       <dbl> 8.1, 13.3, 11.9, NA, NA, NA, NA
#> $ horaracha   <time> 19:10:00, 18:40:00, 19:50:00,       NA,       NA,       NA…
#> $ sol         <dbl> 11.8, 11.5, 11.3, NA, NA, NA, NA
#> $ presMax     <dbl> 1000.6, 1000.4, 995.4, 950.9, 950.0, 949.5, 947.2
#> $ horaPresMax <chr> "10", "09", "00", "10", "00", "08", "00"
#> $ presMin     <dbl> 996.5, 995.8, 987.9, 947.3, 946.1, 945.5, 939.4
#> $ horaPresMin <chr> "16", "17", "17", "17", "15", "Varias", "17"

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.