Daily/annual climatology values

Get climatology values for a station or for all the available stations. Note that aemet_daily_period() and aemet_daily_period_all() are shortcuts of aemet_daily_clim().

aemet_daily_clim(
  station = "all",
  start = Sys.Date() - 7,
  end = Sys.Date(),
  verbose = FALSE,
  return_sf = FALSE
)

aemet_daily_period(
  station,
  start = 2020,
  end = 2020,
  verbose = FALSE,
  return_sf = FALSE
)

aemet_daily_period_all(
  start = 2020,
  end = 2020,
  verbose = FALSE,
  return_sf = FALSE
)

Arguments

station
Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

start, end
Character string with start and end date. See Details.

verbose
Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf
Logical. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

a tibble or a sf object

Details

start and end parameters should be:

For aemet_daily_clim(): A Date object or a string with format: %Y%m%d (2020-12-31).

For aemet_daily_period() and aemet_daily_period_all(): A string representing the year(s) to be extracted: "2020", "2018".

Examples

# Run this example only if AEMET_API_KEY is detected

if (aemet_detect_api_key()) {
  library(tibble)
  obs <- aemet_daily_clim(c("9434", "3195"))
  glimpse(obs)
}

#> Rows: 8
#> Columns: 20
#> $ fecha       <date> 2021-06-02, 2021-06-03, 2021-06-04, 2021-06-05, 2021-06-0…
#> $ indicativo  <chr> "9434", "9434", "9434", "9434", "3195", "3195", "3195", "3…
#> $ nombre      <chr> "ZARAGOZA, AEROPUERTO", "ZARAGOZA, AEROPUERTO", "ZARAGOZA,…
#> $ provincia   <chr> "ZARAGOZA", "ZARAGOZA", "ZARAGOZA", "ZARAGOZA", "MADRID", …
#> $ altitud     <dbl> 249, 249, 249, 249, 667, 667, 667, 667
#> $ tmed        <dbl> 21.2, 23.5, 18.1, 19.4, 19.0, 22.6, 22.0, 18.4
#> $ prec        <dbl> 0.0, 6.2, 0.0, 0.0, 0.0, 0.0, 0.0, 9.6
#> $ tmin        <dbl> 13.3, 17.5, 15.0, 13.3, 12.5, 16.9, 16.0, 15.4
#> $ horatmin    <chr> "Varias", "23:59", "23:59", "04:30", "05:10", "Varias", "0…
#> $ tmax        <dbl> 29.1, 29.5, 21.2, 25.5, 25.6, 28.4, 28.0, 21.3
#> $ horatmax    <time> 16:30:00, 13:50:00, 17:20:00, 15:30:00, 15:50:00, 13:50:00…
#> $ dir         <dbl> 11, 12, 30, 30, NA, NA, NA, NA
#> $ velmedia    <dbl> 3.1, 5.6, 6.1, 3.1, NA, NA, NA, NA
#> $ racha       <dbl> 8.9, 13.3, 13.6, 13.6, NA, NA, NA, NA
#> $ horaracha   <chr> "Varias", "17:10", "17:40", "00:20", NA, NA, NA, NA
#> $ sol         <dbl> 13.5, 7.5, 6.0, 7.6, NA, NA, NA, NA
#> $ presMax     <dbl> 986.1, 992.0, 994.4, 995.0, 938.9, 941.0, 944.1, 945.4
#> $ horaPresMax <chr> "24", "24", "23", "08", "24", "08", "23", "08"
#> $ presMin     <dbl> 982.8, 984.5, 989.4, 989.3, 935.5, 937.1, 940.0, 943.1
#> $ horaPresMin <chr> "Varias", "17", "02", "18", "01", "Varias", "00", "02"

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.