Daily/annual climatology values

Get climatology values for a station or for all the available stations. Note that aemet_daily_period() and aemet_daily_period_all() are shortcuts of aemet_daily_clim().

aemet_daily_clim(
  station = "all",
  start = Sys.Date() - 7,
  end = Sys.Date(),
  verbose = FALSE,
  return_sf = FALSE
)

aemet_daily_period(
  station,
  start = 2020,
  end = 2020,
  verbose = FALSE,
  return_sf = FALSE
)

aemet_daily_period_all(
  start = 2020,
  end = 2020,
  verbose = FALSE,
  return_sf = FALSE
)

Arguments

station
Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

start, end
Character string with start and end date. See Details.

verbose
Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf
Logical. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

a tibble or a sf object

Details

start and end parameters should be:

For aemet_daily_clim(): A Date object or a string with format: %Y%m%d (2020-12-31).

For aemet_daily_period() and aemet_daily_period_all(): A string representing the year(s) to be extracted: "2020", "2018".

Examples

# Run this example only if AEMET_API_KEY is detected

if (aemet_detect_api_key()) {
  library(tibble)
  obs <- aemet_daily_clim(c("9434", "3195"))
  glimpse(obs)
}

#> Rows: 6
#> Columns: 20
#> $ fecha       <date> 2021-06-16, 2021-06-17, 2021-06-18, 2021-06-16, 2021-06-1…
#> $ indicativo  <chr> "9434", "9434", "9434", "3195", "3195", "3195"
#> $ nombre      <chr> "ZARAGOZA, AEROPUERTO", "ZARAGOZA, AEROPUERTO", "ZARAGOZA,…
#> $ provincia   <chr> "ZARAGOZA", "ZARAGOZA", "ZARAGOZA", "MADRID", "MADRID", "M…
#> $ altitud     <dbl> 249, 249, 249, 667, 667, 667
#> $ tmed        <dbl> 26.0, 20.9, 22.7, 23.4, 18.0, 20.6
#> $ prec        <dbl> 1.8, 6.6, 0.0, 40.4, 5.7, 0.0
#> $ tmin        <dbl> 21.1, 16.5, 15.5, 16.4, 15.3, 15.4
#> $ horatmin    <time> 04:30:00, 23:59:00, 04:20:00, 19:30:00, 04:30:00, 01:10:00
#> $ tmax        <dbl> 31.0, 25.3, 29.9, 30.3, 20.7, 25.9
#> $ horatmax    <time> 15:40:00, 11:40:00, 15:10:00, 13:20:00, 16:10:00, 15:20:00
#> $ dir         <dbl> 12, 22, 12, NA, NA, NA
#> $ velmedia    <dbl> 4.2, 3.6, 5.0, NA, NA, NA
#> $ racha       <dbl> 13.1, 17.8, 12.5, NA, NA, NA
#> $ horaracha   <time> 16:30:00, 13:20:00, 17:20:00,       NA,       NA,       NA
#> $ sol         <dbl> 4.4, 5.9, 11.0, NA, NA, NA
#> $ presMax     <dbl> 984.5, 983.2, 985.0, 937.8, 936.3, 938.6
#> $ horaPresMax <chr> "00", "23", "08", "00", "22", "23"
#> $ presMin     <dbl> 978.2, 978.2, 981.1, 931.9, 933.5, 934.5
#> $ horaPresMin <dbl> 19, 13, 18, 17, 16, 15

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.