Daily/annual climatology values

Get climatology values for a station or for all the available stations. Note that aemet_daily_period() and aemet_daily_period_all() are shortcuts of aemet_daily_clim().

aemet_daily_clim(
  station = "all",
  start = Sys.Date() - 7,
  end = Sys.Date(),
  verbose = FALSE,
  return_sf = FALSE
)

aemet_daily_period(
  station,
  start = 2020,
  end = 2020,
  verbose = FALSE,
  return_sf = FALSE
)

aemet_daily_period_all(
  start = 2020,
  end = 2020,
  verbose = FALSE,
  return_sf = FALSE
)

Arguments

station
Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

start, end
Character string with start and end date. See Details.

verbose
Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf
Logical. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

a tibble or a sf object

Details

start and end parameters should be:

For aemet_daily_clim(): A Date object or a string with format: %Y%m%d (2020-12-31).

For aemet_daily_period() and aemet_daily_period_all(): A string representing the year(s) to be extracted: "2020", "2018".

Examples

# Run this example only if AEMET_API_KEY is detected

if (aemet_detect_api_key()) {
  library(tibble)
  obs <- aemet_daily_clim(c("9434", "3195"))
  glimpse(obs)
}

#> Rows: 8
#> Columns: 20
#> $ fecha       <date> 2021-04-21, 2021-04-22, 2021-04-23, 2021-04-24, 2021-04-2…
#> $ indicativo  <chr> "9434", "9434", "9434", "9434", "3195", "3195", "3195", "3…
#> $ nombre      <chr> "ZARAGOZA, AEROPUERTO", "ZARAGOZA, AEROPUERTO", "ZARAGOZA,…
#> $ provincia   <chr> "ZARAGOZA", "ZARAGOZA", "ZARAGOZA", "ZARAGOZA", "MADRID", …
#> $ altitud     <dbl> 249, 249, 249, 249, 667, 667, 667, 667
#> $ tmed        <dbl> 15.5, 13.2, 15.2, 15.3, 13.8, 12.6, 13.1, 12.6
#> $ prec        <dbl> 1.8, 1.2, 0.0, 6.0, 6.4, 1.6, 0.0, 6.6
#> $ tmin        <dbl> 11.4, 9.9, 8.6, 10.0, 9.4, 9.5, 9.6, 9.4
#> $ horatmin    <time> 04:00:00, 07:30:00, 05:20:00, 03:00:00, 05:00:00, 02:20:00…
#> $ tmax        <dbl> 19.6, 16.5, 21.8, 20.6, 18.3, 15.8, 16.6, 15.9
#> $ horatmax    <time> 14:30:00, 16:30:00, 14:40:00, 12:50:00, 13:00:00, 17:10:00…
#> $ dir         <dbl> 12, 13, 13, 11, NA, NA, NA, NA
#> $ velmedia    <dbl> 3.1, 4.7, 5.0, 4.4, NA, NA, NA, NA
#> $ racha       <dbl> 10.8, 10.0, 12.5, 11.9, NA, NA, NA, NA
#> $ horaracha   <time> 20:30:00, 15:10:00, 17:10:00, 11:50:00,       NA,       NA…
#> $ sol         <dbl> 5.2, 1.1, 10.4, 4.4, NA, NA, NA, NA
#> $ presMax     <dbl> 984.0, 989.5, 990.2, 989.9, 935.2, 940.2, 941.6, 940.6
#> $ horaPresMax <chr> "24", "24", "07", "07", "00", "24", "09", "07"
#> $ presMin     <dbl> 980.8, 983.5, 987.2, 986.0, 932.6, 934.0, 939.3, 937.6
#> $ horaPresMin <chr> "15", "Varias", "16", "15", "15", "03", "17", "23"

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.