Daily/annual climatology values

Get climatology values for a station or for all the available stations. Note that aemet_daily_period() and aemet_daily_period_all() are shortcuts of aemet_daily_clim().

Usage

aemet_daily_clim(
  station = "all",
  start = Sys.Date() - 7,
  end = Sys.Date(),
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)

aemet_daily_period(
  station,
  start = as.integer(format(Sys.Date(), "%Y")),
  end = start,
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)

aemet_daily_period_all(
  start = as.integer(format(Sys.Date(), "%Y")),
  end = start,
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)

Arguments

station

Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

start, end

Character string with start and end date. See Details.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf

Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

extract_metadata

Logical TRUE/FALSE. On TRUE the output is a tibble with the description of the fields. See also get_metadata_aemet().

progress

Logical, display a cli::cli_progress_bar() object. If verbose = TRUE won't be displayed.

Value

A tibble or a sf object.

Details

start and end parameters should be:

For aemet_daily_clim(): A Date object or a string with format: YYYY-MM-DD ("2020-12-31") coercible with as.Date().

For aemet_daily_period() and aemet_daily_period_all(): A string representing the year(s) to be extracted: "2020", "2018".

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_daily_clim(c("9434", "3195"))
glimpse(obs)
#> Rows: 8
#> Columns: 25
#> $ fecha       <date> 2024-06-26, 2024-06-27, 2024-06-28, 2024-06-29, 2024-06-2…
#> $ indicativo  <chr> "9434", "9434", "9434", "9434", "3195", "3195", "3195", "3…
#> $ nombre      <chr> "ZARAGOZA, AEROPUERTO", "ZARAGOZA, AEROPUERTO", "ZARAGOZA,…
#> $ provincia   <chr> "ZARAGOZA", "ZARAGOZA", "ZARAGOZA", "ZARAGOZA", "MADRID", …
#> $ altitud     <dbl> 249, 249, 249, 249, 667, 667, 667, 667
#> $ tmed        <dbl> 27.6, 28.0, 27.2, 21.8, 26.2, 26.9, 22.6, 20.4
#> $ prec        <dbl> 0.0, 0.0, 8.6, 1.2, 0.3, 0.0, 0.5, 0.0
#> $ tmin        <dbl> 20.8, 20.0, 20.1, 18.4, 21.9, 20.7, 17.2, 16.3
#> $ horatmin    <time> 04:40:00, 05:10:00, 05:00:00, 01:20:00, 05:40:00, 03:30:00…
#> $ tmax        <dbl> 34.4, 36.1, 34.4, 25.2, 30.4, 33.1, 27.9, 24.5
#> $ horatmax    <time> 14:10:00, 15:40:00, 13:50:00, 15:20:00, 12:00:00, 16:00:00…
#> $ dir         <dbl> 12, 12, 10, 25, 12, 16, 24, 99
#> $ velmedia    <dbl> 3.9, 4.2, 7.2, 5.8, 2.5, 2.5, 2.2, 4.2
#> $ racha       <dbl> 8.9, 11.7, 15.0, 15.8, 10.8, 9.2, 10.3, 9.7
#> $ horaracha   <chr> "14:10", "19:30", "15:50", "01:10", "14:10", "Varias", "16…
#> $ sol         <dbl> 8.0, 11.5, 11.7, 6.2, NA, NA, NA, NA
#> $ presMax     <dbl> 982.2, 984.3, 985.1, 985.1, 935.4, 936.7, 938.0, 938.4
#> $ horaPresMax <chr> "24", "23", "07", "23", "23", "23", "23", "22"
#> $ presMin     <dbl> 978.2, 980.1, 980.0, 980.2, 932.1, 932.9, 936.1, 936.3
#> $ horaPresMin <chr> "17", "16", "18", "00", "00", "Varias", "19", "Varias"
#> $ hrMedia     <dbl> 53, 39, 43, 60, 51, 38, 54, 68
#> $ hrMax       <dbl> 76, 77, 79, 91, 63, 66, 88, 90
#> $ horaHrMax   <time> 23:00:00, 03:30:00, 05:00:00, 10:50:00, 14:40:00, 06:00:00…
#> $ hrMin       <dbl> 33, 21, 25, 53, 37, 21, 36, 55
#> $ horaHrMin   <time> 14:10:00, 16:50:00, 13:30:00, 15:20:00, 11:50:00, 15:40:0…

# Metadata
meta <- aemet_daily_clim(c("9434", "3195"), extract_metadata = TRUE)

glimpse(meta$campos)
#> Rows: 25
#> Columns: 5
#> $ id          <chr> "fecha", "indicativo", "nombre", "provincia", "altitud", "…
#> $ descripcion <chr> "fecha del dia (AAAA-MM-DD)", "indicativo climatológico", …
#> $ tipo_datos  <chr> "string", "string", "string", "string", "float", "float", …
#> $ requerido   <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, …
#> $ unidad      <chr> NA, NA, NA, NA, "m", "°C", "mm (Ip = inferior a 0,1 mm) (A…

Site built with pkgdown 2.0.9.

Template rostemplate by dieghernan, based on Bootstrapious.