Daily/annual climatology values

Get climatology values for a station or for all the available stations. Note that aemet_daily_period() and aemet_daily_period_all() are shortcuts of aemet_daily_clim().

Usage

aemet_daily_clim(
  station = "all",
  start = Sys.Date() - 7,
  end = Sys.Date(),
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)

aemet_daily_period(
  station,
  start = as.integer(format(Sys.Date(), "%Y")),
  end = start,
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)

aemet_daily_period_all(
  start = as.integer(format(Sys.Date(), "%Y")),
  end = start,
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)

Arguments

station

Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

start, end

Character string with start and end date. See Details.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf

Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

extract_metadata

Logical TRUE/FALSE. On TRUE the output is a tibble with the description of the fields. See also get_metadata_aemet().

progress

Logical, display a cli::cli_progress_bar() object. If verbose = TRUE won't be displayed.

Value

A tibble or a sf object.

Details

start and end parameters should be:

For aemet_daily_clim(): A Date object or a string with format: YYYY-MM-DD ("2020-12-31") coercible with as.Date().

For aemet_daily_period() and aemet_daily_period_all(): A string representing the year(s) to be extracted: "2020", "2018".

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_daily_clim(c("9434", "3195"))
glimpse(obs)
#> Rows: 8
#> Columns: 25
#> $ fecha       <date> 2025-03-12, 2025-03-13, 2025-03-14, 2025-03-15, 2025-03-1…
#> $ indicativo  <chr> "9434", "9434", "9434", "9434", "3195", "3195", "3195", "3…
#> $ nombre      <chr> "ZARAGOZA, AEROPUERTO", "ZARAGOZA, AEROPUERTO", "ZARAGOZA,…
#> $ provincia   <chr> "ZARAGOZA", "ZARAGOZA", "ZARAGOZA", "ZARAGOZA", "MADRID", …
#> $ altitud     <dbl> 249, 249, 249, 249, 667, 667, 667, 667
#> $ tmed        <dbl> 12.2, 8.2, 8.2, 7.1, 8.5, 8.0, 7.2, 5.8
#> $ prec        <dbl> 0.8, 0.8, 0.0, 0.0, 0.0, 4.7, 0.0, 6.7
#> $ tmin        <dbl> 8.2, 6.2, 4.8, 2.5, 4.2, 4.3, 4.1, 1.1
#> $ horatmin    <chr> "07:10", "03:40", "23:59", "06:10", "07:00", "23:50", "Var…
#> $ tmax        <dbl> 16.2, 10.2, 11.7, 11.7, 12.8, 11.6, 10.4, 10.6
#> $ horatmax    <time> 13:40:00, 15:20:00, 15:30:00, 15:30:00, 13:20:00, 15:00:00…
#> $ dir         <chr> "32", "30", "30", "31", "23", "05", "05", "03"
#> $ velmedia    <dbl> 4.7, 3.9, 8.1, 5.6, 2.5, 1.4, 3.6, 1.9
#> $ racha       <dbl> 11.4, 11.7, 14.4, 13.9, 10.0, 10.3, 12.2, 6.7
#> $ horaracha   <time> 10:50:00, 23:59:00, 16:40:00, 09:40:00, 15:10:00, 23:10:00…
#> $ sol         <dbl> 7.3, 0.1, 4.9, 10.1, NA, NA, NA, NA
#> $ presMax     <dbl> 973.5, 970.3, 977.1, 985.8, 928.1, 923.5, 929.6, 936.1
#> $ horaPresMax <chr> "Varias", "00", "24", "24", "09", "00", "24", "Varias"
#> $ presMin     <dbl> 969.5, 966.4, 968.3, 977.1, 923.5, 915.9, 919.3, 929.6
#> $ horaPresMin <chr> "Varias", "16", "05", "00", "24", "15", "04", "00"
#> $ hrMedia     <dbl> 69, 90, 55, 57, NA, 89, 57, 60
#> $ hrMax       <dbl> 94, 96, 75, 77, NA, 98, 96, 88
#> $ horaHrMax   <time> 00:00:00, 06:50:00, 06:30:00, 06:10:00,       NA, 18:40:00…
#> $ hrMin       <dbl> 46, 72, 39, 39, NA, 63, 38, 40
#> $ horaHrMin   <time> 13:30:00, 15:40:00, 16:10:00, 14:30:00,       NA, 14:30:00…

# Metadata
meta <- aemet_daily_clim(c("9434", "3195"), extract_metadata = TRUE)

glimpse(meta$campos)
#> Rows: 25
#> Columns: 5
#> $ id          <chr> "fecha", "indicativo", "nombre", "provincia", "altitud", "…
#> $ descripcion <chr> "fecha del dia (AAAA-MM-DD)", "indicativo climatológico", …
#> $ tipo_datos  <chr> "string", "string", "string", "string", "float", "float", …
#> $ requerido   <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, …
#> $ unidad      <chr> NA, NA, NA, NA, "m", "°C", "mm (Ip = inferior a 0,1 mm) (A…

Site built with pkgdown 2.1.1.

Template rostemplate by dieghernan, based on bootstraptemple.com.