Daily/annual climatology values

Get climatology values for a station or for all the available stations. Note that aemet_daily_period() and aemet_daily_period_all() are shortcuts of aemet_daily_clim().

Usage

aemet_daily_clim(
  station = "all",
  start = Sys.Date() - 7,
  end = Sys.Date(),
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)

aemet_daily_period(
  station,
  start = as.integer(format(Sys.Date(), "%Y")),
  end = start,
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)

aemet_daily_period_all(
  start = as.integer(format(Sys.Date(), "%Y")),
  end = start,
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)

Arguments

station

Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

start, end

Character string with start and end date. See Details.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf

Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

extract_metadata

Logical TRUE/FALSE. On TRUE the output is a tibble with the description of the fields. See also get_metadata_aemet().

progress

Logical, display a cli::cli_progress_bar() object. If verbose = TRUE won't be displayed.

Value

A tibble or a sf object.

Details

start and end parameters should be:

For aemet_daily_clim(): A Date object or a string with format: YYYY-MM-DD ("2020-12-31") coercible with as.Date().

For aemet_daily_period() and aemet_daily_period_all(): A string representing the year(s) to be extracted: "2020", "2018".

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_daily_clim(c("9434", "3195"))
glimpse(obs)
#> Rows: 8
#> Columns: 25
#> $ fecha       <date> 2024-12-01, 2024-12-02, 2024-12-03, 2024-12-04, 2024-12-0…
#> $ indicativo  <chr> "9434", "9434", "9434", "9434", "3195", "3195", "3195", "3…
#> $ nombre      <chr> "ZARAGOZA, AEROPUERTO", "ZARAGOZA, AEROPUERTO", "ZARAGOZA,…
#> $ provincia   <chr> "ZARAGOZA", "ZARAGOZA", "ZARAGOZA", "ZARAGOZA", "MADRID", …
#> $ altitud     <dbl> 249, 249, 249, 249, 667, 667, 667, 667
#> $ tmed        <dbl> 5.8, 9.0, 11.8, 10.7, 12.1, 12.3, 10.6, 10.4
#> $ prec        <dbl> 0.2, 0.6, 2.6, 0.0, 0.0, 0.0, 0.5, 0.0
#> $ tmin        <dbl> 4.7, 6.2, 9.0, 7.8, 8.3, 9.5, 6.2, 6.4
#> $ horatmin    <chr> "05:20", "04:10", "01:10", "Varias", "07:40", "23:59", "07…
#> $ tmax        <dbl> 6.9, 11.7, 14.6, 13.6, 15.9, 15.1, 15.1, 14.3
#> $ horatmax    <time> 16:00:00, 15:10:00, 13:40:00, 14:40:00, 15:00:00, 14:50:00…
#> $ dir         <chr> "30", "32", "30", "30", "04", "05", "30", "34"
#> $ velmedia    <dbl> 1.4, 2.5, 5.0, 7.5, 0.6, 0.8, 1.4, 1.7
#> $ racha       <dbl> 3.9, 6.4, 10.6, 16.1, 5.0, 6.4, 9.2, 11.9
#> $ horaracha   <time> 16:30:00, 12:30:00, 12:10:00, 12:10:00, 10:50:00, 03:30:00…
#> $ sol         <dbl> 0.0, 1.8, 3.0, 8.5, NA, NA, NA, NA
#> $ presMax     <dbl> 997.4, 995.2, 992.9, 998.0, 949.1, 947.5, 944.5, 950.1
#> $ horaPresMax <chr> "Varias", "10", "10", "24", "10", "10", "11", "22"
#> $ presMin     <dbl> 994.6, 992.4, 989.7, 988.8, 945.9, 944.1, 941.8, 940.6
#> $ horaPresMin <chr> "Varias", "16", "24", "03", "16", "24", "23", "04"
#> $ hrMedia     <dbl> 100, 99, 78, 68, 81, 83, 79, 60
#> $ hrMax       <dbl> 100, 100, 99, 93, 93, 94, 96, 86
#> $ horaHrMax   <chr> "Varias", "Varias", "00:50", "Varias", "03:30", "06:50", "…
#> $ hrMin       <dbl> 100, 92, 62, 50, 66, 70, 56, 39
#> $ horaHrMin   <chr> "Varias", "15:30", "13:10", "15:00", "14:40", "14:30", "15…

# Metadata
meta <- aemet_daily_clim(c("9434", "3195"), extract_metadata = TRUE)

glimpse(meta$campos)
#> Rows: 25
#> Columns: 5
#> $ id          <chr> "fecha", "indicativo", "nombre", "provincia", "altitud", "…
#> $ descripcion <chr> "fecha del dia (AAAA-MM-DD)", "indicativo climatológico", …
#> $ tipo_datos  <chr> "string", "string", "string", "string", "float", "float", …
#> $ requerido   <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, …
#> $ unidad      <chr> NA, NA, NA, NA, "m", "°C", "mm (Ip = inferior a 0,1 mm) (A…

Site built with pkgdown 2.1.1.

Template rostemplate by dieghernan, based on Bootstrapious.