Last observation values for a station

Get last observation values for a station.

aemet_last_obs(station = "all", verbose = FALSE, return_sf = FALSE)

Arguments

station
Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose
Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf
Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

A tibble or a sf object.

API Key

You need to set your API Key globally using aemet_api_key().

Examples

# Run this example only if AEMET_API_KEY is set

if (aemet_detect_api_key()) {
  library(tibble)
  obs <- aemet_last_obs(c("9434", "3195"))
  glimpse(obs)
}

#> Rows: 46
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943~
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00~
#> $ fint      <dttm> 2021-08-26 19:00:00, 2021-08-26 20:00:00, 2021-08-26 21:00:~
#> $ prec      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, ~
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, ~
#> $ vmax      <dbl> 5.7, 3.6, 5.8, 7.8, 8.1, 7.1, 7.6, 5.8, 8.6, 7.9, 8.2, 6.2, ~
#> $ vv        <dbl> 2.7, 2.2, 4.4, 6.0, 4.2, 4.0, 3.2, 4.5, 6.5, 5.5, 5.2, 3.4, ~
#> $ dv        <dbl> 271, 297, 286, 299, 283, 304, 278, 296, 291, 295, 308, 315, ~
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, ~
#> $ dmax      <dbl> 308, 238, 295, 305, 303, 300, 305, 290, 295, 290, 300, 308, ~
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A~
#> $ pres      <dbl> 982.6, 982.9, 983.5, 984.2, 984.5, 984.6, 984.8, 984.8, 984.~
#> $ hr        <dbl> 33, 33, 35, 38, 48, 57, 62, 68, 68, 67, 66, 70, 61, 55, 46, ~
#> $ stdvv     <dbl> 0.5, 0.3, 0.6, 0.6, 0.4, 0.4, 0.6, 0.4, 0.5, 0.5, 0.6, 0.4, ~
#> $ ts        <dbl> 30.3, 27.7, 27.5, 27.2, 25.4, 24.8, 23.6, 22.9, 22.8, 22.4, ~
#> $ pres_nmar <dbl> 1010.5, 1011.0, 1011.7, 1012.5, 1013.0, 1013.1, 1013.4, 1013~
#> $ tamin     <dbl> 31.3, 29.0, 27.8, 27.3, 25.8, 25.0, 24.1, 23.1, 22.6, 22.2, ~
#> $ ta        <dbl> 31.3, 29.0, 28.1, 27.4, 25.8, 25.2, 24.1, 23.1, 22.6, 22.2, ~
#> $ tamax     <dbl> 33.0, 31.3, 29.0, 28.2, 27.5, 25.9, 25.2, 24.1, 23.1, 22.6, ~
#> $ tpr       <dbl> 13.1, 11.1, 11.2, 11.8, 13.9, 16.0, 16.4, 16.9, 16.4, 15.8, ~
#> $ stddv     <dbl> 8, 6, 6, 9, 5, 7, 13, 8, 4, 6, 6, 8, 5, 8, 7, 8, 10, 11, 12,~
#> $ inso      <dbl> 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 7.8, ~
#> $ tss5cm    <dbl> 34.7, 33.5, 32.5, 31.7, 31.1, 30.6, 30.1, 29.5, 29.1, 28.6, ~
#> $ pacutp    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, ~
#> $ tss20cm   <dbl> 32.6, 32.7, 32.6, 32.5, 32.3, 32.1, 31.8, 31.6, 31.3, 31.1, ~
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, ~

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.