Last observation values for a station

Get last observation values for a station.

Usage,

aemet_last_obs(station = "all", verbose = FALSE, return_sf = FALSE)

Arguments

station

Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf

Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

A tibble or a sf object.

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_last_obs(c("9434", "3195"))
glimpse(obs)
#> Rows: 48
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943…
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00…
#> $ fint      <dttm> 2021-12-01 17:00:00, 2021-12-01 18:00:00, 2021-12-01 19:00:…
#> $ prec      <dbl> 0.2, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, …
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, …
#> $ vmax      <dbl> 5.6, 7.6, 4.5, 6.4, 4.1, 3.9, 4.8, 5.5, 5.0, 7.0, 9.1, 6.8, …
#> $ vv        <dbl> 3.9, 0.8, 2.3, 2.7, 3.3, 2.5, 3.2, 2.5, 3.5, 4.5, 4.9, 4.8, …
#> $ dv        <dbl> 270, 10, 303, 314, 292, 302, 251, 271, 304, 297, 312, 303, 3…
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, …
#> $ dmax      <dbl> 275, 318, 303, 285, 305, 313, 290, 298, 305, 298, 298, 305, …
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A…
#> $ pres      <dbl> 983.9, 984.1, 983.7, 983.4, 983.0, 982.4, 981.8, 981.1, 980.…
#> $ hr        <dbl> 88, 90, 91, 90, 89, 88, 88, 86, 87, 84, 82, 83, 82, 83, 79, …
#> $ stdvv     <dbl> 0.6, 0.3, 0.6, 0.4, 0.3, 0.5, 0.2, 0.4, 0.4, 0.4, 0.6, 0.6, …
#> $ ts        <dbl> 8.1, 7.6, 7.7, 7.4, 6.1, 6.0, 4.7, 5.1, 5.5, 5.5, 5.6, 5.0, …
#> $ pres_nmar <dbl> 1014.2, 1014.5, 1014.1, 1013.8, 1013.5, 1012.9, 1012.4, 1011…
#> $ tamin     <dbl> 8.3, 8.3, 8.0, 7.9, 7.0, 6.9, 5.7, 5.7, 6.0, 6.1, 6.5, 6.1, …
#> $ ta        <dbl> 8.6, 8.3, 8.1, 7.9, 7.0, 7.0, 5.7, 6.3, 6.3, 6.7, 6.8, 6.2, …
#> $ tamax     <dbl> 8.6, 9.2, 8.4, 8.9, 8.0, 7.3, 7.0, 6.7, 6.3, 6.8, 7.0, 6.8, …
#> $ tpr       <dbl> 6.7, 6.7, 6.7, 6.4, 5.3, 5.1, 3.9, 4.1, 4.3, 4.1, 3.9, 3.6, …
#> $ stddv     <dbl> 8, 29, 17, 9, 4, 13, 6, 9, 6, 5, 7, 8, 5, 9, 9, 8, 8, 9, 10,…
#> $ inso      <dbl> 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, …
#> $ tss5cm    <dbl> 8.6, 8.4, 8.2, 8.0, 7.7, 7.3, 7.0, 6.7, 6.4, 6.3, 6.3, 6.1, …
#> $ pacutp    <dbl> 0.18, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, …
#> $ tss20cm   <dbl> 8.6, 8.7, 8.7, 8.7, 8.7, 8.7, 8.6, 8.6, 8.4, 8.3, 8.2, 8.1, …
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …

Site built with pkgdown 2.0.0.

Template by Bootstrapious . Ported to pkgdown by dieghernan.