Last observation values for a station

Get last observation values for a station.

aemet_last_obs(station = "all", verbose = FALSE, return_sf = FALSE)

Arguments

station
Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose
Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf
Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

A tibble or a sf object.

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

# Run this example only if AEMET_API_KEY is set

if (aemet_detect_api_key()) {
  library(tibble)
  obs <- aemet_last_obs(c("9434", "3195"))
  glimpse(obs)
}

#> Rows: 34
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943…
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00…
#> $ fint      <dttm> 2021-09-14 13:00:00, 2021-09-14 14:00:00, 2021-09-14 15:00:…
#> $ prec      <dbl> 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 2.2, 0.8, …
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 667, 667, …
#> $ vmax      <dbl> 4.0, 6.1, 5.1, 2.7, 2.4, 1.6, 2.7, 2.8, 3.2, 3.9, NA, NA, NA…
#> $ vv        <dbl> 1.6, 3.8, 1.3, 1.5, 0.4, 0.5, 1.8, 1.0, 1.7, 3.3, NA, NA, NA…
#> $ dv        <dbl> 177, 279, 320, 315, 32, 306, 248, 280, 234, 302, NA, NA, NA,…
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, …
#> $ dmax      <dbl> 165, 283, 270, 320, 330, 40, 250, 245, 233, 305, NA, NA, NA,…
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A…
#> $ pres      <dbl> 985.9, 985.8, 984.9, 984.8, 984.4, 984.5, 984.6, 984.2, 984.…
#> $ hr        <dbl> 70, 66, 60, 61, 66, 61, 74, 78, 76, 74, 94, 93, 91, 88, 94, …
#> $ stdvv     <dbl> 0.5, 0.8, 0.3, 0.4, 0.2, 0.2, 0.2, 0.3, 0.3, 0.2, NA, NA, NA…
#> $ ts        <dbl> 23.1, 23.1, 24.0, 22.3, 22.1, 21.4, 21.1, 20.2, 20.3, 20.2, …
#> $ pres_nmar <dbl> 1014.6, 1014.6, 1013.7, 1013.6, 1013.2, 1013.4, 1013.5, 1013…
#> $ tamin     <dbl> 23.2, 23.0, 22.9, 22.3, 22.3, 22.0, 21.6, 20.9, 20.6, 20.5, …
#> $ ta        <dbl> 23.2, 23.0, 23.0, 22.5, 22.3, 22.0, 21.6, 20.9, 20.7, 20.7, …
#> $ tamax     <dbl> 24.0, 23.2, 23.1, 23.0, 22.6, 22.3, 22.1, 21.6, 20.9, 20.7, …
#> $ tpr       <dbl> 17.5, 16.3, 14.8, 14.6, 15.6, 14.1, 16.7, 16.9, 16.3, 15.9, …
#> $ stddv     <dbl> 14, 7, 84, 11, 56, 26, 10, 21, 10, 8, NA, NA, NA, NA, NA, NA…
#> $ inso      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ tss5cm    <dbl> 25.2, 25.0, 24.9, 24.6, 24.3, 24.2, 23.9, 23.5, 23.2, 23.0, …
#> $ pacutp    <dbl> 0.05, 0.01, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, …
#> $ tss20cm   <dbl> 26.0, 26.0, 26.0, 26.0, 25.9, 25.8, 25.7, 25.6, 25.5, 25.4, …
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 0, 0, 0, 0, 0, 0, 0,…

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.