Last observation values for a station

Get last observation values for a station.

aemet_last_obs(station = "all", verbose = FALSE, return_sf = FALSE)

Arguments

station
Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose
Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf
Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

A tibble or a sf object.

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_last_obs(c("9434", "3195"))
glimpse(obs)
#> Rows: 48
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943…
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00…
#> $ fint      <dttm> 2021-11-16 13:00:00, 2021-11-16 14:00:00, 2021-11-16 15:00:…
#> $ prec      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, …
#> $ vmax      <dbl> 11.6, 12.2, 12.8, 13.3, 12.3, 9.9, 9.2, 9.3, 9.9, 10.3, 10.3…
#> $ vv        <dbl> 7.2, 7.6, 8.1, 9.1, 5.8, 7.2, 6.5, 7.5, 7.0, 7.4, 5.6, 6.3, …
#> $ dv        <dbl> 321, 320, 322, 315, 314, 303, 293, 297, 301, 293, 297, 299, …
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, …
#> $ dmax      <dbl> 330, 323, 318, 323, 320, 320, 295, 303, 278, 300, 300, 300, …
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A…
#> $ pres      <dbl> 989.7, 989.4, 989.6, 989.5, 989.9, 990.1, 990.4, 990.8, 991.…
#> $ hr        <dbl> 48, 48, 50, 53, 57, 61, 63, 66, 67, 71, 70, 71, 76, 80, 76, …
#> $ stdvv     <dbl> 1.6, 1.4, 1.6, 1.4, 1.1, 0.8, 0.8, 0.6, 0.6, 0.7, 0.9, 1.2, …
#> $ ts        <dbl> 17.6, 16.3, 15.6, 13.1, 11.0, 10.4, 9.7, 9.1, 8.8, 8.2, 8.2,…
#> $ pres_nmar <dbl> 1019.6, 1019.3, 1019.6, 1019.5, 1020.1, 1020.4, 1020.8, 1021…
#> $ tamin     <dbl> 13.7, 14.2, 13.7, 13.0, 11.6, 10.7, 10.1, 9.6, 9.4, 8.6, 8.8…
#> $ ta        <dbl> 14.3, 14.2, 13.7, 13.0, 11.6, 10.7, 10.1, 9.6, 9.4, 8.8, 9.0…
#> $ tamax     <dbl> 19.1, 14.5, 14.2, 13.7, 13.0, 11.6, 10.7, 10.2, 9.6, 9.4, 9.…
#> $ tpr       <dbl> 3.4, 3.4, 3.4, 3.6, 3.4, 3.4, 3.4, 3.6, 3.6, 3.8, 3.8, 3.9, …
#> $ stddv     <dbl> 12, 10, 10, 8, 8, 6, 5, 4, 6, 5, 10, 9, 7, 4, 8, 9, 7, 4, 5,…
#> $ inso      <dbl> 60.0, 60.0, 60.0, 60.0, 27.1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, …
#> $ tss5cm    <dbl> 14.1, 14.7, 14.5, 14.2, 13.6, 12.9, 12.5, 12.1, 11.7, 11.3, …
#> $ pacutp    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ tss20cm   <dbl> 12.3, 12.6, 12.9, 13.1, 13.3, 13.4, 13.4, 13.3, 13.2, 13.1, …
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.