Last observation values for a station

Get last observation values for a station.

Usage,

aemet_last_obs(station = "all", verbose = FALSE, return_sf = FALSE)

Arguments

station

Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf

Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

A tibble or a sf object.

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_last_obs(c("9434", "3195"))
glimpse(obs)
#> Rows: 48
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943…
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00…
#> $ fint      <dttm> 2021-11-30 13:00:00, 2021-11-30 14:00:00, 2021-11-30 15:00:…
#> $ prec      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, …
#> $ vmax      <dbl> 10.7, 10.2, 9.0, 8.4, 6.9, 5.7, 4.0, 6.8, 6.7, 3.9, 2.2, 2.7…
#> $ vv        <dbl> 7.0, 6.7, 5.6, 5.3, 4.1, 3.4, 3.3, 1.7, 3.8, 1.8, 1.4, 1.6, …
#> $ dv        <dbl> 308, 313, 316, 301, 266, 261, 253, 262, 245, 278, 289, 291, …
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, …
#> $ dmax      <dbl> 315, 315, 318, 315, 300, 283, 238, 253, 233, 250, 283, 298, …
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A…
#> $ pres      <dbl> 995.6, 995.2, 995.2, 995.3, 995.3, 995.5, 995.5, 995.5, 995.…
#> $ hr        <dbl> 59, 59, 60, 63, 74, 77, 83, 84, 83, 87, 89, 90, 91, 93, 92, …
#> $ stdvv     <dbl> 1.1, 1.1, 1.0, 0.8, 0.4, 0.3, 0.3, 0.3, 0.6, 0.2, 0.2, 0.4, …
#> $ ts        <dbl> 15.7, 15.5, 14.9, 12.2, 9.6, 8.6, 7.2, 6.0, 6.8, 5.3, 4.2, 4…
#> $ pres_nmar <dbl> 1025.7, 1025.3, 1025.3, 1025.5, 1025.7, 1026.1, 1026.3, 1026…
#> $ tamin     <dbl> 12.7, 13.6, 13.8, 13.0, 10.9, 9.7, 8.0, 7.3, 7.0, 6.5, 5.7, …
#> $ ta        <dbl> 13.7, 14.0, 13.8, 13.0, 10.9, 9.7, 8.1, 7.3, 7.6, 6.5, 5.7, …
#> $ tamax     <dbl> 13.9, 14.5, 14.0, 13.8, 13.0, 10.9, 9.7, 8.3, 7.6, 7.6, 6.5,…
#> $ tpr       <dbl> 5.8, 6.1, 6.1, 6.1, 6.4, 5.9, 5.5, 4.8, 5.0, 4.5, 3.9, 3.8, …
#> $ stddv     <dbl> 11, 11, 10, 9, 7, 6, 7, 17, 9, 6, 14, 11, 25, 12, 30, 12, 43…
#> $ inso      <dbl> 60.0, 60.0, 60.0, 60.0, 12.7, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, …
#> $ tss5cm    <dbl> 11.3, 11.7, 11.3, 11.0, 10.3, 9.7, 9.1, 8.6, 8.2, 7.8, 7.4, …
#> $ pacutp    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ tss20cm   <dbl> 9.0, 9.3, 9.6, 9.8, 9.9, 10.0, 10.0, 9.9, 9.8, 9.6, 9.5, 9.3…
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …

Site built with pkgdown .

Template by Bootstrapious . Ported to pkgdown by dieghernan.