Last observation values for a station

Get last observation values for a station.

aemet_last_obs(station = "all", verbose = FALSE, return_sf = FALSE)

Arguments

station
Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose
Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf
Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

A tibble or a sf object.

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_last_obs(c("9434", "3195"))
glimpse(obs)

#> Rows: 48
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943…
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00…
#> $ fint      <dttm> 2021-10-05 13:00:00, 2021-10-05 14:00:00, 2021-10-05 15:00:…
#> $ prec      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, …
#> $ vmax      <dbl> 6.4, 6.8, 8.4, 7.9, 8.4, 7.0, 5.4, 6.7, 7.5, 9.1, 8.9, 8.9, …
#> $ vv        <dbl> 3.8, 4.1, 5.3, 4.9, 4.9, 4.3, 3.5, 4.5, 5.0, 6.2, 6.4, 6.1, …
#> $ dv        <dbl> 311, 311, 316, 320, 308, 306, 302, 298, 306, 310, 307, 302, …
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, …
#> $ dmax      <dbl> 325, 303, 315, 303, 328, 310, 303, 310, 298, 305, 303, 305, …
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A…
#> $ pres      <dbl> 989.2, 989.1, 989.2, 989.5, 989.8, 990.5, 991.2, 992.2, 992.…
#> $ hr        <dbl> 48, 45, 46, 53, 54, 57, 62, 65, 67, 65, 68, 67, 69, 69, 70, …
#> $ stdvv     <dbl> 0.7, 0.8, 0.7, 1.0, 0.7, 0.4, 0.5, 0.5, 0.7, 1.0, 0.6, 0.7, …
#> $ ts        <dbl> 25.3, 28.9, 26.7, 23.9, 21.4, 20.1, 19.1, 18.0, 17.8, 17.6, …
#> $ pres_nmar <dbl> 1018.3, 1018.1, 1018.2, 1018.5, 1018.9, 1019.7, 1020.5, 1021…
#> $ tamin     <dbl> 20.4, 21.4, 22.3, 21.8, 20.9, 20.5, 19.7, 18.5, 18.1, 17.7, …
#> $ ta        <dbl> 21.4, 22.4, 22.3, 21.8, 21.0, 20.5, 19.7, 18.5, 18.1, 17.8, …
#> $ tamax     <dbl> 21.5, 22.4, 22.7, 22.3, 21.8, 21.1, 20.5, 19.7, 18.5, 18.1, …
#> $ tpr       <dbl> 9.9, 9.9, 10.2, 11.8, 11.3, 11.7, 12.2, 11.8, 11.9, 11.1, 11…
#> $ stddv     <dbl> 10, 12, 9, 11, 7, 7, 15, 8, 8, 9, 6, 7, 6, 8, 8, 10, 10, 9, …
#> $ inso      <dbl> 60.0, 60.0, 58.4, 18.0, 1.3, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0…
#> $ tss5cm    <dbl> 20.8, 22.0, 22.8, 22.3, 21.7, 20.8, 20.3, 19.8, 19.3, 19.0, …
#> $ pacutp    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ tss20cm   <dbl> 17.6, 18.1, 18.6, 19.1, 19.5, 19.8, 20.0, 20.0, 20.0, 19.9, …
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.