Last observation values for a station

Get last observation values for a station.

aemet_last_obs(station = "all", verbose = FALSE, return_sf = FALSE)

Arguments

station

Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf

Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

A tibble or a sf object.

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_last_obs(c("9434", "3195"))
glimpse(obs)
#> Rows: 48
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943…
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00…
#> $ fint      <dttm> 2021-12-07 13:00:00, 2021-12-07 14:00:00, 2021-12-07 15:00:…
#> $ prec      <dbl> 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.4, 0.0, …
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, …
#> $ vmax      <dbl> 5.8, 5.2, 5.3, 4.2, 4.7, 9.2, 3.3, 3.3, 3.8, 7.5, 16.8, 10.7…
#> $ vv        <dbl> 2.0, 2.2, 1.9, 2.3, 1.7, 2.2, 1.5, 1.9, 2.1, 5.0, 2.8, 5.7, …
#> $ dv        <dbl> 102, 159, 89, 96, 133, 188, 175, 96, 99, 264, 246, 264, 322,…
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, …
#> $ dmax      <dbl> 90, 128, 220, 105, 95, 238, 88, 90, 198, 263, 290, 260, 355,…
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A…
#> $ pres      <dbl> 985.4, 984.0, 983.3, 982.6, 982.2, 981.5, 980.5, 979.7, 979.…
#> $ hr        <dbl> 89, 87, 85, 84, 85, 74, 78, 82, 85, 81, 83, 69, 68, 55, 60, …
#> $ stdvv     <dbl> 0.6, 0.5, 0.6, 0.9, 0.5, 1.0, 0.5, 0.3, 0.6, 0.9, 1.4, 1.3, …
#> $ ts        <dbl> 9.6, 9.6, 10.0, 10.5, 9.8, 11.7, 9.8, 9.0, 9.4, 10.5, 9.6, 1…
#> $ pres_nmar <dbl> 1015.6, 1014.2, 1013.4, 1012.6, 1012.3, 1011.3, 1010.4, 1009…
#> $ tamin     <dbl> 9.4, 9.7, 10.0, 10.1, 10.2, 10.3, 11.3, 10.6, 10.0, 9.7, 10.…
#> $ ta        <dbl> 9.8, 10.0, 10.3, 10.8, 10.3, 12.4, 11.4, 10.6, 10.0, 11.5, 1…
#> $ tamax     <dbl> 9.8, 10.0, 10.7, 11.5, 10.8, 12.5, 12.4, 11.6, 10.6, 11.5, 1…
#> $ tpr       <dbl> 8.1, 8.0, 7.8, 8.3, 7.8, 7.8, 7.7, 7.7, 7.6, 8.4, 8.0, 6.5, …
#> $ stddv     <dbl> 24, 18, 59, 20, 18, 47, 34, 18, 19, 7, 20, 11, 17, 10, 6, 7,…
#> $ inso      <dbl> 0.0, 0.0, 0.0, 26.8, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,…
#> $ tss5cm    <dbl> 8.3, 8.3, 8.3, 8.6, 8.6, 8.6, 8.6, 8.4, 8.3, 8.4, 8.6, 8.6, …
#> $ pacutp    <dbl> 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, …
#> $ tss20cm   <dbl> 8.1, 8.2, 8.3, 8.4, 8.4, 8.5, 8.6, 8.6, 8.7, 8.7, 8.7, 8.7, …
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …

Site built with pkgdown 2.0.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.