Last observation values for a station

Get last observation values for a station.

Usage

aemet_last_obs(station = "all", verbose = FALSE, return_sf = FALSE)

Arguments

station

Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf

Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

A tibble or a sf object.

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_last_obs(c("9434", "3195"))
glimpse(obs)
#> Rows: 46
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943…
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00…
#> $ fint      <dttm> 2023-05-24 08:00:00, 2023-05-24 09:00:00, 2023-05-24 10:00:…
#> $ prec      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, …
#> $ vmax      <dbl> 7.6, 8.1, 8.3, 8.2, 7.3, 6.7, 6.4, 11.7, 12.3, 10.6, 8.2, 8.…
#> $ vv        <dbl> 5.6, 5.1, 4.1, 4.2, 3.1, 2.0, 4.6, 7.7, 7.2, 5.9, 5.0, 4.9, …
#> $ dv        <dbl> 308, 315, 313, 310, 304, 249, 301, 323, 326, 314, 314, 323, …
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, …
#> $ dmax      <dbl> 310, 298, 328, 315, 305, 300, 303, 313, 318, 308, 318, 313, …
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A…
#> $ pres      <dbl> 987.0, 986.9, 986.4, 986.1, 985.5, 984.7, 984.0, 983.7, 983.…
#> $ hr        <dbl> 74, 65, 61, 54, 46, 38, 38, 41, 41, 45, 49, 52, 59, 62, 64, …
#> $ stdvv     <dbl> 0.6, 0.7, 1.1, 0.7, 0.7, 1.0, 0.8, 1.5, 1.6, 0.8, 1.1, 0.9, …
#> $ ts        <dbl> 20.6, 22.7, 25.6, 27.7, 29.7, 30.2, 32.3, 28.8, 29.4, 25.2, …
#> $ pres_nmar <dbl> 1016.5, 1016.2, 1015.5, 1015.1, 1014.3, 1013.3, 1012.6, 1012…
#> $ tamin     <dbl> 16.6, 18.0, 19.5, 21.0, 22.6, 24.2, 25.6, 24.9, 25.2, 23.4, …
#> $ ta        <dbl> 18.0, 19.8, 21.0, 22.6, 24.3, 25.6, 26.0, 25.2, 25.6, 23.4, …
#> $ tamax     <dbl> 18.0, 22.3, 21.1, 22.6, 24.3, 25.6, 26.3, 26.8, 26.4, 25.6, …
#> $ tpr       <dbl> 13.3, 13.1, 13.2, 12.8, 11.9, 10.2, 10.5, 11.0, 11.3, 10.8, …
#> $ stddv     <dbl> 9, 10, 13, 14, 22, 43, 14, 10, 12, 9, 10, 11, 12, 9, 9, 6, 7…
#> $ inso      <dbl> 60.0, 60.0, 60.0, 60.0, 60.0, 60.0, 60.0, 41.2, 60.0, 34.7, …
#> $ tss5cm    <dbl> 19.7, 21.1, 23.3, 25.6, 28.1, 30.3, 31.9, 31.8, 31.3, 30.3, …
#> $ pacutp    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ tss20cm   <dbl> 21.5, 21.4, 21.4, 21.5, 21.8, 22.2, 22.8, 23.5, 24.2, 24.7, …
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …

Site built with pkgdown 2.0.7.

Template rostemplate by dieghernan, based on Bootstrapious.