Last observation values for a station

Get last observation values for a station.

aemet_last_obs(station = "all", verbose = FALSE, return_sf = FALSE)

Arguments

station
Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose
Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf
Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

A tibble or a sf object.

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_last_obs(c("9434", "3195"))
glimpse(obs)
#> Rows: 48
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943…
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00…
#> $ fint      <dttm> 2021-11-02 13:00:00, 2021-11-02 14:00:00, 2021-11-02 15:00:…
#> $ prec      <dbl> 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.6, 0.8, 0.2, …
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, …
#> $ vmax      <dbl> 14.1, 11.1, 7.7, 7.2, 11.3, 9.8, 8.9, 7.3, 11.8, 17.1, 10.6,…
#> $ vv        <dbl> 8.4, 6.5, 3.6, 4.8, 5.2, 5.1, 1.7, 4.3, 8.8, 8.5, 5.5, 4.6, …
#> $ dv        <dbl> 272, 277, 260, 245, 241, 223, 205, 180, 250, 274, 304, 300, …
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, …
#> $ dmax      <dbl> 275, 288, 278, 250, 218, 215, 178, 175, 248, 265, 275, 298, …
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A…
#> $ pres      <dbl> 979.3, 978.3, 977.8, 977.1, 976.6, 976.6, 976.2, 975.9, 975.…
#> $ hr        <dbl> 43, 41, 45, 44, 46, 47, 50, 53, 55, 76, 85, 90, 88, 83, 82, …
#> $ stdvv     <dbl> 1.1, 1.0, 0.6, 0.7, 0.9, 1.1, 0.4, 1.1, 1.3, 1.0, 0.7, 0.8, …
#> $ ts        <dbl> 18.7, 18.1, 17.6, 17.8, 16.3, 15.9, 14.6, 14.6, 14.5, 12.0, …
#> $ pres_nmar <dbl> 1008.5, 1007.5, 1007.0, 1006.3, 1005.8, 1005.9, 1005.5, 1005…
#> $ tamin     <dbl> 17.5, 17.6, 17.6, 17.3, 17.0, 16.5, 15.9, 15.5, 14.8, 12.8, …
#> $ ta        <dbl> 17.6, 17.8, 17.7, 17.7, 17.0, 16.6, 15.9, 15.5, 14.9, 12.8, …
#> $ tamax     <dbl> 17.9, 18.0, 17.8, 17.7, 17.8, 17.0, 16.6, 16.2, 15.5, 15.0, …
#> $ tpr       <dbl> 4.8, 4.3, 5.6, 5.3, 5.3, 5.3, 5.5, 5.9, 5.9, 8.7, 8.9, 8.8, …
#> $ stddv     <dbl> 8, 10, 9, 7, 11, 15, 27, 20, 8, 9, 7, 6, 8, 7, 7, 5, 8, 8, 5…
#> $ inso      <dbl> 31.0, 6.3, 0.0, 11.9, 4.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0…
#> $ tss5cm    <dbl> 15.4, 15.7, 15.6, 15.2, 14.9, 14.5, 14.2, 13.9, 13.7, 13.5, …
#> $ pacutp    <dbl> 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.70, …
#> $ tss20cm   <dbl> 14.6, 14.8, 15.1, 15.2, 15.3, 15.4, 15.4, 15.4, 15.3, 15.3, …
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.