Last observation values for a station

Get last observation values for a station.

aemet_last_obs(station = "all", verbose = FALSE, return_sf = FALSE)

Arguments

station
Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose
Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf
Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

A tibble or a sf object.

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_last_obs(c("9434", "3195"))
glimpse(obs)
#> Rows: 48
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943…
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00…
#> $ fint      <dttm> 2021-11-23 13:00:00, 2021-11-23 14:00:00, 2021-11-23 15:00:…
#> $ prec      <dbl> 6.0, 1.2, 0.6, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, …
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, …
#> $ vmax      <dbl> 5.5, 5.4, 7.6, 5.3, 4.0, 3.5, 4.6, 7.1, 5.3, 4.1, 5.1, 4.0, …
#> $ vv        <dbl> 4.1, 2.9, 2.7, 3.2, 2.0, 2.7, 2.9, 2.6, 2.0, 2.7, 2.1, 3.2, …
#> $ dv        <dbl> 276, 255, 285, 281, 319, 302, 285, 294, 239, 260, 320, 297, …
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, …
#> $ dmax      <dbl> 275, 278, 223, 265, 295, 303, 280, 235, 248, 263, 260, 288, …
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A…
#> $ pres      <dbl> 981.0, 981.2, 981.7, 982.0, 982.7, 982.7, 983.3, 983.2, 983.…
#> $ hr        <dbl> 95, 93, 94, 89, 92, 92, 86, 85, 86, 86, 87, 89, 88, 89, 91, …
#> $ stdvv     <dbl> 0.5, 0.6, 0.4, 0.4, 0.5, 0.3, 0.4, 0.8, 0.3, 0.5, 0.4, 0.3, …
#> $ ts        <dbl> 8.4, 7.8, 8.4, 7.8, 7.4, 7.1, 6.8, 6.1, 6.4, 6.6, 6.9, 7.2, …
#> $ pres_nmar <dbl> 1011.3, 1011.5, 1012.0, 1012.4, 1013.1, 1013.1, 1013.8, 1013…
#> $ tamin     <dbl> 8.1, 7.9, 7.8, 7.8, 7.4, 7.3, 7.1, 6.8, 6.9, 6.9, 6.9, 7.0, …
#> $ ta        <dbl> 8.1, 7.9, 8.0, 7.8, 7.4, 7.3, 7.2, 6.9, 7.0, 6.9, 7.0, 7.2, …
#> $ tamax     <dbl> 8.6, 8.1, 8.0, 8.1, 7.9, 7.5, 7.4, 7.2, 7.4, 7.0, 7.1, 7.2, …
#> $ tpr       <dbl> 7.3, 6.8, 7.1, 6.1, 6.2, 6.1, 5.0, 4.5, 4.8, 4.6, 5.0, 5.5, …
#> $ stddv     <dbl> 7, 12, 11, 6, 14, 8, 8, 16, 9, 10, 12, 5, 7, 5, 6, 6, 6, 8, …
#> $ inso      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ tss5cm    <dbl> 9.2, 9.3, 9.3, 9.3, 9.2, 9.1, 8.9, 8.6, 8.5, 8.4, 8.3, 8.4, …
#> $ pacutp    <dbl> 5.63, 1.15, 0.61, 0.00, 0.01, 0.00, 0.00, 0.00, 0.00, 0.00, …
#> $ tss20cm   <dbl> 10.2, 10.2, 10.2, 10.2, 10.2, 10.2, 10.2, 10.1, 10.1, 10.0, …
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.