Last observation values for a station

Get last observation values for a station.

Usage

aemet_last_obs(
  station = "all",
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)

Arguments

station

Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf

Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

extract_metadata

Logical TRUE/FALSE. On TRUE the output is a tibble with the description of the fields. See also get_metadata_aemet().

progress

Logical, display a cli::cli_progress_bar() object. If verbose = TRUE won't be displayed.

Value

A tibble or a sf object

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_last_obs(c("9434", "3195"))
glimpse(obs)
#> Rows: 24
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943…
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00…
#> $ fint      <dttm> 2024-06-23 00:00:00, 2024-06-23 01:00:00, 2024-06-23 02:00:…
#> $ prec      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, …
#> $ vmax      <dbl> 11.0, 11.8, 12.2, 12.6, 13.2, 12.5, 15.5, 15.4, 16.1, 16.3, …
#> $ vv        <dbl> 8.4, 8.5, 8.0, 9.3, 10.1, 7.4, 11.5, 11.7, 10.5, 10.7, 11.6,…
#> $ dv        <dbl> 305, 296, 304, 301, 302, 292, 297, 305, 312, 315, 316, 314, …
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, …
#> $ dmax      <dbl> 318, 290, 298, 305, 303, 305, 298, 303, 308, 310, 310, 313, …
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A…
#> $ pres      <dbl> 991.1, 990.6, 990.1, 989.5, 989.3, 989.3, 989.2, 989.0, 988.…
#> $ hr        <dbl> 60, 62, 62, 65, 66, 69, 66, 64, 60, 57, 52, 48, 45, 49, 53, …
#> $ stdvv     <dbl> 1.0, 1.0, 1.4, 1.4, 0.9, 0.6, 1.6, 1.1, 1.6, 1.8, 2.0, 1.7, …
#> $ ts        <dbl> 18.7, 18.1, 17.8, 17.3, 17.4, 17.3, 20.1, 23.7, 26.6, 28.6, …
#> $ pres_nmar <dbl> 1020.7, 1020.2, 1019.7, 1019.1, 1018.9, 1018.9, 1018.7, 1018…
#> $ tamin     <dbl> 18.5, 17.9, 17.8, 17.2, 17.0, 17.0, 17.1, 18.4, 19.6, 21.0, …
#> $ ta        <dbl> 18.5, 18.0, 17.8, 17.2, 17.4, 17.1, 18.4, 19.6, 21.0, 22.3, …
#> $ tamax     <dbl> 18.7, 18.5, 18.1, 17.8, 17.5, 17.4, 18.4, 19.6, 21.0, 22.3, …
#> $ tpr       <dbl> 10.5, 10.6, 10.4, 10.5, 11.0, 11.3, 11.9, 12.6, 13.0, 13.4, …
#> $ stddv     <dbl> 6, 6, 11, 8, 6, 5, 6, 6, 9, 9, 10, 9, NA, NA, NA, NA, NA, NA…
#> $ inso      <dbl> 0.0, 0.0, 0.0, 0.0, 0.0, 4.9, 60.0, 60.0, 60.0, 60.0, 60.0, …
#> $ tss5cm    <dbl> 26.2, 25.7, 25.1, 24.6, 24.1, 23.8, 23.5, 23.9, 24.9, 26.4, …
#> $ pacutp    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, NA, NA, NA, NA, NA, NA, …
#> $ tss20cm   <dbl> 28.6, 28.3, 28.0, 27.7, 27.3, 27.0, 26.7, 26.4, 26.2, 26.1, …
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 102, 78, 79,…

Site built with pkgdown 2.0.9.

Template rostemplate by dieghernan, based on Bootstrapious.