Last observation values for a station

Get last observation values for a station.

Usage

aemet_last_obs(
  station = "all",
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE
)

Arguments

station

Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf

Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

extract_metadata

Logical TRUE/FALSE. On TRUE the output is a tibble with the description of the fields. See also get_metadata_aemet().

Value

A tibble or a sf object

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_last_obs(c("9434", "3195"))
glimpse(obs)
#> Rows: 48
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943…
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00…
#> $ fint      <dttm> 2023-08-15 13:00:00, 2023-08-15 14:00:00, 2023-08-15 15:00:…
#> $ prec      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, …
#> $ vmax      <dbl> 7.1, 8.1, 7.9, 8.7, 10.0, 11.3, 10.6, 9.5, 9.9, 7.0, 2.4, 4.…
#> $ vv        <dbl> 3.5, 5.2, 5.3, 5.2, 7.0, 6.3, 5.8, 6.0, 3.6, 2.3, 1.2, 3.5, …
#> $ dv        <dbl> 105, 115, 112, 122, 113, 114, 116, 120, 114, 106, 42, 307, 2…
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, …
#> $ dmax      <dbl> 115, 135, 120, 105, 118, 113, 115, 110, 120, 113, 43, 300, 2…
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A…
#> $ pres      <dbl> 984.6, 983.6, 983.0, 982.3, 982.2, 982.1, 982.6, 983.1, 983.…
#> $ hr        <dbl> 40, 38, 35, 32, 35, 36, 36, 35, 36, 43, 44, 61, 64, 65, 67, …
#> $ stdvv     <dbl> 0.9, 1.1, 1.0, 0.8, 1.2, 1.1, 0.9, 1.1, 0.8, 0.8, 0.4, 0.5, …
#> $ ts        <dbl> 37.8, 38.2, 38.8, 39.1, 37.0, 33.5, 31.6, 30.0, 28.4, 26.7, …
#> $ pres_nmar <dbl> 1012.4, 1011.3, 1010.6, 1009.8, 1009.8, 1009.9, 1010.5, 1011…
#> $ tamin     <dbl> 30.7, 33.2, 34.5, 35.5, 35.1, 33.2, 31.7, 30.2, 28.8, 27.3, …
#> $ ta        <dbl> 33.2, 34.5, 35.5, 36.1, 35.1, 33.2, 31.7, 30.2, 28.8, 27.3, …
#> $ tamax     <dbl> 33.2, 34.5, 35.5, 36.2, 36.1, 35.1, 33.2, 31.7, 30.2, 28.8, …
#> $ tpr       <dbl> 17.8, 18.1, 17.7, 16.8, 17.4, 16.1, 14.8, 13.1, 12.2, 13.7, …
#> $ stddv     <dbl> 23, 14, 10, 14, 10, 10, 9, 10, 12, 13, 22, 16, 10, 17, 9, 5,…
#> $ inso      <dbl> 60.0, 60.0, 60.0, 60.0, 60.0, 32.6, 0.0, 0.0, 0.0, 0.0, 0.0,…
#> $ tss5cm    <dbl> 40.9, 42.2, 42.7, 42.7, 41.8, 40.2, 38.4, 36.8, 35.6, 34.6, …
#> $ pacutp    <dbl> 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, …
#> $ tss20cm   <dbl> 33.0, 33.5, 34.0, 34.6, 35.1, 35.5, 35.8, 35.9, 35.9, 35.8, …
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …

Site built with pkgdown 2.0.7.

Template rostemplate by dieghernan, based on Bootstrapious.