Last observation values for a station

Get last observation values for a station.

aemet_last_obs(station = "all", verbose = FALSE, return_sf = FALSE)

Arguments

station
Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose
Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf
Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

A tibble or a sf object.

API Key

You need to set your API Key globally using aemet_api_key().

Examples

# Run this example only if AEMET_API_KEY is set

if (aemet_detect_api_key()) {
  library(tibble)
  obs <- aemet_last_obs(c("9434", "3195"))
  glimpse(obs)
}

#> Rows: 48
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943…
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00…
#> $ fint      <dttm> 2021-08-03 14:00:00, 2021-08-03 15:00:00, 2021-08-03 16:00:…
#> $ prec      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, …
#> $ vmax      <dbl> 6.8, 7.2, 7.5, 6.7, 6.1, 5.3, 2.8, 5.5, 9.3, 8.6, 9.2, 6.2, …
#> $ vv        <dbl> 4.6, 3.6, 4.2, 3.0, 3.9, 1.9, 2.0, 3.4, 6.2, 5.6, 4.9, 3.9, …
#> $ dv        <dbl> 108, 86, 110, 110, 122, 92, 77, 102, 316, 310, 313, 314, 302…
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, …
#> $ dmax      <dbl> 105, 83, 100, 98, 120, 100, 93, 95, 310, 305, 313, 320, 308,…
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A…
#> $ pres      <dbl> 983.6, 982.8, 982.2, 981.7, 981.3, 981.4, 981.7, 982.2, 982.…
#> $ hr        <dbl> 35, 30, 29, 33, 33, 35, 39, 45, 45, 46, 45, 46, 46, 52, 57, …
#> $ stdvv     <dbl> 0.9, 1.0, 1.1, 0.6, 0.5, 0.4, 0.3, 0.6, 1.1, 0.8, 0.8, 0.4, …
#> $ ts        <dbl> 35.4, 36.3, 34.9, 32.2, 30.9, 29.4, 27.4, 27.4, 26.3, 25.2, …
#> $ pres_nmar <dbl> 1011.5, 1010.6, 1010.0, 1009.5, 1009.2, 1009.4, 1009.8, 1010…
#> $ tamin     <dbl> 29.4, 31.3, 32.4, 31.6, 30.7, 29.8, 28.4, 27.8, 25.7, 24.6, …
#> $ ta        <dbl> 31.3, 32.5, 32.5, 31.6, 30.7, 29.8, 28.4, 27.8, 25.7, 24.6, …
#> $ tamax     <dbl> 31.3, 32.5, 32.9, 32.5, 31.6, 30.7, 29.8, 28.4, 27.8, 25.7, …
#> $ tpr       <dbl> 14.0, 12.6, 12.2, 13.4, 12.5, 12.6, 13.2, 14.8, 12.8, 12.2, …
#> $ stddv     <dbl> 13, 19, 15, 12, 9, 12, 13, 10, 8, 8, 8, 8, 7, 11, 8, 6, 6, 8…
#> $ inso      <dbl> 60.0, 59.2, 57.7, 19.4, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0…
#> $ tss5cm    <dbl> 35.5, 36.5, 36.8, 36.1, 35.0, 34.0, 33.1, 32.3, 31.7, 31.0, …
#> $ pacutp    <dbl> 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, …
#> $ tss20cm   <dbl> 29.6, 30.1, 30.7, 31.2, 31.6, 31.9, 32.0, 32.0, 31.9, 31.8, …
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.