Last observation values for a station

Get last observation values for a station.

Usage

aemet_last_obs(station = "all", verbose = FALSE, return_sf = FALSE)

Arguments

station

Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf

Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

A tibble or a sf object.

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_last_obs(c("9434", "3195"))
glimpse(obs)
#> Rows: 46
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943…
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00…
#> $ fint      <dttm> 2023-02-13 14:00:00, 2023-02-13 15:00:00, 2023-02-13 16:00:…
#> $ prec      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, …
#> $ vmax      <dbl> 5.9, 6.5, 6.6, 5.5, 5.0, 4.9, 1.1, 1.8, 1.5, 2.0, 3.4, 3.6, …
#> $ vv        <dbl> 3.5, 3.3, 3.2, 3.5, 2.0, 0.9, 0.6, 1.4, 1.0, 0.7, 2.4, 1.6, …
#> $ dv        <dbl> 122, 116, 141, 140, 99, 106, 269, 189, 224, 173, 144, 117, 1…
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, …
#> $ dmax      <dbl> 118, 108, 123, 130, 125, 345, 275, 195, 190, 185, 153, 148, …
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A…
#> $ pres      <dbl> 1003.3, 1003.0, 1002.8, 1002.8, 1003.2, 1003.4, 1003.5, 1003…
#> $ hr        <dbl> 54, 50, 47, 51, 58, 65, 70, 76, 78, 83, 84, 83, 86, 88, 91, …
#> $ stdvv     <dbl> 0.6, 0.9, 0.7, 0.6, 0.5, 0.3, 0.1, 0.3, 0.1, 0.2, 0.4, 0.1, …
#> $ ts        <dbl> 13.8, 14.1, 13.7, 11.0, 8.0, 5.2, 4.1, 2.4, 2.3, 1.5, 2.9, 1…
#> $ pres_nmar <dbl> 1034.0, 1033.6, 1033.4, 1033.5, 1034.0, 1034.4, 1034.8, 1035…
#> $ tamin     <dbl> 10.0, 11.6, 12.6, 11.8, 9.6, 8.3, 6.3, 5.0, 4.0, 3.7, 3.3, 3…
#> $ ta        <dbl> 11.6, 12.6, 12.8, 11.8, 11.3, 8.3, 6.3, 5.0, 4.0, 3.8, 3.8, …
#> $ tamax     <dbl> 11.7, 12.6, 12.9, 12.9, 11.8, 12.4, 8.3, 6.3, 5.0, 4.1, 3.8,…
#> $ tpr       <dbl> 2.7, 2.5, 1.7, 1.9, 3.4, 2.1, 1.3, 1.1, 0.4, 1.3, 1.3, 0.6, …
#> $ stddv     <dbl> 14, 18, 15, 8, 68, 34, 16, 6, 14, 14, 7, 5, 5, 24, 8, 12, 5,…
#> $ inso      <dbl> 60.0, 60.0, 60.0, 60.0, 23.2, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, …
#> $ tss5cm    <dbl> 9.7, 10.3, 10.4, 10.0, 9.2, 8.3, 7.5, 6.8, 6.3, 5.8, 5.3, 5.…
#> $ pacutp    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ tss20cm   <dbl> 5.8, 6.3, 6.8, 7.2, 7.6, 7.8, 7.8, 7.8, 7.7, 7.5, 7.4, 7.2, …
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …

Site built with pkgdown 2.0.7.

Template rostemplate by dieghernan, based on Bootstrapious.