Last observation values for a station

Get last observation values for a station.

Usage

aemet_last_obs(
  station = "all",
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE
)

Arguments

station

Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf

Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

extract_metadata

Logical TRUE/FALSE. On TRUE the output is a tibble with the description of the fields. See also get_metadata_aemet().

Value

A tibble or a sf object.

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_last_obs(c("9434", "3195"))
glimpse(obs)
#> Rows: 48
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943…
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00…
#> $ fint      <dttm> 2023-08-01 13:00:00, 2023-08-01 14:00:00, 2023-08-01 15:00:…
#> $ prec      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, …
#> $ vmax      <dbl> 12.9, 12.2, 11.5, 11.5, 9.3, 8.1, 5.8, 4.0, 3.5, 5.8, 6.0, 5…
#> $ vv        <dbl> 6.8, 7.1, 6.5, 5.1, 4.2, 3.3, 2.4, 2.1, 2.7, 3.2, 4.0, 3.8, …
#> $ dv        <dbl> 323, 320, 324, 331, 330, 323, 334, 333, 250, 282, 292, 280, …
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, …
#> $ dmax      <dbl> 303, 315, 310, 338, 315, 313, 318, 325, 328, 255, 263, 265, …
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A…
#> $ pres      <dbl> 985.1, 984.5, 983.9, 983.7, 983.3, 983.3, 983.2, 983.4, 983.…
#> $ hr        <dbl> 39, 37, 34, 34, 34, 33, 36, 40, 43, 44, 48, 50, 51, 60, 63, …
#> $ stdvv     <dbl> 1.5, 1.1, 1.6, 1.0, 1.1, 0.7, 0.5, 0.5, 0.3, 0.4, 0.5, 0.4, …
#> $ ts        <dbl> 38.4, 38.9, 38.8, 37.9, 36.8, 35.7, 30.6, 28.2, 26.8, 26.3, …
#> $ pres_nmar <dbl> 1013.2, 1012.5, 1011.8, 1011.6, 1011.3, 1011.3, 1011.3, 1011…
#> $ tamin     <dbl> 29.4, 30.8, 31.7, 32.0, 31.8, 31.6, 30.7, 28.9, 27.5, 26.7, …
#> $ ta        <dbl> 31.0, 31.7, 32.3, 32.2, 31.8, 31.7, 30.7, 28.9, 27.5, 26.8, …
#> $ tamax     <dbl> 31.2, 32.0, 32.4, 32.4, 32.5, 32.1, 31.7, 30.7, 28.9, 27.5, …
#> $ tpr       <dbl> 15.4, 15.3, 14.4, 14.3, 14.0, 13.5, 13.9, 13.9, 13.8, 13.6, …
#> $ stddv     <dbl> 15, 11, 15, 14, 20, 16, 13, 12, 9, 11, 6, 6, 5, 8, 5, 30, 16…
#> $ inso      <dbl> 60.0, 60.0, 60.0, 60.0, 60.0, 60.0, 60.0, 5.7, 0.0, 0.0, 0.0…
#> $ tss5cm    <dbl> 38.9, 40.2, 40.9, 40.7, 40.1, 39.0, 37.4, 35.8, 34.6, 33.5, …
#> $ pacutp    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ tss20cm   <dbl> 31.6, 32.1, 32.7, 33.3, 33.8, 34.2, 34.5, 34.7, 34.6, 34.5, …
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …

Site built with pkgdown 2.0.7.

Template rostemplate by dieghernan, based on Bootstrapious.