Last observation values for a station

Get last observation values for a station.

Usage

aemet_last_obs(station = "all", verbose = FALSE, return_sf = FALSE)

Arguments

station

Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf

Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

A tibble or a sf object.

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_last_obs(c("9434", "3195"))
glimpse(obs)
#> Rows: 48
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943…
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00…
#> $ fint      <dttm> 2022-11-15 13:00:00, 2022-11-15 14:00:00, 2022-11-15 15:00:…
#> $ prec      <dbl> 0.0, 0.0, 0.4, 0.2, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, …
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, …
#> $ vmax      <dbl> 4.4, 4.6, 4.0, 3.8, 9.7, 12.8, 9.2, 6.8, 11.1, 10.2, 7.7, 6.…
#> $ vv        <dbl> 3.4, 1.5, 1.7, 0.8, 7.0, 6.4, 5.0, 4.3, 8.3, 5.4, 4.1, 3.5, …
#> $ dv        <dbl> 119, 87, 71, 94, 252, 249, 264, 257, 246, 247, 271, 287, 272…
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, …
#> $ dmax      <dbl> 128, 108, 148, 90, 258, 250, 263, 265, 243, 233, 240, 283, 2…
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A…
#> $ pres      <dbl> 980.0, 979.7, 979.6, 979.3, 979.0, 979.2, 979.6, 979.6, 979.…
#> $ hr        <dbl> 80, 75, 84, 88, 63, 60, 66, 70, 70, 72, 73, 69, 74, 78, 77, …
#> $ stdvv     <dbl> 0.4, 0.9, 0.5, 0.2, 0.9, 0.9, 1.1, 0.8, 1.2, 0.8, 0.4, 0.3, …
#> $ ts        <dbl> 16.7, 17.5, 15.8, 15.2, 17.2, 17.6, 16.3, 15.8, 16.1, 15.3, …
#> $ pres_nmar <dbl> 1009.4, 1009.0, 1009.0, 1008.8, 1008.2, 1008.4, 1008.9, 1009…
#> $ tamin     <dbl> 16.8, 16.8, 16.5, 15.5, 15.6, 17.3, 17.0, 16.6, 16.1, 16.1, …
#> $ ta        <dbl> 16.8, 17.3, 16.5, 15.6, 17.9, 18.5, 17.2, 16.6, 16.7, 16.1, …
#> $ tamax     <dbl> 17.4, 17.3, 17.3, 16.5, 17.9, 18.6, 18.5, 17.2, 16.7, 16.8, …
#> $ tpr       <dbl> 13.4, 12.8, 13.8, 13.7, 10.8, 10.5, 10.8, 11.1, 11.2, 11.0, …
#> $ stddv     <dbl> 10, 54, 21, 49, 9, 10, 8, 9, 7, 9, 7, 8, 16, 12, 11, 12, 6, …
#> $ inso      <dbl> 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, …
#> $ tss5cm    <dbl> 15.4, 15.6, 15.5, 15.3, 15.2, 14.8, 14.6, 14.5, 14.3, 14.2, …
#> $ pacutp    <dbl> 0.05, 0.00, 0.58, 0.15, 0.00, 0.00, 0.00, 0.05, 0.00, 0.00, …
#> $ tss20cm   <dbl> 14.7, 14.8, 15.0, 15.1, 15.2, 15.2, 15.3, 15.3, 15.2, 15.2, …
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …

Site built with pkgdown 2.0.6.

Template rostemplate by dieghernan, based on Bootstrapious.