Last observation values for a station

Get last observation values for a station.

Usage

aemet_last_obs(station = "all", verbose = FALSE, return_sf = FALSE)

Arguments

station

Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf

Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

A tibble or a sf object.

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_last_obs(c("9434", "3195"))
glimpse(obs)
#> Rows: 48
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943…
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00…
#> $ fint      <dttm> 2022-02-20 15:00:00, 2022-02-20 16:00:00, 2022-02-20 17:00:…
#> $ prec      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, …
#> $ vmax      <dbl> 13.9, 12.4, 12.1, 9.6, 7.1, 8.6, 9.6, 8.7, 7.8, 5.4, 6.2, 8.…
#> $ vv        <dbl> 7.1, 6.3, 5.3, 4.6, 4.9, 6.8, 6.5, 6.2, 3.4, 2.6, 4.2, 4.3, …
#> $ dv        <dbl> 323, 319, 316, 326, 299, 290, 287, 290, 294, 316, 306, 295, …
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, …
#> $ dmax      <dbl> 308, 313, 308, 320, 330, 293, 288, 285, 308, 278, 305, 280, …
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A…
#> $ pres      <dbl> 997.0, 996.7, 996.9, 997.1, 997.8, 998.0, 998.2, 998.6, 998.…
#> $ hr        <dbl> 49, 48, 52, 58, 63, 66, 66, 68, 71, 72, 77, 75, 78, 83, 84, …
#> $ stdvv     <dbl> 1.2, 1.2, 1.1, 1.0, 0.5, 0.5, 0.6, 1.0, 0.6, 0.7, 0.9, 0.5, …
#> $ ts        <dbl> 18.3, 18.4, 14.6, 12.3, 11.2, 10.5, 10.2, 9.8, 8.8, 8.5, 7.4…
#> $ pres_nmar <dbl> 1027.0, 1026.6, 1026.9, 1027.3, 1028.2, 1028.5, 1028.7, 1029…
#> $ tamin     <dbl> 15.3, 15.7, 15.1, 13.5, 12.1, 11.2, 11.1, 10.4, 10.0, 9.6, 8…
#> $ ta        <dbl> 15.7, 15.9, 15.1, 13.5, 12.2, 11.3, 11.1, 10.8, 10.0, 9.6, 8…
#> $ tamax     <dbl> 15.7, 16.0, 15.9, 15.1, 13.5, 12.2, 11.6, 11.3, 11.0, 10.0, …
#> $ tpr       <dbl> 5.0, 5.0, 5.3, 5.5, 5.3, 5.1, 5.0, 5.1, 5.0, 4.8, 4.6, 4.8, …
#> $ stddv     <dbl> 12, 11, 9, 14, 7, 4, 5, 9, 10, 23, 7, 5, 24, 15, 15, 9, 10, …
#> $ inso      <dbl> 59.5, 58.7, 45.2, 20.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0…
#> $ tss5cm    <dbl> 13.9, 13.9, 13.7, 12.9, 12.2, 11.7, 11.3, 10.9, 10.6, 10.3, …
#> $ pacutp    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ tss20cm   <dbl> 10.2, 10.6, 10.9, 11.2, 11.3, 11.4, 11.4, 11.3, 11.3, 11.2, …
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …

Site built with pkgdown 2.0.2.

Template rostemplate by dieghernan, based on Bootstrapious.