Last observation values for a station

Get last observation values for a station.

aemet_last_obs(station = "all", verbose = FALSE, return_sf = FALSE)

Arguments

station
Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose
Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf
Logical. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

a tibble or a sf object.

Examples

# Run this example only if AEMET_API_KEY is set

if (aemet_detect_api_key()) {
  library(tibble)
  obs <- aemet_last_obs(c("9434", "3195"))
  glimpse(obs)
}

#> Rows: 48
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943…
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00…
#> $ fint      <dttm> 2021-04-05 01:00:00, 2021-04-05 02:00:00, 2021-04-05 03:00:…
#> $ prec      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, …
#> $ vmax      <dbl> 7.5, 7.7, 8.0, 7.1, 8.2, 6.9, 8.4, 10.2, 8.4, 8.5, 8.4, 9.0,…
#> $ vv        <dbl> 5.8, 5.0, 4.8, 4.9, 5.5, 4.7, 6.7, 5.6, 4.9, 6.0, 5.4, 4.5, …
#> $ dv        <dbl> 292, 278, 294, 284, 285, 289, 307, 317, 307, 323, 314, 320, …
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, …
#> $ dmax      <dbl> 295, 285, 280, 275, 303, 285, 310, 313, 298, 298, 305, 295, …
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A…
#> $ pres      <dbl> 989.2, 989.0, 988.8, 989.0, 989.4, 989.6, 990.0, 990.1, 989.…
#> $ hr        <dbl> 53, 53, 53, 52, 55, 58, 53, 50, 40, 37, 37, 36, 35, 30, 24, …
#> $ stdvv     <dbl> 0.6, 0.7, 0.4, 0.7, 0.4, 0.4, 0.7, 0.8, 1.0, 0.9, 0.9, 1.0, …
#> $ ts        <dbl> 9.2, 7.9, 7.9, 7.4, 7.0, 6.8, 9.6, 11.7, 14.9, 16.3, 19.0, 2…
#> $ pres_nmar <dbl> 1019.7, 1019.6, 1019.4, 1019.6, 1020.1, 1020.3, 1020.6, 1020…
#> $ tamin     <dbl> 9.4, 8.4, 8.2, 7.8, 7.4, 7.1, 7.1, 8.6, 9.6, 11.7, 13.5, 15.…
#> $ ta        <dbl> 9.4, 8.5, 8.2, 7.9, 7.4, 7.1, 8.6, 9.6, 11.7, 13.5, 15.1, 17…
#> $ tamax     <dbl> 10.3, 9.4, 8.7, 8.2, 7.9, 7.4, 8.6, 9.6, 11.7, 13.5, 15.1, 1…
#> $ tpr       <dbl> 0.2, -0.5, -0.7, -1.4, -1.0, -0.7, -0.5, -0.3, -1.4, -1.0, 0…
#> $ stddv     <dbl> 5, 7, 5, 6, 4, 5, 6, 6, 9, 10, 9, 16, 20, 13, 13, 9, 10, 9, …
#> $ inso      <dbl> 0.0, 0.0, 0.0, 0.0, 0.0, 2.2, 60.0, 60.0, 60.0, 60.0, 60.0, …
#> $ tss5cm    <dbl> 15.5, 15.0, 14.6, 14.2, 13.8, 13.5, 13.2, 13.2, 13.8, 14.6, …
#> $ pacutp    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ tss20cm   <dbl> 17.1, 16.9, 16.7, 16.4, 16.2, 16.0, 15.8, 15.5, 15.4, 15.2, …
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.