Last observation values for a station

Get last observation values for a station.

aemet_last_obs(
  station = "all",
  apikey = NULL,
  verbose = FALSE,
  return_sf = FALSE
)

Arguments

station
Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

apikey
Character string as personal API key. It can be set on the environment variable AEMET_API_KEY, see aemet_api_key().

verbose
Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf
Logical. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

a tibble or a sf object.

Examples

# Run this example only if AEMET_API_KEY is set

if (aemet_detect_api_key()) {
  library(tibble)
  obs <- aemet_last_obs(c("9434", "3195"))
  glimpse(obs)
}

#> Rows: 48
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943…
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00…
#> $ fint      <dttm> 2021-04-01 08:00:00, 2021-04-01 09:00:00, 2021-04-01 10:00:…
#> $ prec      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, …
#> $ vmax      <dbl> 3.7, 4.9, 7.2, 8.3, 9.1, 7.6, 6.7, 7.6, 8.1, 7.4, 9.5, 5.8, …
#> $ vv        <dbl> 2.6, 3.3, 4.3, 5.6, 5.8, 4.5, 2.4, 4.8, 4.4, 4.0, 3.7, 4.1, …
#> $ dv        <dbl> 116, 104, 110, 110, 107, 115, 111, 108, 109, 109, 111, 119, …
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, …
#> $ dmax      <dbl> 105, 95, 105, 98, 120, 123, 133, 125, 128, 108, 123, 120, 12…
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A…
#> $ pres      <dbl> 989.9, 989.6, 989.2, 988.2, 987.2, 987.0, 986.1, 984.6, 984.…
#> $ hr        <dbl> 53, 46, 47, 40, 34, 34, 31, 27, 27, 28, 27, 33, 35, 37, 38, …
#> $ stdvv     <dbl> 0.3, 0.7, 0.8, 1.0, 0.8, 0.9, 0.7, 0.9, 0.9, 0.8, 0.7, 0.5, …
#> $ ts        <dbl> 11.1, 15.2, 17.9, 19.9, 21.9, 22.9, 24.6, 22.2, 25.3, 23.2, …
#> $ pres_nmar <dbl> 1020.3, 1019.7, 1019.0, 1017.7, 1016.5, 1016.1, 1015.0, 1013…
#> $ tamin     <dbl> 8.9, 10.4, 13.1, 15.3, 17.5, 19.6, 21.0, 22.5, 22.3, 23.5, 2…
#> $ ta        <dbl> 10.4, 13.1, 15.3, 17.5, 19.6, 21.0, 22.5, 22.5, 24.0, 23.5, …
#> $ tamax     <dbl> 10.4, 13.1, 15.3, 17.5, 19.6, 21.0, 22.5, 23.5, 24.0, 24.2, …
#> $ tpr       <dbl> 1.3, 1.7, 4.1, 3.8, 3.2, 4.5, 4.5, 2.5, 3.8, 3.9, 1.7, 3.0, …
#> $ stddv     <dbl> 8, 20, 13, 15, 11, 10, 16, 10, 18, 10, 12, 10, 10, 11, 8, 6,…
#> $ inso      <dbl> 8.7, 60.0, 60.0, 60.0, 60.0, 57.6, 60.0, 31.2, 25.4, 46.4, 1…
#> $ tss5cm    <dbl> 13.6, 14.0, 14.9, 16.2, 17.6, 18.6, 19.4, 19.9, 19.4, 19.7, …
#> $ pacutp    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ tss20cm   <dbl> 15.2, 15.1, 15.0, 14.9, 15.0, 15.2, 15.5, 15.9, 16.2, 16.5, …
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.