Last observation values for a station

Get last observation values for a station.

aemet_last_obs(station = "all", verbose = FALSE, return_sf = FALSE)

Arguments

station
Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose
Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf
Logical. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

a tibble or a sf object.

Examples

# Run this example only if AEMET_API_KEY is set

if (aemet_detect_api_key()) {
  library(tibble)
  obs <- aemet_last_obs(c("9434", "3195"))
  glimpse(obs)
}

#> Rows: 46
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943…
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00…
#> $ fint      <dttm> 2021-05-29 09:00:00, 2021-05-29 10:00:00, 2021-05-29 11:00:…
#> $ prec      <dbl> 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.6, 0.2, …
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, …
#> $ vmax      <dbl> 4.7, 7.0, 7.4, 6.9, 5.6, 6.4, 4.6, 5.7, 4.9, 12.6, 8.0, 9.7,…
#> $ vv        <dbl> 2.3, 3.2, 5.1, 3.1, 2.8, 1.2, 2.1, 2.3, 3.1, 4.9, 6.4, 6.8, …
#> $ dv        <dbl> 336, 310, 305, 314, 305, 185, 353, 323, 337, 305, 317, 286, …
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, …
#> $ dmax      <dbl> 333, 313, 310, 295, 313, 338, 313, 340, 340, 293, 320, 293, …
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A…
#> $ pres      <dbl> 989.9, 989.7, 989.3, 988.5, 987.8, 987.2, 986.7, 986.1, 986.…
#> $ hr        <dbl> 53, 52, 47, 40, 34, 28, 22, 27, 32, 55, 71, 80, 72, 80, 84, …
#> $ stdvv     <dbl> 1.0, 0.8, 0.8, 1.0, 0.8, 0.6, 0.7, 0.9, 0.6, 0.7, 0.8, 1.1, …
#> $ ts        <dbl> 27.7, 30.1, 31.8, 33.8, 35.5, 36.4, 37.3, 37.3, 30.2, 24.6, …
#> $ pres_nmar <dbl> 1018.8, 1018.4, 1017.8, 1016.9, 1016.0, 1015.2, 1014.7, 1014…
#> $ tamin     <dbl> 21.7, 23.1, 24.9, 26.5, 28.2, 30.0, 31.6, 32.1, 29.8, 24.5, …
#> $ ta        <dbl> 23.2, 24.9, 26.5, 28.2, 30.0, 31.6, 32.2, 32.3, 29.8, 24.9, …
#> $ tamax     <dbl> 23.2, 25.0, 26.5, 28.2, 30.0, 31.6, 32.4, 32.9, 32.4, 29.8, …
#> $ tpr       <dbl> 13.1, 14.3, 14.2, 13.4, 12.4, 10.9, 7.8, 10.9, 11.3, 15.3, 1…
#> $ stddv     <dbl> 26, 19, 11, 13, 22, 97, 29, 38, 12, 12, 8, 9, 7, 7, 13, 10, …
#> $ inso      <dbl> 60.0, 60.0, 60.0, 60.0, 60.0, 60.0, 60.0, 60.0, 12.2, 0.0, 0…
#> $ tss5cm    <dbl> 25.5, 26.9, 28.5, 30.1, 31.8, 33.1, 34.0, 34.3, 33.5, 31.8, …
#> $ pacutp    <dbl> 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.32, …
#> $ tss20cm   <dbl> 24.5, 24.5, 24.6, 24.9, 25.3, 25.8, 26.4, 27.0, 27.6, 28.0, …
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.