Last observation values for a station

Get last observation values for a station.

aemet_last_obs(station = "all", verbose = FALSE, return_sf = FALSE)

Arguments

station
Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose
Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf
Logical. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

a tibble or a sf object.

Examples

# Run this example only if AEMET_API_KEY is set

if (aemet_detect_api_key()) {
  library(tibble)
  obs <- aemet_last_obs(c("9434", "3195"))
  glimpse(obs)
}

#> Rows: 46
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943…
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00…
#> $ fint      <dttm> 2021-04-14 11:00:00, 2021-04-14 12:00:00, 2021-04-14 13:00:…
#> $ prec      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, …
#> $ vmax      <dbl> 5.4, 3.6, 3.9, 4.0, 4.2, 4.1, 2.7, 3.7, 1.9, 2.5, 5.0, 5.3, …
#> $ vv        <dbl> 1.5, 1.4, 1.6, 1.3, 2.3, 2.2, 0.9, 1.2, 0.9, 1.7, 3.5, 2.5, …
#> $ dv        <dbl> 122, 140, 100, 60, 128, 107, 138, 93, 69, 143, 113, 144, 33,…
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, …
#> $ dmax      <dbl> 180, 100, 135, 115, 55, 113, 120, 115, 78, 153, 123, 135, 40…
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A…
#> $ pres      <dbl> 989.6, 988.8, 988.3, 987.7, 986.9, 986.2, 986.2, 986.1, 986.…
#> $ hr        <dbl> 65, 57, 52, 44, 39, 37, 40, 43, 50, 61, 73, 75, 80, 81, 66, …
#> $ stdvv     <dbl> 0.8, 0.6, 0.7, 0.5, 0.6, 0.5, 0.3, 0.3, 0.2, 0.2, 0.5, 0.9, …
#> $ ts        <dbl> 21.0, 22.0, 23.7, 24.3, 23.8, 21.4, 20.5, 18.4, 12.5, 13.2, …
#> $ pres_nmar <dbl> 1019.1, 1018.2, 1017.6, 1016.9, 1016.0, 1015.4, 1015.4, 1015…
#> $ tamin     <dbl> 14.9, 17.3, 18.2, 19.4, 20.1, 20.4, 20.0, 19.3, 17.3, 15.9, …
#> $ ta        <dbl> 17.3, 18.2, 19.5, 20.1, 21.1, 20.4, 20.1, 19.3, 17.3, 15.9, …
#> $ tamax     <dbl> 17.3, 18.2, 19.6, 20.2, 21.2, 21.4, 20.5, 20.1, 19.3, 17.3, …
#> $ tpr       <dbl> 10.6, 9.6, 9.4, 7.4, 6.5, 5.1, 6.1, 6.4, 6.8, 8.4, 9.3, 8.8,…
#> $ stddv     <dbl> 42, 86, 34, 94, 21, 20, 31, 13, 14, 8, 8, 15, 35, 50, 10, 12…
#> $ inso      <dbl> 46.6, 39.7, 55.0, 60.0, 60.0, 60.0, 24.8, 23.6, 0.0, 0.0, 0.…
#> $ tss5cm    <dbl> 16.2, 17.6, 19.1, 20.1, 21.1, 21.4, 20.9, 20.2, 19.2, 18.2, …
#> $ pacutp    <dbl> 0.00, 0.00, 0.01, 0.00, 0.00, 0.00, 0.00, 0.01, 0.00, 0.00, …
#> $ tss20cm   <dbl> 14.2, 14.4, 14.7, 15.1, 15.6, 16.1, 16.6, 17.0, 17.2, 17.3, …
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.