Last observation values for a station

Get last observation values for a station.

Usage

aemet_last_obs(
  station = "all",
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)

Arguments

station

Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf

Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

extract_metadata

Logical TRUE/FALSE. On TRUE the output is a tibble with the description of the fields. See also get_metadata_aemet().

progress

Logical, display a cli::cli_progress_bar() object. If verbose = TRUE won't be displayed.

Value

A tibble or a sf object

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_last_obs(c("9434", "3195"))
glimpse(obs)
#> Rows: 26
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943…
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00…
#> $ fint      <dttm> 2024-07-03 00:00:00, 2024-07-03 01:00:00, 2024-07-03 02:00:…
#> $ prec      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, …
#> $ vmax      <dbl> 13.7, 13.2, 13.2, 13.1, 12.4, 12.6, 14.1, 14.1, 14.3, 13.7, …
#> $ vv        <dbl> 10.7, 9.6, 8.9, 9.6, 8.9, 9.4, 9.1, 10.1, 10.0, 9.1, 7.5, 7.…
#> $ dv        <dbl> 293, 287, 295, 301, 295, 294, 300, 297, 298, 305, 300, 311, …
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, …
#> $ dmax      <dbl> 290, 298, 303, 293, 288, 293, 300, 298, 300, 305, 288, 320, …
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A…
#> $ pres      <dbl> 988.9, 988.3, 987.8, 987.6, 987.5, 987.6, 987.7, 988.0, 987.…
#> $ hr        <dbl> 60, 63, 67, 68, 71, 71, 69, 62, 56, 53, 45, 33, 19, 31, 35, …
#> $ stdvv     <dbl> 1.0, 0.8, 1.1, 1.2, 1.0, 1.0, 1.3, 1.3, 1.3, 1.3, 1.8, 1.6, …
#> $ ts        <dbl> 17.7, 17.3, 16.6, 16.1, 15.6, 15.4, 16.5, 19.6, 22.3, 25.0, …
#> $ pres_nmar <dbl> 1018.5, 1017.9, 1017.5, 1017.3, 1017.3, 1017.4, 1017.5, 1017…
#> $ tamin     <dbl> 17.5, 17.0, 16.5, 16.2, 15.5, 15.3, 15.2, 15.7, 16.9, 18.3, …
#> $ ta        <dbl> 17.5, 17.1, 16.5, 16.2, 15.5, 15.3, 15.7, 17.0, 18.4, 19.4, …
#> $ tamax     <dbl> 18.0, 17.5, 17.1, 16.5, 16.2, 15.5, 15.7, 17.2, 18.4, 19.4, …
#> $ tpr       <dbl> 9.7, 10.0, 10.3, 10.3, 10.3, 10.0, 10.0, 9.7, 9.4, 9.6, 9.6,…
#> $ stddv     <dbl> 5, 6, 6, 7, 6, 5, 7, 6, 8, 7, 12, 12, 9, NA, NA, NA, NA, NA,…
#> $ inso      <dbl> 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 27.5, 60.0, 60.0, 60.0, 60.0, …
#> $ tss5cm    <dbl> 25.6, 25.0, 24.4, 23.9, 23.4, 22.9, 22.6, 22.7, 23.3, 24.4, …
#> $ pacutp    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, NA, NA, NA, NA, NA, N…
#> $ tss20cm   <dbl> 28.6, 28.3, 27.9, 27.6, 27.2, 26.9, 26.5, 26.2, 25.9, 25.7, …
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 122, 119…

Site built with pkgdown 2.0.9.

Template rostemplate by dieghernan, based on Bootstrapious.