Last observation values for a station

Get last observation values for a station.

Usage

aemet_last_obs(
  station = "all",
  verbose = FALSE,
  return_sf = FALSE,
  extract_metadata = FALSE,
  progress = TRUE
)

Arguments

station

Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf

Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

extract_metadata

Logical TRUE/FALSE. On TRUE the output is a tibble with the description of the fields. See also get_metadata_aemet().

progress

Logical, display a cli::cli_progress_bar() object. If verbose = TRUE won't be displayed.

Value

A tibble or a sf object

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_last_obs(c("9434", "3195"))
#> HTTP 500: Hit API Limits. Retrying...
#> Waiting 4s for retry backoff ■■■■■■■■■                       
#> Waiting 4s for retry backoff ■■■■■■■■■■■■■■■■■■■■■■          
#> Waiting 4s for retry backoff ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  
#> Waiting 13s for retry backoff ■■■                             
#> Waiting 13s for retry backoff ■■■■■■                          
#> Waiting 13s for retry backoff ■■■■■■■■■■■■■                   
#> Waiting 13s for retry backoff ■■■■■■■■■■■■■■■■■■■■            
#> Waiting 13s for retry backoff ■■■■■■■■■■■■■■■■■■■■■■■■■■■     
#> Waiting 13s for retry backoff ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 
glimpse(obs)
#> Rows: 24
#> Columns: 26
#> $ idema     <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "943…
#> $ lon       <dbl> -1.004167, -1.004167, -1.004167, -1.004167, -1.004167, -1.00…
#> $ fint      <dttm> 2024-12-07 09:00:00, 2024-12-07 10:00:00, 2024-12-07 11:00:…
#> $ prec      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ alt       <dbl> 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, 249, …
#> $ vmax      <dbl> 4.1, 8.8, 10.8, 19.8, 16.0, 16.3, 17.1, 15.9, 15.7, 12.3, 14…
#> $ vv        <dbl> 2.8, 5.4, 7.3, 10.7, 9.7, 10.8, 10.2, 11.1, 7.6, 6.6, 8.3, 5…
#> $ dv        <dbl> 285, 315, 309, 323, 317, 323, 320, 321, 315, 309, 317, 321, …
#> $ lat       <dbl> 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, 41.66056, …
#> $ dmax      <dbl> 285, 318, 303, 325, 313, 328, 313, 315, 323, 333, 323, 315, …
#> $ ubi       <chr> "ZARAGOZA  AEROPUERTO", "ZARAGOZA  AEROPUERTO", "ZARAGOZA  A…
#> $ pres      <dbl> 991.5, 992.4, 992.1, 992.0, 991.4, 991.3, 991.4, 992.1, 992.…
#> $ hr        <dbl> 74, 74, 74, 71, 56, 53, 53, 54, 60, 61, 66, 74, 88, 78, 68, …
#> $ stdvv     <dbl> 0.3, 1.0, 1.2, 1.9, 2.0, 1.9, 1.9, 2.0, 1.6, 1.5, 1.6, 1.4, …
#> $ ts        <dbl> 13.4, 14.6, 15.5, 15.3, 16.8, 15.9, 15.0, 12.2, 10.9, 10.5, …
#> $ pres_nmar <dbl> 1021.6, 1022.5, 1022.1, 1022.1, 1021.3, 1021.3, 1021.5, 1022…
#> $ tamin     <dbl> 12.9, 13.5, 13.9, 13.7, 14.0, 14.9, 13.9, 12.4, 11.2, 10.6, …
#> $ ta        <dbl> 13.5, 14.1, 14.7, 14.0, 15.3, 14.9, 14.1, 12.4, 11.2, 10.7, …
#> $ tamax     <dbl> 13.5, 14.9, 14.7, 14.7, 15.3, 15.4, 15.0, 14.2, 12.4, 11.2, …
#> $ tpr       <dbl> 8.9, 9.6, 10.2, 8.8, 6.5, 5.5, 4.6, 3.4, 3.8, 3.4, 3.4, 3.9,…
#> $ stddv     <dbl> 6, 10, 9, 21, 11, 16, 16, 16, 12, 11, 12, 11, NA, NA, NA, NA…
#> $ inso      <dbl> 16.9, 35.2, 29.5, 25.4, 59.8, 55.3, 52.5, 42.6, 0.0, 0.0, 0.…
#> $ tss5cm    <dbl> 11.7, 12.3, 12.7, 12.9, 13.4, 13.8, 13.4, 13.1, 12.3, 11.7, …
#> $ pacutp    <dbl> 0.00, 0.00, 0.00, 0.02, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, …
#> $ tss20cm   <dbl> 12.1, 12.1, 12.1, 12.2, 12.3, 12.4, 12.6, 12.7, 12.8, 12.8, …
#> $ rviento   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 77, 59, 146,…

Site built with pkgdown 2.1.1.

Template rostemplate by dieghernan, based on Bootstrapious.