Normal climatology values

Get normal climatology values for a station (or all the stations with aemet_normal_clim_all(). Standard climatology from 1981 to 2010.

Usage

aemet_normal_clim(station = NULL, verbose = FALSE, return_sf = FALSE)

aemet_normal_clim_all(verbose = FALSE, return_sf = FALSE)

Arguments

station

Character string with station identifier code(s) (see aemet_stations()) or "all" for all the stations.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

return_sf

Logical TRUE or FALSE. Should the function return an sf spatial object? If FALSE (the default value) it returns a tibble. Note that you need to have the sf package installed.

Value

A tibble or a sf object.

Note

Code modified from project https://github.com/SevillaR/aemet

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

library(tibble)
obs <- aemet_normal_clim(c("9434", "3195"))
glimpse(obs)
#> Rows: 26
#> Columns: 475
#> $ indicativo  <chr> "9434", "9434", "9434", "9434", "9434", "9434", "9434", "9…
#> $ w_racha_max <dbl> 29.7, 35.0, 30.3, 26.7, 28.3, 30.8, 28.1, 24.7, 25.8, 28.3…
#> $ np_010_n    <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ np_010_s    <dbl> 3.35, 2.40, 2.29, 3.09, 3.04, 2.63, 1.72, 1.86, 2.01, 3.43…
#> $ q_max_s     <dbl> 4.50, 4.57, 3.61, 2.95, 2.75, 2.04, 1.94, 1.60, 1.85, 2.33…
#> $ n_tor_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ n_tor_s     <dbl> 0.00, 0.25, 0.60, 1.40, 2.67, 2.38, 2.10, 2.21, 1.86, 0.89…
#> $ q_max_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ tm_min_q4   <dbl> 4.3, 4.5, 7.2, 8.7, 12.9, 16.6, 19.2, 19.3, 16.0, 12.1, 7.…
#> $ tm_min_q1   <dbl> 1.1, 2.4, 4.7, 7.2, 10.6, 15.0, 17.2, 17.2, 14.2, 10.0, 5.…
#> $ tm_min_q3   <dbl> 3.3, 4.0, 5.7, 8.3, 12.2, 15.8, 18.6, 18.7, 15.3, 11.1, 6.…
#> $ tm_min_q2   <dbl> 2.2, 2.9, 5.4, 7.8, 11.7, 15.5, 17.7, 18.1, 14.9, 10.8, 5.…
#> $ q_mar_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ n_des_min   <dbl> 0, 1, 1, 1, 0, 2, 7, 6, 3, 1, 0, 0, 56, 0, 2, 0, 0, 0, 1, …
#> $ q_mar_s     <dbl> 5.81, 4.85, 3.76, 2.14, 2.09, 1.54, 1.29, 1.33, 2.05, 2.86…
#> $ q_med_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ ts_20_q1    <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ ts_20_q2    <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ e_cv        <dbl> 0.11, 0.13, 0.11, 0.11, 0.10, 0.09, 0.09, 0.12, 0.12, 0.10…
#> $ ts_20_q4    <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ e_min       <dbl> 61, 55, 67, 75, 97, 115, 132, 132, 119, 107, 69, 60, 103, …
#> $ np_300_q4   <dbl> 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0…
#> $ np_300_q1   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ np_300_q3   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0…
#> $ hr_max      <dbl> 83, 78, 70, 69, 67, 62, 57, 61, 69, 80, 84, 84, 66, 81, 73…
#> $ np_300_cv   <dbl> NA, NA, 3.81, 2.59, 3.05, 2.59, 3.81, 3.05, 2.27, 2.59, 3.…
#> $ n_nub_max   <dbl> 24, 23, 26, 24, 26, 23, 23, 22, 23, 25, 26, 24, 245, 22, 2…
#> $ tm_min_cv   <dbl> 0.64, 0.41, 0.21, 0.12, 0.11, 0.08, 0.06, 0.05, 0.07, 0.10…
#> $ n_des_mn    <dbl> 4.5, 5.0, 6.5, 4.0, 3.0, 7.0, 14.5, 11.0, 8.0, 6.0, 3.5, 3…
#> $ n_des_md    <dbl> 4.6, 5.1, 6.7, 4.6, 4.5, 8.2, 14.6, 10.9, 8.0, 5.4, 4.0, 4…
#> $ ts_20_s     <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ q_med_mn    <dbl> 990.1, 989.0, 987.9, 984.2, 984.5, 986.1, 986.0, 985.8, 98…
#> $ evap_cv     <dbl> 0.41, 0.39, 0.24, 0.27, 0.23, 0.25, 0.21, 0.23, 0.24, 0.29…
#> $ q_med_md    <dbl> 991.0, 989.1, 987.5, 983.9, 984.5, 985.8, 985.9, 985.6, 98…
#> $ nt_30_cv    <dbl> NA, NA, NA, 2.65, 0.79, 0.37, 0.20, 0.24, 0.66, 2.53, NA, …
#> $ mes         <chr> "01", "02", "03", "04", "05", "06", "07", "08", "09", "10"…
#> $ ts_20_cv    <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ inso_max    <dbl> 6.9, 7.7, 9.6, 9.2, 10.8, 12.0, 13.0, 11.4, 9.6, 7.9, 7.6,…
#> $ ts_20_max   <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ np_001_max  <dbl> 20, 14, 13, 18, 20, 15, 9, 9, 12, 16, 15, 20, 103, 20, 18,…
#> $ n_llu_md    <dbl> 8.7, 7.8, 7.4, 10.0, 10.8, 7.5, 5.0, 5.1, 7.1, 10.1, 9.9, …
#> $ nv_1000_s   <dbl> 3.06, 2.24, 0.31, 0.18, 0.35, 0.00, 0.00, 0.00, 0.31, 0.78…
#> $ ta_min_mn   <dbl> -3.4, -2.3, -0.6, 3.0, 6.5, 10.6, 13.8, 13.6, 9.8, 5.5, -0…
#> $ tm_mes_max  <dbl> 9.4, 12.1, 14.6, 15.4, 20.5, 26.6, 28.2, 27.9, 24.1, 18.8,…
#> $ ta_max_mn   <dbl> 17.1, 19.2, 24.0, 27.5, 32.1, 36.1, 38.8, 37.6, 33.9, 27.5…
#> $ nv_1000_n   <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 28, 29…
#> $ ta_max_md   <dbl> 17.1, 19.1, 24.5, 27.2, 31.8, 36.4, 38.9, 37.5, 33.6, 27.6…
#> $ nw_91_min   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ ta_max_n    <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ ta_max_s    <dbl> 1.87, 1.79, 1.96, 2.49, 2.56, 2.13, 1.90, 2.01, 2.42, 2.20…
#> $ ts_10_md    <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ nw_55_mn    <dbl> 7, 8, 8, 8, 6, 6, 5, 5, 4, 4, 5, 6, 77, 2, 1, 2, 2, 1, 1, …
#> $ ta_min_s    <dbl> 2.35, 2.09, 2.25, 1.44, 1.77, 1.70, 1.58, 1.57, 1.72, 1.97…
#> $ nw_55_md    <dbl> 7.9, 8.1, 9.4, 8.8, 6.6, 6.1, 6.6, 5.3, 4.6, 4.7, 6.0, 6.7…
#> $ np_300_mn   <dbl> 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0…
#> $ evap_q4     <dbl> 1048, 1519, 2071, 2286, 2787, 3542, 4016, 3812, 2651, 1758…
#> $ evap_q1     <dbl> 545, 707, 1291, 1381, 1788, 2199, 2478, 2339, 1652, 1025, …
#> $ np_300_md   <dbl> 0.0, 0.0, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.2, 0.1, 0.1, 0.0…
#> $ evap_q3     <dbl> 850, 1188, 1879, 1922, 2406, 2937, 3529, 3215, 2324, 1507,…
#> $ evap_q2     <dbl> 727, 807, 1607, 1638, 2168, 2852, 3173, 2938, 1933, 1289, …
#> $ p_mes_min   <dbl> 0.0, 0.0, 0.0, 3.8, 5.2, 0.0, 0.3, 0.1, 0.0, 0.6, 0.4, 0.7…
#> $ ts_min_min  <dbl> 5.2, 5.5, 8.5, 10.5, 13.4, 17.8, 20.0, 19.2, 17.1, 13.5, 9…
#> $ n_des_s     <dbl> 3.54, 2.78, 4.86, 3.44, 3.05, 3.76, 3.62, 3.13, 3.37, 3.36…
#> $ nv_0100_mn  <dbl> 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0…
#> $ p_mes_md    <dbl> 21.0, 21.5, 19.1, 39.3, 43.7, 26.4, 17.3, 16.6, 29.5, 36.4…
#> $ n_tor_cv    <dbl> NA, 3.81, 1.99, 1.03, 0.65, 0.60, 0.56, 0.59, 0.66, 0.86, …
#> $ nv_0100_md  <dbl> 0.2, 0.2, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.1, 0.4…
#> $ q_min_max   <dbl> 992.1, 988.6, 988.8, 976.6, 978.3, 980.1, 979.7, 981.9, 98…
#> $ n_des_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 26, 27…
#> $ p_sol_s     <dbl> 10.16, 10.83, 8.41, 8.93, 7.77, 7.62, 6.96, 5.58, 7.49, 8.…
#> $ ts_20_n     <dbl> 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 0, 0, …
#> $ ts_10_cv    <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ nw_91_s     <dbl> 1.35, 0.38, 0.50, 0.31, 0.42, 0.46, 0.56, 0.00, 0.19, 0.38…
#> $ nw_91_n     <dbl> 28, 29, 30, 29, 27, 26, 29, 29, 27, 27, 28, 28, 19, 27, 27…
#> $ nt_00_min   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 0, 0, 0, 0, 0, 0, 0…
#> $ p_sol_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 29, 29, 30, 30, 30, 29, 23, 24…
#> $ n_nie_mn    <dbl> 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0…
#> $ n_nub_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 26, 27…
#> $ n_nie_md    <dbl> 0.7, 0.4, 0.2, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.1, 0.5…
#> $ tm_mes_md   <dbl> 6.6, 8.2, 11.6, 13.8, 18.0, 22.6, 25.3, 25.0, 21.2, 16.2, …
#> $ p_sol_cv    <dbl> 0.23, 0.20, 0.14, 0.16, 0.13, 0.11, 0.09, 0.08, 0.12, 0.16…
#> $ n_nub_s     <dbl> 4.19, 3.27, 3.70, 4.00, 2.91, 3.21, 3.47, 3.08, 3.14, 2.50…
#> $ w_racha_min <dbl> 15.8, 16.9, 15.0, 16.7, 13.6, 12.5, 14.4, 14.4, 14.7, 12.2…
#> $ np_300_max  <dbl> 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 4, 1, 1, 1, 0, 2, 0, 1…
#> $ nv_0050_md  <dbl> 0.1, 0.2, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.2…
#> $ ta_min_min  <dbl> -6.6, -7.6, -6.0, -0.8, 1.7, 7.4, 10.8, 10.4, 7.0, 1.0, -5…
#> $ glo_mn      <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ inso_s      <dbl> 0.96, 1.14, 0.99, 1.18, 1.14, 1.17, 1.05, 0.79, 0.93, 0.99…
#> $ glo_md      <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ n_nie_q1    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ w_med_q4    <dbl> 21.6, 20.0, 21.0, 20.0, 20.0, 20.0, 20.6, 19.8, 17.0, 17.0…
#> $ n_tor_max   <dbl> 0, 1, 2, 5, 10, 9, 9, 8, 8, 3, 1, 2, 32, 1, 1, 3, 5, 11, 9…
#> $ w_med_q2    <dbl> 14.4, 16.0, 17.0, 18.0, 16.0, 16.4, 16.4, 16.0, 13.0, 13.0…
#> $ w_med_q3    <dbl> 17.6, 18.0, 20.0, 19.0, 18.0, 18.0, 19.0, 17.6, 15.0, 15.0…
#> $ n_gra_md    <dbl> 0.0, 0.0, 0.1, 0.0, 0.1, 0.0, 0.1, 0.1, 0.1, 0.0, 0.0, 0.0…
#> $ inso_q4     <dbl> 5.0, 6.9, 7.9, 8.7, 10.0, 11.3, 12.3, 10.7, 9.1, 7.3, 5.8,…
#> $ q_max_min   <dbl> 994.7, 992.3, 992.2, 990.2, 989.5, 989.5, 990.5, 989.8, 99…
#> $ ts_10_max   <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ nt_30_min   <dbl> 0, 0, 0, 0, 0, 3, 14, 11, 0, 0, 0, 0, 49, 0, 0, 0, 0, 0, 1…
#> $ n_nub_cv    <dbl> 0.24, 0.18, 0.20, 0.21, 0.14, 0.17, 0.23, 0.17, 0.17, 0.12…
#> $ p_mes_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ p_mes_q1    <dbl> 4.3, 5.3, 4.6, 14.5, 19.4, 7.2, 3.9, 2.3, 3.7, 10.0, 13.5,…
#> $ p_mes_q2    <dbl> 10.1, 14.0, 9.8, 28.1, 31.8, 13.4, 7.6, 6.6, 15.4, 23.4, 1…
#> $ p_mes_q3    <dbl> 16.2, 22.6, 19.2, 41.8, 40.0, 24.3, 17.5, 10.5, 26.6, 40.3…
#> $ p_mes_q4    <dbl> 35.2, 33.3, 31.1, 49.1, 68.8, 54.7, 33.9, 30.2, 55.0, 60.7…
#> $ hr_cv       <dbl> 0.08, 0.08, 0.08, 0.09, 0.09, 0.10, 0.08, 0.10, 0.08, 0.08…
#> $ n_tor_md    <dbl> 0.0, 0.1, 0.3, 1.4, 4.1, 3.9, 3.8, 3.7, 2.8, 1.0, 0.1, 0.1…
#> $ p_mes_s     <dbl> 22.15, 18.02, 16.84, 30.17, 31.03, 23.59, 16.13, 20.64, 28…
#> $ nw_91_cv    <dbl> 2.90, 2.23, 2.16, 3.00, 1.91, 3.02, 2.70, NA, 5.20, 5.20, …
#> $ n_tor_mn    <dbl> 0.0, 0.0, 0.0, 1.0, 4.0, 4.0, 3.0, 4.0, 3.0, 1.0, 0.0, 0.0…
#> $ nv_1000_max <dbl> 12, 9, 1, 1, 1, 0, 0, 0, 1, 2, 9, 12, 23, 7, 6, 4, 2, 1, 0…
#> $ n_gra_cv    <dbl> 5.00, NA, 3.26, NA, 3.96, 5.39, 3.60, 3.81, 2.94, NA, 5.48…
#> $ w_racha_n   <dbl> 28, 29, 30, 29, 27, 26, 29, 29, 27, 27, 29, 28, 20, 27, 27…
#> $ w_med_min   <dbl> 8, 8, 11, 9, 12, 13, 13, 9, 11, 9, 9, 9, 14, 4, 5, 6, 5, 6…
#> $ w_racha_s   <dbl> 3.77, 3.51, 3.63, 2.51, 3.99, 4.05, 3.83, 2.68, 2.82, 3.27…
#> $ np_100_max  <dbl> 3, 3, 2, 4, 5, 3, 2, 3, 4, 4, 3, 1, 17, 5, 5, 3, 5, 5, 3, …
#> $ e_s         <dbl> 8.42, 9.71, 9.57, 10.25, 11.51, 11.98, 13.93, 19.25, 17.16…
#> $ w_med_s     <dbl> 5.25, 4.45, 3.29, 3.63, 3.74, 2.96, 2.88, 2.99, 2.37, 2.76…
#> $ p_sol_md    <dbl> 44, 55, 59, 57, 61, 68, 76, 74, 65, 57, 50, 42, 59, 50, 52…
#> $ n_cub_q4    <dbl> 11.8, 7.0, 8.0, 8.0, 8.8, 5.0, 2.0, 2.8, 4.8, 7.8, 9.0, 13…
#> $ n_cub_q3    <dbl> 10.6, 6.0, 6.0, 7.6, 6.0, 3.0, 1.0, 2.0, 4.0, 6.0, 8.6, 9.…
#> $ n_cub_q2    <dbl> 8.0, 4.0, 4.4, 5.0, 5.0, 2.0, 0.0, 1.0, 2.0, 4.0, 6.0, 8.0…
#> $ n_cub_q1    <dbl> 6.0, 3.0, 3.0, 4.0, 4.0, 1.0, 0.0, 0.0, 1.2, 2.2, 3.2, 7.0…
#> $ nw_55_max   <dbl> 22, 17, 23, 21, 17, 18, 18, 13, 14, 11, 16, 20, 173, 7, 7,…
#> $ p_sol_mn    <dbl> 43, 56, 57, 58, 62, 70, 77, 74, 63, 59, 53, 42, 60, 47, 51…
#> $ n_cub_s     <dbl> 3.47, 3.24, 2.83, 3.07, 2.72, 2.12, 1.58, 2.06, 1.94, 3.46…
#> $ hr_mn       <dbl> 75, 66, 59, 56, 53, 48, 47, 51, 57, 68, 73, 77, 61, 72, 65…
#> $ q_med_max   <dbl> 1002.5, 996.5, 997.4, 986.7, 988.0, 987.9, 987.9, 988.5, 9…
#> $ hr_md       <dbl> 75, 67, 59, 57, 54, 49, 47, 51, 57, 67, 73, 76, 61, 71, 64…
#> $ np_010_md   <dbl> 4.0, 3.9, 3.7, 5.7, 6.4, 4.0, 2.6, 2.3, 3.2, 5.4, 5.1, 4.8…
#> $ n_cub_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 26, 27…
#> $ np_010_mn   <dbl> 3.0, 4.5, 4.0, 5.5, 6.0, 3.5, 2.0, 2.0, 3.5, 4.0, 5.0, 4.0…
#> $ nw_91_q4    <dbl> 0.2, 0.0, 0.8, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0…
#> $ p_max_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ nw_91_q2    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0…
#> $ nw_91_q3    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 0, 0, 0, 0, 0, 0, 0…
#> $ nw_91_q1    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ n_gra_q4    <dbl> 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0…
#> $ n_gra_q1    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ n_gra_q3    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0…
#> $ nw_55_s     <dbl> 5.68, 3.79, 5.18, 4.25, 4.39, 3.82, 4.79, 3.50, 3.08, 2.80…
#> $ nw_55_n     <dbl> 28, 29, 30, 29, 27, 26, 29, 29, 27, 27, 28, 28, 19, 27, 27…
#> $ w_med_mn    <dbl> 15.5, 17.0, 18.0, 18.0, 17.0, 17.5, 17.5, 16.5, 14.0, 14.0…
#> $ n_nub_min   <dbl> 10, 9, 11, 10, 15, 10, 8, 12, 9, 16, 9, 10, 186, 9, 10, 7,…
#> $ p_max_s     <dbl> 6.27, 8.05, 9.23, 12.61, 11.57, 13.67, 11.87, 12.46, 16.90…
#> $ ts_20_min   <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ inso_min    <dbl> 2.2, 3.3, 5.0, 5.3, 6.0, 7.6, 9.0, 7.7, 6.1, 4.5, 2.8, 2.2…
#> $ n_gra_max   <dbl> 1, 0, 2, 0, 2, 1, 1, 2, 1, 0, 1, 0, 5, 1, 1, 1, 3, 3, 2, 1…
#> $ p_sol_q1    <dbl> 36, 48, 51, 46, 54, 58, 70, 69, 57, 47, 39, 34, 55, 38, 42…
#> $ p_sol_q3    <dbl> 45, 60, 62, 62, 64, 72, 78, 76, 65, 60, 55, 45, 61, 53, 57…
#> $ p_sol_q2    <dbl> 41, 52, 55, 54, 59, 67, 75, 72, 62, 54, 50, 39, 60, 44, 48…
#> $ p_sol_q4    <dbl> 52, 65, 66, 65, 69, 74, 83, 77, 73, 66, 59, 52, 62, 65, 63…
#> $ n_fog_q1    <dbl> 2.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 2.0, 4.0…
#> $ ta_min_md   <dbl> -3.1, -2.4, -0.4, 2.5, 6.2, 10.8, 13.8, 13.8, 9.8, 4.9, -0…
#> $ n_fog_q3    <dbl> 7.0, 3.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 4.6, 8.0…
#> $ n_fog_q2    <dbl> 5.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 2.0, 6.0…
#> $ n_fog_q4    <dbl> 11.8, 5.8, 1.0, 0.0, 0.8, 0.0, 0.0, 0.0, 0.0, 2.0, 6.8, 10…
#> $ w_med_cv    <dbl> 0.32, 0.26, 0.18, 0.19, 0.21, 0.17, 0.16, 0.18, 0.17, 0.20…
#> $ p_max_max   <dbl> 23.2, 29.0, 32.5, 57.9, 46.2, 64.5, 57.7, 51.9, 70.8, 45.4…
#> $ e_n         <dbl> 29, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 29, 30, 30…
#> $ q_min_md    <dbl> 972.5, 971.3, 971.4, 969.3, 973.6, 975.7, 976.3, 976.6, 97…
#> $ n_fog_cv    <dbl> 0.75, 1.07, 1.34, 3.05, 2.19, 4.03, 5.48, NA, 2.77, 1.14, …
#> $ p_mes_cv    <dbl> 1.06, 0.84, 0.88, 0.77, 0.71, 0.89, 0.93, 1.25, 0.95, 0.78…
#> $ nv_0100_max <dbl> 2, 2, 0, 0, 0, 0, 0, 0, 0, 1, 1, 3, 6, 1, 1, 0, 0, 0, 0, 1…
#> $ np_300_q2   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0…
#> $ q_min_mn    <dbl> 973.4, 972.3, 973.2, 971.0, 973.8, 976.3, 976.2, 976.5, 97…
#> $ ti_max_min  <dbl> -0.6, 0.8, 4.0, 5.2, 10.6, 15.8, 19.6, 18.5, 15.4, 9.5, 2.…
#> $ np_100_cv   <dbl> 1.76, 1.49, 1.64, 1.03, 1.08, 1.17, 1.31, 1.93, 1.23, 1.18…
#> $ hr_q3       <dbl> 75, 68, 60, 58, 55, 49, 48, 52, 58, 70, 74, 78, 62, 73, 67…
#> $ hr_q2       <dbl> 73, 65, 58, 55, 53, 47, 46, 48, 56, 65, 71, 75, 60, 69, 63…
#> $ hr_q1       <dbl> 69, 61, 55, 53, 49, 46, 43, 45, 53, 62, 68, 71, 59, 64, 59…
#> $ nv_0050_s   <dbl> 0.40, 0.46, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.18…
#> $ hr_q4       <dbl> 81, 72, 62, 63, 58, 52, 51, 56, 61, 72, 78, 80, 63, 76, 69…
#> $ evap_s      <dbl> 342.57, 418.08, 399.72, 486.87, 523.22, 738.14, 685.87, 71…
#> $ w_med_q1    <dbl> 11.2, 13.0, 16.0, 16.2, 14.0, 16.0, 15.0, 14.0, 12.0, 11.2…
#> $ evap_n      <dbl> 29, 29, 28, 30, 30, 30, 29, 30, 30, 30, 30, 26, 23, 27, 27…
#> $ nv_0050_n   <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 28, 29…
#> $ nv_0050_q2  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ nv_0050_q3  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ q_max_md    <dbl> 1003.6, 1002.5, 1000.0, 996.0, 994.7, 993.9, 994.3, 993.6,…
#> $ nv_0050_q1  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ nv_0050_q4  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0…
#> $ q_max_mn    <dbl> 1003.1, 1002.9, 1000.5, 996.5, 994.7, 994.0, 993.9, 993.3,…
#> $ n_llu_min   <dbl> 0, 0, 0, 0, 3, 2, 2, 1, 1, 2, 2, 1, 65, 0, 1, 0, 5, 3, 1, …
#> $ ts_min_max  <dbl> 12.6, 12.4, 15.3, 15.6, 20.2, 24.7, 24.0, 24.3, 23.4, 19.5…
#> $ evap_max    <dbl> 1860, 2140, 2503, 2978, 3139, 4881, 4766, 4664, 3071, 2446…
#> $ ta_min_q1   <dbl> -5.4, -3.6, -2.2, 1.1, 5.0, 9.5, 12.4, 12.6, 7.8, 2.8, -2.…
#> $ ta_min_q3   <dbl> -2.7, -1.5, 0.1, 3.2, 6.6, 10.8, 14.1, 14.3, 10.1, 5.8, 0.…
#> $ n_cub_cv    <dbl> 0.39, 0.63, 0.52, 0.47, 0.47, 0.73, 1.35, 1.17, 0.62, 0.64…
#> $ ta_min_q4   <dbl> -0.4, -0.9, 1.6, 3.8, 7.7, 12.5, 15.2, 15.5, 11.4, 6.5, 2.…
#> $ tm_mes_n    <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ tm_max_s    <dbl> 1.86, 2.08, 1.69, 1.52, 2.07, 2.05, 1.59, 1.68, 1.44, 1.80…
#> $ nt_30_md    <dbl> 0.0, 0.0, 0.0, 0.3, 3.6, 14.4, 22.3, 21.1, 7.1, 0.4, 0.0, …
#> $ tm_max_n    <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ tm_mes_s    <dbl> 1.73, 1.59, 1.32, 1.18, 1.62, 1.61, 1.33, 1.28, 1.17, 1.35…
#> $ nt_30_mn    <dbl> 0.0, 0.0, 0.0, 0.0, 4.0, 14.0, 23.0, 22.0, 7.5, 0.0, 0.0, …
#> $ ts_50_q4    <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ ts_50_q3    <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ ts_50_q2    <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ ts_50_q1    <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ q_mar_mn    <dbl> 1021.5, 1020.4, 1018.8, 1014.5, 1014.4, 1015.6, 1015.1, 10…
#> $ np_100_q4   <dbl> 1.0, 1.0, 1.0, 2.0, 2.0, 2.0, 1.0, 1.0, 2.0, 2.0, 2.0, 1.0…
#> $ np_100_q3   <dbl> 0.0, 0.6, 0.0, 1.0, 1.0, 1.0, 0.6, 0.0, 1.0, 1.0, 1.0, 0.0…
#> $ np_100_q2   <dbl> 0.0, 0.0, 0.0, 1.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0…
#> $ np_100_q1   <dbl> 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0…
#> $ n_des_cv    <dbl> 0.78, 0.54, 0.72, 0.75, 0.68, 0.46, 0.25, 0.29, 0.42, 0.63…
#> $ q_mar_md    <dbl> 1022.5, 1020.3, 1018.3, 1014.2, 1014.4, 1015.2, 1015.1, 10…
#> $ tm_mes_q4   <dbl> 8.2, 9.9, 12.9, 14.9, 19.5, 23.8, 26.7, 26.1, 22.4, 17.6, …
#> $ tm_mes_q2   <dbl> 6.2, 7.3, 11.1, 13.6, 17.6, 22.1, 25.0, 24.8, 20.8, 15.8, …
#> $ tm_mes_q3   <dbl> 7.2, 8.4, 11.6, 14.3, 18.5, 22.8, 25.6, 25.4, 21.5, 16.5, …
#> $ np_300_min  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ p_sol_min   <dbl> 23, 31, 41, 39, 41, 50, 61, 56, 49, 41, 28, 23, 50, 26, 34…
#> $ n_cub_min   <dbl> 1, 0, 0, 1, 2, 0, 0, 0, 0, 1, 2, 5, 37, 0, 1, 1, 3, 2, 0, …
#> $ n_gra_s     <dbl> 0.20, 0.00, 0.43, 0.00, 0.41, 0.19, 0.27, 0.42, 0.31, 0.00…
#> $ w_racha_mn  <dbl> 22.1, 22.5, 21.6, 21.4, 19.4, 19.7, 20.6, 19.7, 18.9, 19.4…
#> $ p_max_cv    <dbl> 0.82, 0.82, 1.00, 0.75, 0.66, 1.00, 1.11, 1.32, 1.06, 0.87…
#> $ n_des_q4    <dbl> 7.8, 7.8, 11.0, 7.8, 8.0, 11.0, 18.0, 13.8, 10.0, 8.8, 5.8…
#> $ n_des_q1    <dbl> 1.0, 2.2, 2.0, 2.0, 2.0, 5.0, 12.0, 8.2, 5.2, 2.0, 1.2, 1.…
#> $ n_des_q3    <dbl> 5.0, 5.6, 7.0, 4.6, 5.0, 8.0, 15.0, 11.6, 8.6, 6.0, 4.0, 4…
#> $ n_des_q2    <dbl> 3.0, 4.0, 4.4, 3.0, 3.0, 6.4, 14.0, 10.0, 7.0, 3.4, 3.0, 2…
#> $ ti_max_max  <dbl> 8.5, 13.4, 16.5, 15.2, 21.6, 26.4, 29.8, 28.3, 24.6, 20.2,…
#> $ n_gra_n     <dbl> 25, 29, 30, 27, 29, 29, 27, 27, 28, 29, 30, 26, 22, 30, 30…
#> $ tm_min_n    <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ nt_30_max   <dbl> 0, 0, 0, 3, 9, 24, 30, 29, 16, 4, 0, 0, 94, 0, 0, 0, 0, 6,…
#> $ q_max_max   <dbl> 1011.1, 1013.9, 1011.0, 1001.1, 1002.5, 997.2, 997.8, 996.…
#> $ tm_min_s    <dbl> 1.71, 1.36, 1.20, 0.96, 1.25, 1.22, 1.12, 0.99, 1.02, 1.09…
#> $ nt_00_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ nv_0050_min <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ ts_10_s     <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ n_nub_q4    <dbl> 21.8, 20.0, 21.8, 23.0, 23.8, 22.0, 17.8, 21.0, 21.0, 22.0…
#> $ n_nub_q1    <dbl> 13.4, 16.0, 15.2, 15.0, 18.0, 16.0, 12.2, 14.4, 17.0, 18.2…
#> $ n_nub_q2    <dbl> 16.0, 18.0, 18.4, 19.0, 19.4, 18.4, 14.4, 18.4, 18.4, 19.0…
#> $ n_nub_q3    <dbl> 18.6, 19.0, 20.6, 20.6, 21.6, 21.0, 16.0, 20.0, 19.0, 20.6…
#> $ ts_50_s     <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ ts_10_n     <dbl> 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 0, 0, …
#> $ nt_00_s     <dbl> 5.99, 3.77, 1.69, 0.25, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00…
#> $ ts_10_min   <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ tm_mes_mn   <dbl> 6.7, 8.2, 11.3, 14.0, 18.2, 22.5, 25.3, 25.1, 21.1, 16.2, …
#> $ np_100_min  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 0, 0, 0, 0, 0, 0, 0…
#> $ nv_1000_min <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 0, 0, 0, 0, 0, 0, 0…
#> $ nt_30_s     <dbl> 0.00, 0.00, 0.00, 0.79, 2.85, 5.33, 4.52, 5.15, 4.69, 0.93…
#> $ nt_00_q3    <dbl> 8.0, 5.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 2.0, 6.6…
#> $ nt_00_q2    <dbl> 6.0, 4.0, 0.4, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 4.0…
#> $ nt_00_q1    <dbl> 1.2, 1.2, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 2.0…
#> $ n_tor_min   <dbl> 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 9, 0, 0, 0, 0, 0, 0, 0…
#> $ nt_00_q4    <dbl> 12.8, 8.0, 3.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 3.8, 11…
#> $ ts_20_q3    <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ nt_00_max   <dbl> 20, 14, 7, 1, 0, 0, 0, 0, 0, 0, 8, 21, 49, 15, 11, 8, 3, 0…
#> $ nw_55_min   <dbl> 1, 2, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 26, 0, 0, 0, 0, 0, 0, …
#> $ n_llu_mn    <dbl> 10.0, 9.0, 7.5, 9.0, 11.0, 8.0, 4.0, 5.0, 7.0, 10.0, 10.5,…
#> $ ts_50_mn    <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ glo_min     <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ ts_50_md    <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ q_med_min   <dbl> 978.6, 977.3, 980.4, 977.8, 979.5, 981.2, 983.7, 982.2, 98…
#> $ nv_0050_cv  <dbl> 4.03, 2.77, NA, NA, NA, NA, NA, NA, NA, 5.48, 5.48, 2.68, …
#> $ inso_q2     <dbl> 3.9, 5.5, 6.5, 7.3, 8.7, 10.2, 11.1, 10.0, 7.8, 6.0, 4.9, …
#> $ inso_q3     <dbl> 4.3, 6.4, 7.4, 8.2, 9.3, 10.8, 11.5, 10.5, 8.1, 6.7, 5.5, …
#> $ inso_q1     <dbl> 3.5, 5.1, 6.2, 6.2, 7.7, 8.8, 10.5, 9.5, 7.1, 5.2, 3.8, 3.…
#> $ n_fog_max   <dbl> 17, 12, 1, 2, 2, 2, 1, 0, 2, 4, 12, 14, 36, 11, 11, 6, 5, …
#> $ np_100_s    <dbl> 0.82, 0.89, 0.55, 1.17, 1.26, 0.90, 0.57, 0.77, 1.19, 1.22…
#> $ np_300_s    <dbl> 0.00, 0.00, 0.25, 0.35, 0.31, 0.35, 0.25, 0.31, 0.38, 0.35…
#> $ ta_min_cv   <dbl> -0.76, -0.86, -5.83, 0.58, 0.28, 0.16, 0.11, 0.11, 0.18, 0…
#> $ tm_mes_cv   <dbl> 0.26, 0.19, 0.11, 0.09, 0.09, 0.07, 0.05, 0.05, 0.06, 0.08…
#> $ ts_min_md   <dbl> 9.0, 9.6, 11.3, 13.1, 16.8, 20.8, 22.2, 22.0, 19.7, 16.2, …
#> $ ts_min_mn   <dbl> 9.3, 10.2, 11.0, 13.6, 16.8, 20.6, 22.3, 22.0, 19.7, 16.5,…
#> $ inso_cv     <dbl> 0.23, 0.20, 0.14, 0.16, 0.13, 0.11, 0.09, 0.08, 0.12, 0.16…
#> $ np_300_n    <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ np_100_n    <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ ts_min_n    <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ tm_max_md   <dbl> 10.5, 13.1, 17.3, 19.6, 24.1, 29.3, 32.4, 31.7, 27.1, 21.4…
#> $ tm_max_mn   <dbl> 10.9, 13.1, 17.3, 19.9, 24.3, 29.1, 32.4, 32.0, 27.1, 21.4…
#> $ ts_min_s    <dbl> 1.92, 1.87, 1.60, 1.46, 1.60, 1.40, 1.05, 1.18, 1.36, 1.62…
#> $ tm_mes_q1   <dbl> 4.7, 6.9, 10.7, 12.9, 16.8, 21.7, 24.2, 23.6, 20.1, 15.1, …
#> $ p_max_q1    <dbl> 2.4, 2.8, 2.5, 6.3, 7.4, 3.7, 2.0, 1.0, 1.7, 5.0, 6.2, 3.4…
#> $ p_max_q3    <dbl> 7.9, 10.4, 7.3, 16.4, 19.3, 12.8, 10.3, 6.8, 16.3, 12.5, 1…
#> $ p_max_q2    <dbl> 5.7, 5.8, 4.1, 11.3, 12.7, 7.4, 5.7, 3.8, 6.3, 8.3, 7.9, 5…
#> $ p_max_q4    <dbl> 12.8, 18.2, 15.8, 22.3, 26.3, 20.5, 15.1, 13.8, 26.8, 26.5…
#> $ n_nie_s     <dbl> 1.28, 0.83, 0.48, 0.19, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00…
#> $ n_gra_min   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ nt_30_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ n_nie_n     <dbl> 25, 29, 30, 27, 29, 29, 27, 27, 28, 29, 30, 26, 22, 30, 30…
#> $ ts_50_n     <dbl> 11, 11, 11, 11, 11, 11, 10, 10, 9, 11, 11, 11, 9, 0, 0, 0,…
#> $ nv_0100_min <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ q_mar_q1    <dbl> 1017.5, 1016.0, 1014.5, 1012.3, 1012.5, 1013.7, 1013.9, 10…
#> $ q_mar_q2    <dbl> 1020.4, 1019.6, 1017.6, 1014.0, 1013.7, 1015.3, 1014.7, 10…
#> $ q_mar_q3    <dbl> 1024.1, 1021.5, 1019.1, 1015.2, 1014.8, 1015.7, 1015.5, 10…
#> $ q_mar_q4    <dbl> 1027.9, 1025.4, 1021.1, 1016.0, 1016.2, 1016.6, 1016.6, 10…
#> $ p_max_min   <dbl> 0.0, 0.0, 0.0, 1.4, 2.7, 0.0, 0.2, 0.1, 0.0, 0.5, 0.4, 0.6…
#> $ n_llu_cv    <dbl> 0.62, 0.53, 0.50, 0.48, 0.32, 0.47, 0.50, 0.48, 0.49, 0.48…
#> $ n_nub_mn    <dbl> 18.0, 19.0, 19.0, 20.0, 21.0, 20.5, 16.0, 19.5, 19.0, 20.0…
#> $ w_racha_q1  <dbl> 19.2, 20.6, 18.9, 18.9, 18.1, 17.6, 17.5, 16.9, 17.5, 18.1…
#> $ w_racha_q3  <dbl> 23.4, 23.3, 22.7, 21.9, 21.0, 20.8, 21.7, 20.3, 20.5, 20.9…
#> $ w_racha_q2  <dbl> 21.3, 22.2, 21.1, 20.0, 18.9, 18.8, 20.0, 18.6, 18.6, 19.2…
#> $ n_nub_md    <dbl> 17.5, 17.9, 18.8, 18.9, 20.7, 18.9, 15.2, 18.3, 18.8, 20.2…
#> $ w_racha_q4  <dbl> 24.9, 24.7, 25.6, 23.1, 25.8, 24.7, 24.7, 22.2, 22.5, 22.2…
#> $ nt_00_cv    <dbl> 0.79, 0.74, 1.24, 3.81, NA, NA, NA, NA, NA, NA, 1.25, 0.78…
#> $ np_010_q4   <dbl> 7.0, 6.0, 5.0, 8.8, 8.0, 6.0, 4.0, 3.8, 5.0, 9.0, 7.0, 8.0…
#> $ np_010_q1   <dbl> 1.2, 2.0, 2.0, 3.2, 4.0, 2.0, 1.0, 1.0, 1.0, 2.0, 2.0, 3.0…
#> $ np_010_q3   <dbl> 4.0, 5.0, 4.0, 6.0, 7.0, 4.6, 2.6, 2.6, 4.0, 6.2, 5.6, 4.0…
#> $ np_010_q2   <dbl> 3.0, 3.0, 3.0, 4.0, 5.4, 3.0, 2.0, 2.0, 2.4, 4.0, 5.0, 3.0…
#> $ nv_1000_q3  <dbl> 4.6, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 2.6, 4.0…
#> $ ts_50_cv    <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ nv_0100_s   <dbl> 0.46, 0.48, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.18…
#> $ n_llu_max   <dbl> 17, 16, 17, 20, 20, 18, 11, 11, 13, 21, 16, 19, 120, 19, 1…
#> $ tm_max_min  <dbl> 6.8, 10.0, 13.7, 15.2, 17.9, 24.2, 29.0, 28.6, 24.7, 17.3,…
#> $ tm_min_md   <dbl> 2.7, 3.3, 5.8, 7.9, 11.8, 15.8, 18.3, 18.3, 15.2, 11.0, 6.…
#> $ q_mar_cv    <dbl> 0.01, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00…
#> $ nv_0100_n   <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 28, 29…
#> $ tm_min_mn   <dbl> 2.8, 3.2, 5.6, 8.2, 12.0, 15.7, 18.1, 18.2, 15.1, 10.9, 6.…
#> $ ts_10_mn    <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ w_racha_cv  <dbl> 0.17, 0.15, 0.16, 0.12, 0.19, 0.20, 0.18, 0.14, 0.14, 0.16…
#> $ evap_min    <dbl> 358, 519, 947, 1004, 1327, 1374, 2298, 1732, 1146, 813, 41…
#> $ ta_max_cv   <dbl> 0.11, 0.09, 0.08, 0.09, 0.08, 0.06, 0.05, 0.05, 0.07, 0.08…
#> $ n_llu_q4    <dbl> 12.0, 11.0, 10.8, 14.4, 14.0, 10.0, 8.0, 7.0, 11.0, 14.0, …
#> $ n_llu_q2    <dbl> 8.4, 6.0, 7.0, 9.0, 10.0, 6.0, 4.0, 4.0, 5.6, 10.0, 9.0, 9…
#> $ n_llu_q3    <dbl> 10.6, 10.0, 8.0, 11.0, 11.0, 8.0, 4.8, 5.8, 7.4, 11.0, 11.…
#> $ n_llu_q1    <dbl> 3.0, 4.0, 5.0, 5.6, 9.0, 4.0, 3.0, 3.0, 4.8, 5.0, 6.0, 7.0…
#> $ q_min_q4    <dbl> 980.9, 977.7, 978.5, 974.0, 977.2, 977.6, 978.3, 978.3, 97…
#> $ hr_min      <dbl> 63, 60, 50, 49, 46, 40, 40, 41, 49, 57, 55, 66, 56, 53, 52…
#> $ w_med_max   <dbl> 28, 27, 25, 26, 29, 27, 25, 21, 19, 19, 22, 25, 21, 11, 13…
#> $ nw_91_mn    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0…
#> $ q_min_q1    <dbl> 964.5, 966.2, 964.9, 965.3, 970.3, 973.6, 974.6, 974.4, 97…
#> $ q_min_q2    <dbl> 969.1, 969.9, 970.1, 968.4, 972.7, 975.4, 975.8, 976.3, 97…
#> $ q_min_q3    <dbl> 974.5, 973.5, 974.1, 972.7, 975.5, 976.9, 977.0, 977.1, 97…
#> $ q_min_n     <dbl> 30, 30, 30, 29, 29, 30, 30, 30, 29, 29, 29, 29, 28, 30, 30…
#> $ nw_91_md    <dbl> 0.5, 0.2, 0.2, 0.1, 0.2, 0.2, 0.2, 0.0, 0.0, 0.1, 0.0, 0.2…
#> $ q_min_s     <dbl> 8.91, 8.38, 7.57, 5.16, 3.43, 2.49, 2.03, 2.24, 3.99, 7.01…
#> $ n_tor_q4    <dbl> 0.0, 0.0, 1.0, 2.8, 6.0, 6.0, 6.0, 5.0, 4.0, 1.8, 0.0, 0.0…
#> $ n_tor_q1    <dbl> 0.0, 0.0, 0.0, 0.0, 2.0, 2.0, 2.0, 1.0, 1.0, 0.0, 0.0, 0.0…
#> $ n_tor_q2    <dbl> 0.0, 0.0, 0.0, 1.0, 3.0, 3.0, 3.0, 3.4, 2.0, 1.0, 0.0, 0.0…
#> $ n_tor_q3    <dbl> 0.0, 0.0, 0.0, 1.0, 4.0, 4.6, 4.0, 4.6, 3.0, 1.0, 0.0, 0.0…
#> $ tm_mes_min  <dbl> 3.2, 5.5, 8.9, 10.4, 13.2, 18.5, 22.7, 22.9, 19.1, 13.6, 8…
#> $ n_cub_max   <dbl> 16, 14, 12, 14, 12, 8, 6, 8, 8, 14, 16, 19, 102, 16, 11, 1…
#> $ nw_55_cv    <dbl> 0.72, 0.47, 0.55, 0.48, 0.67, 0.62, 0.73, 0.66, 0.66, 0.59…
#> $ tm_min_min  <dbl> -0.5, 0.8, 4.1, 5.5, 8.5, 12.8, 16.3, 16.5, 13.5, 8.6, 4.3…
#> $ q_mar_max   <dbl> 1034.6, 1027.7, 1028.5, 1017.4, 1018.3, 1017.4, 1017.4, 10…
#> $ nv_0100_q4  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 2, 0, 0, 0, 0, 0, 0, 0…
#> $ nv_0100_q1  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ nv_0100_q2  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ nv_0100_q3  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0…
#> $ q_max_q2    <dbl> 1001.7, 1001.5, 998.9, 994.8, 994.2, 993.9, 993.5, 993.0, …
#> $ q_max_q3    <dbl> 1005.0, 1004.3, 1001.1, 996.9, 995.3, 994.3, 994.5, 994.1,…
#> $ q_max_q1    <dbl> 1000.6, 998.0, 996.7, 993.5, 992.3, 992.1, 992.8, 992.6, 9…
#> $ q_max_q4    <dbl> 1007.9, 1006.2, 1002.8, 998.8, 996.8, 995.7, 996.5, 995.2,…
#> $ nv_0100_cv  <dbl> 2.77, 2.42, NA, NA, NA, NA, NA, NA, NA, 5.48, 2.59, 1.57, …
#> $ ts_10_q3    <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ ts_10_q2    <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ ts_10_q1    <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ ts_10_q4    <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ q_max_cv    <dbl> 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00…
#> $ ta_max_max  <dbl> 20.0, 22.2, 28.0, 31.7, 36.5, 40.5, 43.1, 42.8, 38.0, 31.2…
#> $ np_010_cv   <dbl> 0.83, 0.62, 0.61, 0.55, 0.47, 0.66, 0.67, 0.80, 0.62, 0.64…
#> $ ta_max_q4   <dbl> 18.9, 20.6, 26.4, 29.8, 33.9, 38.5, 40.5, 39.0, 35.7, 29.7…
#> $ ta_max_q3   <dbl> 17.7, 19.9, 24.7, 28.3, 33.0, 36.8, 39.1, 38.0, 34.5, 28.5…
#> $ ta_max_q2   <dbl> 16.3, 18.8, 23.9, 26.7, 31.5, 36.0, 38.3, 37.0, 33.2, 27.0…
#> $ ta_max_q1   <dbl> 15.8, 17.0, 23.1, 25.2, 29.7, 34.9, 37.6, 35.6, 31.2, 25.9…
#> $ np_010_min  <dbl> 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 37, 0, 0, 0, 2, 1, 0, …
#> $ w_med_n     <dbl> 30, 30, 30, 30, 29, 30, 30, 30, 30, 30, 30, 30, 29, 28, 28…
#> $ nv_1000_q4  <dbl> 6.0, 3.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 4.0, 7.8…
#> $ nv_1000_q2  <dbl> 2, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 3, 11, 2, 0, 0, 0, 0, 0, …
#> $ n_nie_max   <dbl> 4, 3, 2, 1, 0, 0, 0, 0, 0, 0, 1, 4, 7, 5, 5, 2, 4, 0, 0, 0…
#> $ nv_1000_q1  <dbl> 0.2, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.2, 2.0…
#> $ nv_1000_cv  <dbl> 0.87, 1.43, 3.05, 5.48, 2.59, NA, NA, NA, 3.05, 1.46, 0.93…
#> $ glo_max     <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ n_fog_s     <dbl> 4.87, 3.06, 0.49, 0.61, 0.58, 0.40, 0.18, 0.00, 0.46, 1.14…
#> $ ti_max_q2   <dbl> 3.3, 6.1, 9.5, 11.4, 14.5, 19.4, 24.1, 24.0, 19.1, 14.5, 8…
#> $ n_fog_n     <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ n_nie_cv    <dbl> 1.77, 1.85, 2.42, 5.20, NA, NA, NA, NA, NA, NA, 3.05, 1.91…
#> $ np_001_min  <dbl> 0, 0, 0, 2, 2, 0, 1, 1, 0, 2, 1, 1, 59, 0, 1, 0, 4, 2, 1, …
#> $ nw_55_q4    <dbl> 13.2, 11.0, 14.8, 13.0, 11.4, 8.6, 11.0, 8.0, 6.8, 7.4, 10…
#> $ n_fog_min   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 13, 0, 0, 0, 0, 0, 0, …
#> $ nw_55_q2    <dbl> 4.6, 7.0, 7.0, 7.0, 5.0, 5.0, 4.0, 4.0, 4.0, 4.0, 4.6, 5.6…
#> $ nw_55_q3    <dbl> 8.4, 9.0, 9.6, 10.0, 6.0, 6.0, 6.0, 5.0, 4.8, 5.0, 6.4, 7.…
#> $ nw_55_q1    <dbl> 2.0, 5.0, 5.0, 5.0, 2.6, 3.0, 3.0, 2.0, 2.0, 2.0, 2.8, 2.8…
#> $ hr_n        <dbl> 29, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 29, 28, 27…
#> $ hr_s        <dbl> 5.66, 5.45, 4.48, 5.18, 4.83, 4.90, 3.80, 4.97, 4.61, 5.53…
#> $ q_med_s     <dbl> 5.53, 4.77, 3.65, 2.11, 2.11, 1.51, 1.20, 1.26, 2.05, 2.80…
#> $ nv_0050_max <dbl> 2, 2, 0, 0, 0, 0, 0, 0, 0, 1, 1, 3, 3, 0, 0, 0, 0, 0, 0, 9…
#> $ ta_min_q2   <dbl> -4.1, -2.9, -0.8, 2.5, 6.0, 10.4, 13.4, 13.4, 9.4, 4.5, -1…
#> $ n_des_max   <dbl> 14, 10, 20, 16, 10, 18, 21, 19, 20, 13, 16, 12, 107, 21, 1…
#> $ n_fog_mn    <dbl> 6.0, 1.5, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 3.5, 7.0…
#> $ p_sol_max   <dbl> 72, 72, 80, 69, 74, 79, 87, 82, 76, 71, 77, 64, 66, 77, 70…
#> $ np_001_s    <dbl> 4.89, 3.59, 3.14, 3.76, 3.57, 3.20, 2.13, 2.22, 2.84, 4.07…
#> $ n_fog_md    <dbl> 6.5, 2.9, 0.4, 0.2, 0.3, 0.1, 0.0, 0.0, 0.2, 1.0, 3.9, 7.1…
#> $ q_min_cv    <dbl> 0.01, 0.01, 0.01, 0.01, 0.00, 0.00, 0.00, 0.00, 0.00, 0.01…
#> $ np_001_n    <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ e_max       <dbl> 97, 99, 104, 115, 143, 170, 192, 219, 181, 157, 121, 99, 1…
#> $ nt_00_md    <dbl> 7.6, 5.1, 1.4, 0.1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.9, 6.5…
#> $ np_100_mn   <dbl> 0.0, 0.0, 0.0, 1.0, 1.0, 0.5, 0.0, 0.0, 0.5, 1.0, 0.0, 0.0…
#> $ nt_30_q1    <dbl> 0.0, 0.0, 0.0, 0.0, 0.2, 10.2, 17.2, 15.4, 2.0, 0.0, 0.0, …
#> $ nt_00_mn    <dbl> 6.5, 5.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 5.0…
#> $ glo_q4      <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ glo_q2      <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ glo_q3      <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ np_100_md   <dbl> 0.5, 0.6, 0.3, 1.1, 1.2, 0.8, 0.4, 0.4, 1.0, 1.0, 0.7, 0.4…
#> $ inso_n      <dbl> 30, 30, 30, 30, 30, 30, 30, 29, 29, 30, 30, 30, 29, 23, 24…
#> $ q_med_q1    <dbl> 986.2, 984.9, 983.7, 982.1, 982.6, 984.3, 984.6, 984.7, 98…
#> $ np_001_md   <dbl> 7.8, 6.9, 6.3, 8.5, 9.0, 6.1, 4.0, 3.9, 5.2, 8.1, 8.2, 9.0…
#> $ np_001_mn   <dbl> 7.0, 6.5, 6.0, 7.0, 8.0, 5.5, 3.5, 4.0, 5.0, 8.0, 8.0, 8.0…
#> $ ti_max_cv   <dbl> 0.74, 0.41, 0.30, 0.18, 0.20, 0.15, 0.10, 0.08, 0.12, 0.18…
#> $ nv_1000_mn  <dbl> 2.5, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.5, 4.0…
#> $ nv_1000_md  <dbl> 3.5, 1.6, 0.1, 0.0, 0.1, 0.0, 0.0, 0.0, 0.1, 0.5, 2.2, 4.5…
#> $ evap_md     <dbl> 830, 1077, 1691, 1814, 2278, 2905, 3333, 3074, 2142, 1451,…
#> $ ta_min_max  <dbl> 2.6, 1.6, 4.8, 4.0, 9.8, 15.1, 17.0, 16.5, 13.5, 9.4, 3.8,…
#> $ e_md        <dbl> 75, 75, 83, 94, 118, 141, 158, 164, 147, 126, 95, 78, 113,…
#> $ e_mn        <dbl> 74, 73, 84, 93, 116, 141, 158, 162, 146, 124, 96, 79, 112,…
#> $ nt_30_q3    <dbl> 0.0, 0.0, 0.0, 0.0, 5.0, 15.0, 24.0, 23.6, 9.0, 0.0, 0.0, …
#> $ ti_max_md   <dbl> 3.6, 6.9, 10.0, 11.4, 15.7, 20.6, 24.5, 24.2, 19.5, 14.7, …
#> $ n_llu_n     <dbl> 25, 29, 30, 27, 29, 29, 27, 27, 28, 29, 30, 26, 22, 30, 30…
#> $ ti_max_mn   <dbl> 4.1, 6.5, 10.4, 11.7, 16.0, 20.8, 24.6, 24.4, 19.5, 14.8, …
#> $ glo_s       <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ glo_n       <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ ts_min_cv   <dbl> 0.21, 0.19, 0.14, 0.11, 0.09, 0.07, 0.05, 0.05, 0.07, 0.10…
#> $ n_llu_s     <dbl> 5.37, 4.14, 3.72, 4.82, 3.48, 3.53, 2.52, 2.44, 3.47, 4.87…
#> $ nv_0050_mn  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ n_nie_q3    <dbl> 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0…
#> $ n_nie_q2    <dbl> 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0…
#> $ n_nie_q4    <dbl> 1.8, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0…
#> $ glo_cv      <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ tm_max_max  <dbl> 13.2, 18.0, 20.7, 21.4, 27.3, 33.8, 35.6, 35.5, 30.7, 24.5…
#> $ ts_20_mn    <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ tm_max_cv   <dbl> 0.18, 0.16, 0.10, 0.08, 0.09, 0.07, 0.05, 0.05, 0.05, 0.08…
#> $ ts_20_md    <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ np_001_cv   <dbl> 0.63, 0.52, 0.50, 0.44, 0.40, 0.52, 0.53, 0.57, 0.54, 0.50…
#> $ p_max_mn    <dbl> 6.5, 6.7, 5.2, 13.2, 16.1, 10.6, 7.2, 4.5, 9.8, 11.0, 9.1,…
#> $ nw_91_max   <dbl> 6, 1, 2, 1, 1, 2, 2, 0, 1, 2, 1, 3, 10, 0, 0, 0, 1, 0, 0, …
#> $ p_max_md    <dbl> 7.7, 9.8, 9.2, 16.8, 17.5, 13.7, 10.7, 9.4, 16.0, 14.8, 12…
#> $ q_med_cv    <dbl> 0.01, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00…
#> $ ts_50_max   <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ ts_50_min   <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ ta_min_n    <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ e_q4        <dbl> 82, 85, 92, 103, 128, 151, 168, 176, 158, 137, 104, 88, 11…
#> $ e_q3        <dbl> 78, 77, 85, 98, 119, 142, 160, 163, 151, 127, 98, 81, 113,…
#> $ e_q2        <dbl> 72, 71, 80, 90, 113, 135, 152, 159, 140, 119, 93, 74, 111,…
#> $ e_q1        <dbl> 68, 67, 73, 86, 107, 131, 146, 150, 131, 116, 82, 69, 109,…
#> $ glo_q1      <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ w_racha_md  <dbl> 22.2, 22.7, 22.2, 21.3, 20.8, 20.7, 20.9, 19.6, 19.8, 19.9…
#> $ q_min_min   <dbl> 955.2, 947.9, 953.7, 957.5, 966.7, 970.3, 972.0, 971.1, 96…
#> $ ts_min_q3   <dbl> 9.5, 10.6, 11.4, 13.9, 17.3, 21.1, 22.5, 22.3, 20.2, 16.6,…
#> $ ts_min_q2   <dbl> 8.4, 9.2, 10.9, 12.8, 16.5, 20.3, 22.0, 21.5, 19.3, 16.2, …
#> $ ts_min_q1   <dbl> 7.1, 7.9, 10.2, 11.7, 15.5, 19.6, 21.1, 20.9, 18.4, 14.1, …
#> $ p_mes_max   <dbl> 81.0, 70.5, 62.7, 126.6, 141.9, 82.8, 58.3, 88.6, 101.4, 1…
#> $ ts_min_q4   <dbl> 10.8, 11.5, 12.4, 14.4, 18.1, 21.8, 23.2, 23.0, 20.6, 17.6…
#> $ q_mar_min   <dbl> 1009.4, 1008.3, 1010.7, 1008.0, 1009.8, 1010.6, 1012.6, 10…
#> $ tm_max_q3   <dbl> 11.2, 13.4, 17.6, 20.4, 25.1, 29.6, 32.6, 32.1, 27.5, 21.7…
#> $ tm_max_q2   <dbl> 10.1, 11.9, 16.8, 19.3, 23.4, 28.8, 32.0, 31.4, 26.9, 20.8…
#> $ tm_max_q1   <dbl> 8.8, 11.2, 16.2, 18.3, 22.8, 28.3, 30.8, 30.0, 25.6, 19.9,…
#> $ p_mes_mn    <dbl> 14.1, 21.3, 13.2, 33.7, 34.0, 18.4, 10.5, 8.4, 22.5, 27.3,…
#> $ tm_max_q4   <dbl> 12.3, 15.1, 18.6, 21.0, 26.0, 31.0, 34.1, 33.1, 28.3, 23.2…
#> $ n_gra_mn    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ tm_min_max  <dbl> 6.0, 6.1, 8.9, 9.6, 13.7, 19.4, 20.8, 20.2, 17.5, 13.4, 9.…
#> $ np_001_q4   <dbl> 12.6, 10.0, 9.0, 11.0, 11.8, 8.8, 6.0, 5.8, 7.8, 11.6, 11.…
#> $ np_001_q3   <dbl> 9.0, 8.0, 7.0, 9.6, 9.0, 7.0, 4.0, 4.0, 5.0, 8.6, 9.0, 9.0…
#> $ np_001_q2   <dbl> 6.0, 6.0, 5.4, 7.0, 8.0, 5.0, 3.0, 3.0, 4.0, 7.4, 7.4, 7.4…
#> $ np_001_q1   <dbl> 3.0, 4.0, 4.0, 5.0, 6.0, 4.0, 2.0, 2.0, 3.0, 4.2, 5.0, 5.2…
#> $ ti_max_n    <dbl> 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30, 30…
#> $ w_med_md    <dbl> 16.5, 16.9, 18.5, 18.6, 17.6, 17.8, 17.8, 16.4, 14.3, 14.0…
#> $ ti_max_q4   <dbl> 5.8, 8.5, 12.0, 13.2, 18.4, 23.3, 26.4, 25.7, 21.6, 16.8, …
#> $ ti_max_q3   <dbl> 4.5, 7.0, 11.2, 11.8, 17.0, 21.5, 25.2, 24.7, 20.1, 15.5, …
#> $ n_gra_q2    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ ti_max_q1   <dbl> 1.0, 4.7, 7.0, 10.2, 12.8, 17.6, 22.4, 22.5, 17.1, 12.1, 6…
#> $ ti_max_s    <dbl> 2.67, 2.79, 3.05, 2.07, 3.09, 3.03, 2.48, 2.02, 2.37, 2.67…
#> $ q_med_q2    <dbl> 989.2, 988.2, 986.7, 983.6, 983.9, 985.8, 985.7, 985.1, 98…
#> $ q_med_q3    <dbl> 992.7, 990.2, 988.2, 984.9, 985.0, 986.3, 986.4, 986.1, 98…
#> $ nt_30_q2    <dbl> 0.0, 0.0, 0.0, 0.0, 2.0, 13.4, 22.0, 21.0, 6.0, 0.0, 0.0, …
#> $ n_nie_min   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ nt_30_q4    <dbl> 0.0, 0.0, 0.0, 0.0, 7.0, 19.8, 26.0, 26.0, 11.0, 0.0, 0.0,…
#> $ q_med_q4    <dbl> 996.1, 994.0, 990.0, 985.7, 986.5, 987.2, 987.3, 986.5, 98…
#> $ n_cub_md    <dbl> 8.9, 5.2, 5.5, 6.5, 5.8, 2.9, 1.2, 1.8, 3.1, 5.4, 7.1, 9.7…
#> $ n_cub_mn    <dbl> 9.0, 4.0, 5.5, 6.0, 5.0, 2.5, 1.0, 1.0, 3.0, 4.5, 7.0, 8.5…
#> $ ta_max_min  <dbl> 12.8, 15.9, 19.6, 22.6, 24.8, 31.6, 34.4, 33.2, 29.6, 22.8…
#> $ np_010_max  <dbl> 15, 8, 10, 15, 16, 11, 6, 7, 7, 13, 12, 16, 73, 14, 13, 11…
#> $ inso_md     <dbl> 4.3, 5.9, 7.0, 7.6, 8.9, 10.3, 11.2, 10.2, 8.1, 6.3, 4.9, …
#> $ inso_mn     <dbl> 4.1, 5.9, 6.8, 7.8, 9.1, 10.6, 11.4, 10.2, 7.9, 6.5, 5.2, …
#> $ evap_mn     <dbl> 774, 1018, 1720, 1782, 2328, 2877, 3422, 3025, 2166, 1405,…

Site built with pkgdown 2.0.7.

Template rostemplate by dieghernan, based on Bootstrapious.