Station climate stripes graph

Plot climate stripes graph for a station.

Usage

climatestripes_station(
  station,
  start = 1950,
  end = 2020,
  with_labels = "yes",
  verbose = FALSE,
  ...
)

Arguments

station

Character string with station identifier code(s) (see aemet_stations()).

start

Numeric value as start year (format: YYYY).

end

Numeric value as end year (format: YYYY).

with_labels

Character string as yes/no. Indicates whether to use labels for the graph or not.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

...

Arguments passed on to ggstripes

n_temp

Numeric value as the number of colors of the palette. (default 11).

col_pal

Character string indicating the name of the hcl.pals() color palette to be used for plotting.

Value

A ggplot2 object

API Key

You need to set your API Key globally using aemet_api_key().

See also

Examples

# \donttest{

# Don't run example
if (FALSE) {
  # Data download may take a few minutes...
  climatestripes_station(
    "9434",
    start = 2020,
    end = 2024,
    with_labels = "yes",
    col_pal = "Inferno"
  )
}
# }

Site built with pkgdown 2.2.0.

Template rostemplate by dieghernan, based on Universal Theme for Hugo.