Client tool for AEMET API

Client tool to get data and metadata from AEMET and convert json to tibble.

Usage

get_data_aemet(apidest, verbose = FALSE)

get_metadata_aemet(apidest, verbose = FALSE)

Source

https://opendata.aemet.es/dist/index.html

Arguments

apidest

Character string as destination URL. See https://opendata.aemet.es/dist/index.html.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

Value

A tibble (if possible) or the results of the query as provided by httr2::resp_body_raw() or httr2::resp_body_string().

See also

Examples

# Run this example only if AEMET_API_KEY is detected

url <- "/api/valores/climatologicos/inventarioestaciones/todasestaciones"

get_data_aemet(url)
#> # A tibble: 291 × 7
#>    latitud provincia     altitud indicativo nombre             indsinop longitud
#>    <chr>   <chr>         <chr>   <chr>      <chr>              <chr>    <chr>   
#>  1 394924N ILLES BALEARS 490     B013X      "ESCORCA, LLUC"    08304    025309E 
#>  2 393319N ILLES BALEARS 3       B228       "PALMA, PUERTO"    08301    023731E 
#>  3 394406N ILLES BALEARS 1030    B248       "SIERRA DE ALFABI… 08303    024247E 
#>  4 393339N ILLES BALEARS 8       B278       "PALMA DE MALLORC… 08306    024412E 
#>  5 393118N ILLES BALEARS 120     B346X      "PORRERES"         08310    030109E 
#>  6 392452N ILLES BALEARS 17      B434X      "PORTOCOLOM  "     08309    031618E 
#>  7 394300N ILLES BALEARS 57      B569X      "CAPDEPERA"        08311    032841E 
#>  8 394457N ILLES BALEARS 40      B691Y      "SA POBLA"         08312    030100E 
#>  9 395117N ILLES BALEARS 91      B893       "MENORCA, AEROPUE… 08314    041256E 
#> 10 385235N ILLES BALEARS 6       B954       "IBIZA, AEROPUERT… 08373    012304E 
#> # ℹ 281 more rows


# Metadata

get_metadata_aemet(url)
#> # A tibble: 7 × 7
#>   unidad_generadora         periodicidad descripcion formato copyright notaLegal
#>   <chr>                     <chr>        <chr>       <chr>   <chr>     <chr>    
#> 1 Servicio del Banco de Da… 1 vez al día Inventario… applic… © AEMET.… https://…
#> 2 Servicio del Banco de Da… 1 vez al día Inventario… applic… © AEMET.… https://…
#> 3 Servicio del Banco de Da… 1 vez al día Inventario… applic… © AEMET.… https://…
#> 4 Servicio del Banco de Da… 1 vez al día Inventario… applic… © AEMET.… https://…
#> 5 Servicio del Banco de Da… 1 vez al día Inventario… applic… © AEMET.… https://…
#> 6 Servicio del Banco de Da… 1 vez al día Inventario… applic… © AEMET.… https://…
#> 7 Servicio del Banco de Da… 1 vez al día Inventario… applic… © AEMET.… https://…
#> # ℹ 1 more variable: campos <df[,4]>

# We can get data from any API endpoint

# Plain text

plain <- get_data_aemet("/api/prediccion/nacional/hoy")
#> 
#> Results are MIME type: text/plain
#> Returning data as string

cat(plain)
#> AGENCIA ESTATAL DE METEOROLOGÍA
#> PREDICCIÓN GENERAL PARA ESPAÑA 
#> DÍA 03 DE MARZO DE 2024 A LAS 08:42 HORA OFICIAL
#> PREDICCIÓN VÁLIDA PARA EL DOMINGO 3
#> 
#> A.- FENÓMENOS SIGNIFICATIVOS
#> Precipitaciones localmente persistentes en el Pirineo y
#> acumulaciones significativas de nieve en áreas de montaña del
#> norte y centro peninsulares y zonas próximas. Intervalos de
#> viento fuerte con rachas muy fuertes en áreas de litoral,
#> interiores del tercio este peninsular y en Baleares.
#> 
#> B.- PREDICCIÓN
#> En la Península y Baleares se prevé una circulación atlántica
#> con la acción de dos frentes. El primero terminará de recorrer
#> el territorio, dejando cielos nubosos o cubiertos con
#> precipitaciones a primeras horas en el Estrecho, Alborán y,
#> acompañadas de tormenta, en Cataluña y Baleares. Tras su paso se
#> mantendrá un ambiente inestable, con cielos nubosos y chubascos
#> en Pirineos y, ocasionalmente acompañados de tormenta, en Galicia
#> y Cantábrico. Intervalos nubosos en el resto de la Península,
#> pudiendo producirse precipitaciones dispersas y ocasionales, más
#> probables en montaña, tendiendo a remitir de oeste a este. La
#> aproximación y entrada de un frente cálido aumentará la
#> nubosidad en el noroeste, con lluvias débiles en Galicia, sin
#> descartar que se extiendan hacia el este. Serán de nieve en
#> montaña y proximidades. Cielos nubosos en el norte de Canarias,
#> con precipitaciones débiles en las islas de mayor relieve, sin
#> descartar las orientales e intervalos nubosos en el sur.
#> 
#> Probables brumas matinales en montaña.
#> 
#> Cota de nieve: en el noroeste 500/900 m subiendo a 1000/1800 m,
#> en el centro norte 600/800 m, en el nordeste 700/1200 m bajando a
#> cualquier cota, en el centro 600/800 m subiendo a 800/1200 m y en
#> el sudeste 700/1000 m subiendo a 1000/1400 m.
#> 
#> Las temperaturas máximas tenderán a descender en las mitades
#> este y sudeste peninsulares y Baleares y a aumentar en el resto.
#> Las mínimas tenderán a descender en general, salvo en el tercio
#> sur donde aumentarán y Canarias sin grandes cambios. Heladas
#> débiles en amplias zonas de interior de la mitad norte peninsular
#> y sierras del sudeste, más intensas en montaña.
#> 
#> Soplará viento intenso de componente oeste en la Península y
#> Baleares, con intervalos de fuerte y/o rachas muy fuertes en
#> áreas de litoral e interiores del tercio este peninsular. Alisio
#> moderado en Canarias.
#> 

# An image

image <- get_data_aemet("/api/mapasygraficos/analisis")
#> 
#> Results are MIME type: image/gif
#> Returning raw data

# Write and read
tmp <- tempfile(fileext = ".gif")

writeBin(image, tmp)

gganimate::gif_file(tmp)
#> Error in shell.exec(url): file association for 'C:\Users\RUNNER~1\AppData\Local\Temp\RtmpIlXgmT\file98029ca3e8.gif' not available or invalid

Site built with pkgdown 2.0.7.

Template rostemplate by dieghernan, based on Bootstrapious.