Client tool for AEMET API

Client tool to get data and metadata from AEMET and convert json to tibble.

Usage

get_data_aemet(apidest, verbose = FALSE)

get_metadata_aemet(apidest, verbose = FALSE)

Source

https://opendata.aemet.es/dist/index.html

Arguments

apidest

Character string as destination URL. See https://opendata.aemet.es/dist/index.html.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

Value

A tibble (if possible) or the results of the query as provided by httr::content().

See also

Examples

# Run this example only if AEMET_API_KEY is detected

url <- "/api/valores/climatologicos/inventarioestaciones/todasestaciones"

get_data_aemet(url)
#> # A tibble: 291 × 7
#>    latitud provincia altitud indicativo nombre                 indsinop longitud
#>    <chr>   <chr>     <chr>   <chr>      <chr>                  <chr>    <chr>   
#>  1 413515N BARCELONA 74      0252D      ARENYS DE MAR          "08186"  023224E 
#>  2 411734N BARCELONA 4       0076       BARCELONA AEROPUERTO   "08181"  020412E 
#>  3 412506N BARCELONA 408     0200E      BARCELONA, FABRA       ""       020727E 
#>  4 412326N BARCELONA 6       0201D      BARCELONA              "08180"  021200E 
#>  5 414312N BARCELONA 291     0149X      MANRESA                "08174"  015025E 
#>  6 413125N BARCELONA 146     0229I      SABADELL AEROPUERTO    "08192"  020611E 
#>  7 413903N BARCELONA 40      0255B      SANTA SUSANNA          "08188"  024149E 
#>  8 415442N GIRONA    143     0367       GIRONA AEROPUERTO      "08184"  024548E 
#>  9 415848N GIRONA    95      0370B      GIRONA, ANTIC INSTITUT ""       024931E 
#> 10 420616N GIRONA    157     0372C      PORQUERES              "08120"  024549E 
#> # … with 281 more rows

# Metadata

get_metadata_aemet(url)
#> # A tibble: 7 × 7
#>   unidad_generadora              perio…¹ descr…² formato copyr…³ notaL…⁴ campo…⁵
#>   <chr>                          <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 Servicio del Banco de Datos N… 1 vez … Invent… applic… © AEME… https:… latitud
#> 2 Servicio del Banco de Datos N… 1 vez … Invent… applic… © AEME… https:… provin…
#> 3 Servicio del Banco de Datos N… 1 vez … Invent… applic… © AEME… https:… indica…
#> 4 Servicio del Banco de Datos N… 1 vez … Invent… applic… © AEME… https:… altitud
#> 5 Servicio del Banco de Datos N… 1 vez … Invent… applic… © AEME… https:… nombre 
#> 6 Servicio del Banco de Datos N… 1 vez … Invent… applic… © AEME… https:… indsin…
#> 7 Servicio del Banco de Datos N… 1 vez … Invent… applic… © AEME… https:… longit…
#> # … with 3 more variables: campos$descripcion <chr>, $tipo_datos <chr>,
#> #   $requerido <lgl>, and abbreviated variable names ¹​periodicidad,
#> #   ²​descripcion, ³​copyright, ⁴​notaLegal, ⁵​campos$id

# We can get data from any API endpoint

# Plain text

plain <- get_data_aemet("/api/prediccion/nacional/hoy")
#> 
#> Returning raw results. MIME type: text/plain

cat(plain)
#> AGENCIA ESTATAL DE METEOROLOGÍA
#> PREDICCIÓN GENERAL PARA ESPAÑA 
#> DÍA 21 DE ENERO DE 2023 A LAS 08:50 HORA OFICIAL
#> PREDICCIÓN VÁLIDA PARA EL SÁBADO 21
#> 
#> A.- FENÓMENOS SIGNIFICATIVOS
#> Nevadas en el Valle de Arán y Mallorca. Temperaturas mínimas
#> bajas en Baleares y el nordeste peninsular. Viento del norte
#> fuerte o con rachas muy fuertes en Pirineos, Ampurdán y norte de
#> Baleares.
#> 
#> B.- PREDICCIÓN
#> Predominarán los cielos poco nubosos o con intervalos nubosos en
#> la Península, salvo en el cuadrante noroeste, donde habrá cielos
#> nubosos o cubiertos. Precipitaciones débiles, en Galicia y mitad
#> occidental del área cantábrica durante la primera mitad del
#> día, sin descartarlas durante la segunda, ni tampoco en el resto
#> del área cantábrica. Con baja probabilidad, se espera alguna
#> precipitación débil a primeras horas en la meseta norte, algo
#> más probables, en los sistemas Central e Ibérico y, al final del
#> día, en los litorales de Andalucía oriental. Durante la segunda
#> mitad del día aumentará a nuboso en el Pirineo con chubascos.
#> Intervalos nubosos y probabilidad de chubascos ocasionales en
#> Baleares y Melilla. En Canarias nuboso con lluvias en el norte de
#> las islas de mayor relieve, y poco nuboso tendiendo a intervalos
#> nubosos en el resto. Las cotas de nieve se situarán entre 300 y
#> 600 m en Baleares, entre 200 y 500 m el tercio nordeste
#> peninsular, y en los 1500 m bajando a unos 1000 m en el resto de
#> la Península, pero solo en Pirineos y el norte de Mallorca se
#> esperan acumulaciones significativas.
#> 
#> Probables bancos de nieblas matinales en los principales sistemas
#> montañosos de la Península.
#> 
#> Temperaturas diurnas en ascenso en los tercios sur y oeste
#> peninsulares y en Canarias y en descenso en el resto. Mínimas en
#> descenso generalizado, salvo en Canarias, con pocos cambios.
#> Heladas en la mitad nordeste peninsular y en la meseta norte, así
#> como en zonas de montaña del resto de la Península. Serán
#> fuertes en Pirineos.
#> 
#> Viento de componente norte, salvo en el Cantábrico donde
#> predominara la componente este y en el Estrecho y Alborán donde
#> lo hará el poniente. El viento será en general flojo en zonas de
#> interior salvo en el valle del Ebro e Ibérica oriental, y
#> llegará a fuerte o con rachas muy fuertes en Pirineos, Ampurdán
#> y norte de Baleares. En Canarias, alisios con intervalos fuertes
#> entre islas.
#> 

# An image

image <- get_data_aemet("/api/satelites/producto/nvdi")
#> Not Found (HTTP 404).
#> 
#> Warning: With: /api/satelites/producto/nvdi: Error al obtener los datos

# Write and read
tmp <- tempfile(fileext = ".gif")

writeBin(image, tmp)
#> Error in writeBin(image, tmp): can only write vector objects

gganimate::gif_file(tmp)

Site built with pkgdown 2.0.7.

Template rostemplate by dieghernan, based on Bootstrapious.