Client tool for AEMET API

Client tool to get data and metadata from AEMET and convert json to tibble.

get_data_aemet(apidest, verbose = FALSE)

get_metadata_aemet(apidest, verbose = FALSE)

Arguments

apidest
Character string as destination URL. See https://opendata.aemet.es/dist/index.html.

verbose
Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

Source

https://opendata.aemet.es/dist/index.html

Value

A tibble or an empty tibble if no valid results from the API.

Examples

# Run this example only if AEMET_API_KEY is detected

url <- "/api/valores/climatologicos/inventarioestaciones/todasestaciones"

if (aemet_detect_api_key()) {
  get_data_aemet(url)
}

#> # A tibble: 291 x 7
#>    latitud provincia altitud indicativo nombre                 indsinop longitud
#>    <chr>   <chr>     <chr>   <chr>      <chr>                  <chr>    <chr>   
#>  1 413515N BARCELONA 74      0252D      ARENYS DE MAR          "08186"  023224E 
#>  2 411734N BARCELONA 4       0076       BARCELONA AEROPUERTO   "08181"  020412E 
#>  3 412506N BARCELONA 408     0200E      BARCELONA, FABRA       ""       020727E 
#>  4 412326N BARCELONA 6       0201D      BARCELONA              "08180"  021200E 
#>  5 414312N BARCELONA 291     0149X      MANRESA                "08174"  015025E 
#>  6 413125N BARCELONA 146     0229I      SABADELL AEROPUERTO    "08192"  020611E 
#>  7 413903N BARCELONA 40      0255B      SANTA SUSANNA          "08188"  024149E 
#>  8 415442N GIRONA    143     0367       GIRONA AEROPUERTO      "08184"  024548E 
#>  9 415848N GIRONA    95      0370B      GIRONA, ANTIC INSTITUT ""       024931E 
#> 10 420616N GIRONA    157     0372C      PORQUERES              "08120"  024549E 
#> # … with 281 more rows

# Metadata
if (aemet_detect_api_key()) {
  get_metadata_aemet(url)
}

#> # A tibble: 7 x 7
#>   unidad_generadora  periodicidad descripcion    formato copyright    notaLegal 
#>   <chr>              <chr>        <chr>          <chr>   <chr>        <chr>     
#> 1 Servicio del Banc… 1 vez al día Inventario de… applic… © AEMET. Au… http://ww…
#> 2 Servicio del Banc… 1 vez al día Inventario de… applic… © AEMET. Au… http://ww…
#> 3 Servicio del Banc… 1 vez al día Inventario de… applic… © AEMET. Au… http://ww…
#> 4 Servicio del Banc… 1 vez al día Inventario de… applic… © AEMET. Au… http://ww…
#> 5 Servicio del Banc… 1 vez al día Inventario de… applic… © AEMET. Au… http://ww…
#> 6 Servicio del Banc… 1 vez al día Inventario de… applic… © AEMET. Au… http://ww…
#> 7 Servicio del Banc… 1 vez al día Inventario de… applic… © AEMET. Au… http://ww…
#> # … with 1 more variable: campos <df[,4]>

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.