Client tool for AEMET API

Client tool to get data and metadata from AEMET and convert json to tibble.

Usage

get_data_aemet(apidest, verbose = FALSE)

get_metadata_aemet(apidest, verbose = FALSE)

Source

https://opendata.aemet.es/dist/index.html.

Arguments

apidest

Character string as destination URL. See https://opendata.aemet.es/dist/index.html.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

Value

A tibble (if possible) or the results of the query as provided by httr2::resp_body_raw() or httr2::resp_body_string().

See also

Examples

# Run this example only if AEMET_API_KEY is detected

url <- "/api/valores/climatologicos/inventarioestaciones/todasestaciones"

get_data_aemet(url)
#> # A tibble: 947 × 7
#>    latitud provincia     altitud indicativo nombre             indsinop longitud
#>    <chr>   <chr>         <chr>   <chr>      <chr>              <chr>    <chr>   
#>  1 394924N ILLES BALEARS 490     B013X      ESCORCA, LLUC      "08304"  025309E 
#>  2 394744N ILLES BALEARS 5       B051A      SÓLLER, PUERTO     "08316"  024129E 
#>  3 394121N ILLES BALEARS 60      B087X      BANYALBUFAR        ""       023046E 
#>  4 393445N ILLES BALEARS 52      B103B      ANDRATX - SANT ELM "99103"  022208E 
#>  5 393305N ILLES BALEARS 50      B158X      CALVIÀ, ES CAPDEL… ""       022759E 
#>  6 393319N ILLES BALEARS 3       B228       PALMA, PUERTO      "08301"  023731E 
#>  7 393832N ILLES BALEARS 95      B236C      PALMA, UNIVERSIDAD ""       023837E 
#>  8 394406N ILLES BALEARS 1030    B248       SIERRA DE ALFABIA… "08303"  024247E 
#>  9 393621N BALEARES      49      B275E      SON BONET, AEROPU… "08302"  024224E 
#> 10 393339N ILLES BALEARS 8       B278       PALMA DE MALLORCA… "08306"  024412E 
#> # ℹ 937 more rows


# Metadata

get_metadata_aemet(url)
#> # A tibble: 7 × 7
#>   unidad_generadora         periodicidad descripcion formato copyright notaLegal
#>   <chr>                     <chr>        <chr>       <chr>   <chr>     <chr>    
#> 1 Servicio del Banco de Da… 1 vez al día Inventario… applic… © AEMET.… https://…
#> 2 Servicio del Banco de Da… 1 vez al día Inventario… applic… © AEMET.… https://…
#> 3 Servicio del Banco de Da… 1 vez al día Inventario… applic… © AEMET.… https://…
#> 4 Servicio del Banco de Da… 1 vez al día Inventario… applic… © AEMET.… https://…
#> 5 Servicio del Banco de Da… 1 vez al día Inventario… applic… © AEMET.… https://…
#> 6 Servicio del Banco de Da… 1 vez al día Inventario… applic… © AEMET.… https://…
#> 7 Servicio del Banco de Da… 1 vez al día Inventario… applic… © AEMET.… https://…
#> # ℹ 1 more variable: campos <df[,4]>

# We can get data from any API endpoint

# Plain text

plain <- get_data_aemet("/api/prediccion/nacional/hoy")
#> 
#> Results are MIME type: text/plain
#> Returning data as string

cat(plain)
#> AGENCIA ESTATAL DE METEOROLOGÍA
#> PREDICCIÓN GENERAL PARA ESPAÑA 
#> DÍA 03 DE JULIO DE 2024 A LAS 07:32 HORA OFICIAL
#> PREDICCIÓN VÁLIDA PARA EL MIÉRCOLES 3
#> 
#> A.- FENÓMENOS SIGNIFICATIVOS
#> Temperaturas significativamente elevadas en los valles del
#> Guadalquivir, Guadiana, Genil y Segura. No se descartan rachas muy
#> fuertes de viento del noroeste en puntos del sur de Tarragona a
#> primeras horas.
#> 
#> B.- PREDICCIÓN
#> Continuará una situación estable en la mayor parte del país,
#> con cielos poco nubosos o despejados. Únicamente en el nordeste
#> de Cataluña y norte de Baleares persistirá una cierta
#> inestabilidad, con cielos nubosos y chubascos desplazándose de la
#> Península al archipiélago, y con tendencia a quedar poco nuboso
#> a lo largo de la tarde. Asimismo se prevén intervalos de
#> nubosidad baja, con probables brumas y bancos de niebla matinales,
#> en otras zonas de Cataluña, norte de Galicia, área cantábrica,
#> alto Ebro, Estrecho y litorales de Alborán, con tendencia a ir
#> levantando en general. Esta nubosidad será algo más persistente
#> en el Cantábrico, pudiendo dejar alguna llovizna en su extremo
#> oriental. También se dará nubosidad baja en el norte de
#> Canarias.
#> 
#> Posibilidad de calima en Melilla y este de Canarias.
#> 
#> Las temperaturas máximas aumentarán de forma prácticamente
#> generalizada, notablemente en áreas de Castilla y León y Ebro, y
#> exceptuando algunos descensos ligeros en Baleares, Cataluña,
#> Cantábrico y Galicia. Con ello se podrán superar los 35 grados
#> en interiores del cuadrante suroeste peninsular, extremo sureste y
#> depresiones del nordeste, incluso rebasar los 38 en los valles del
#> Guadiana, Guadalquivir, Genil y Segura. En general las mínimas
#> descenderán en Galicia, Cantábrico, Baleares y entorno del Ebro-
#> Ibérica, predominando los aumentos en el resto, y con pocos
#> cambios en Canarias.
#> 
#> Soplará la tramontana en el norte de Baleares, y con intervalos
#> de fuerte en Ampurdán. Alisios en Canarias también con
#> intervalos de fuerte en zonas expuestas, cierzo en el Ebro (no se
#> descartan rachas muy fuertes en puntos del sur de Tarragaona a
#> primeras horas), norte y nordeste en litorales de Galicia y
#> levante en el Estrecho. En el resto vientos flojos, variables en
#> el extremo sur peninsular, y en general con predominio de las
#> componentes norte y este rolando a norte y oeste en el resto.
#> 

# An image

image <- get_data_aemet("/api/mapasygraficos/analisis")
#> 
#> Results are MIME type: image/gif
#> Returning raw data

# Write and read
tmp <- tempfile(fileext = ".gif")

writeBin(image, tmp)

gganimate::gif_file(tmp)
#> Error in shell.exec(url): file association for 'C:\Users\RUNNER~1\AppData\Local\Temp\RtmpekMvFN\file18f065cd71ac.gif' not available or invalid

Site built with pkgdown 2.0.9.

Template rostemplate by dieghernan, based on Bootstrapious.