Client tool for AEMET API

Client tool to get data and metadata from AEMET and convert json to tibble.

Usage

get_data_aemet(apidest, verbose = FALSE)

get_metadata_aemet(apidest, verbose = FALSE)

Source

https://opendata.aemet.es/dist/index.html.

Arguments

apidest

Character string as destination URL. See https://opendata.aemet.es/dist/index.html.

verbose

Logical TRUE/FALSE. Provides information about the flow of information between the client and server.

Value

A tibble (if possible) or the results of the query as provided by httr2::resp_body_raw() or httr2::resp_body_string().

See also

Examples

# Run this example only if AEMET_API_KEY is detected

url <- "/api/valores/climatologicos/inventarioestaciones/todasestaciones"

get_data_aemet(url)
#> # A tibble: 947 × 7
#>    latitud provincia     altitud indicativo nombre             indsinop longitud
#>    <chr>   <chr>         <chr>   <chr>      <chr>              <chr>    <chr>   
#>  1 394924N ILLES BALEARS 490     B013X      ESCORCA, LLUC      "08304"  025309E 
#>  2 394744N ILLES BALEARS 5       B051A      SÓLLER, PUERTO     "08316"  024129E 
#>  3 394121N ILLES BALEARS 60      B087X      BANYALBUFAR        ""       023046E 
#>  4 393315N ILLES BALEARS 3       B228       PALMA, PUERTO      "08301"  023731E 
#>  5 393446N ILLES BALEARS 52      B103B      ANDRATX - SANT ELM ""       022208E 
#>  6 393305N ILLES BALEARS 50      B158X      CALVIÀ, ES CAPDEL… ""       022759E 
#>  7 393832N ILLES BALEARS 95      B236C      PALMA, UNIVERSITAT ""       023838E 
#>  8 394406N ILLES BALEARS 1030    B248       SIERRA DE ALFABIA… "08303"  024247E 
#>  9 393621N BALEARES      47      B275E      SON BONET, AEROPU… "08302"  024224E 
#> 10 393339N BALEARES      5       B278       PALMA DE MALLORCA… "08306"  024412E 
#> # ℹ 937 more rows


# Metadata

get_metadata_aemet(url)
#> # A tibble: 7 × 7
#>   unidad_generadora         periodicidad descripcion formato copyright notaLegal
#>   <chr>                     <chr>        <chr>       <chr>   <chr>     <chr>    
#> 1 Servicio del Banco de Da… 1 vez al día Inventario… applic… © AEMET.… https://…
#> 2 Servicio del Banco de Da… 1 vez al día Inventario… applic… © AEMET.… https://…
#> 3 Servicio del Banco de Da… 1 vez al día Inventario… applic… © AEMET.… https://…
#> 4 Servicio del Banco de Da… 1 vez al día Inventario… applic… © AEMET.… https://…
#> 5 Servicio del Banco de Da… 1 vez al día Inventario… applic… © AEMET.… https://…
#> 6 Servicio del Banco de Da… 1 vez al día Inventario… applic… © AEMET.… https://…
#> 7 Servicio del Banco de Da… 1 vez al día Inventario… applic… © AEMET.… https://…
#> # ℹ 1 more variable: campos <df[,4]>

# We can get data from any API endpoint

# Plain text

plain <- get_data_aemet("/api/prediccion/nacional/hoy")
#> 
#> Results are MIME type: text/plain
#> Returning data as string

cat(plain)
#> AGENCIA ESTATAL DE METEOROLOGÍA
#> PREDICCIÓN GENERAL PARA ESPAÑA 
#> DÍA 03 DE JUNIO DE 2025 A LAS 08:17 HORA OFICIAL
#> PREDICCIÓN VÁLIDA PARA EL MARTES 3
#> 
#> A.- FENÓMENOS SIGNIFICATIVOS
#> Chubascos y tormentas con probabilidad de ser fuertes en ambas
#> mesetas, áreas aledañas, este peninsular, sierras del nordeste
#> de Andalucía, medio Ebro e Ibérica, donde podrían ser muy
#> fuertes e incluso con granizo grande. Descenso notable de las
#> temperaturas máximas en la meseta Sur y nordeste de Andalucía.
#> 
#> B.- PREDICCIÓN
#> Se mantendrá una situación de inestabilidad en la Península con
#> la formación de una pequeña dana. Se esperan cielos nubosos o
#> cubiertos en el centro norte y noroeste peninsular con
#> precipitaciones en el entorno del Cantábrico oriental, nubosidad
#> baja matinal en los valles atlánticos tendiendo a levantar, e
#> intervalos de nubes medias y altas en el resto. A partir de
#> mediodía se desarrollará nubosidad de evolución con chubascos y
#> tormentas en amplias zonas de interior, siendo probable que
#> lleguen a fuertes y ocasionalmente vayan acompañadas de granizo y
#> rachas muy fuertes en amplias zonas de ambas mesetas y regiones
#> aledañas, así como en áreas del este peninsular, sierras del
#> nordeste de Andalucía, medio Ebro y especialmente en el entorno
#> del sistema Ibérico, Bajo Aragón y este de la meseta norte,
#> donde podrían ser muy fuertes e ir acompañadas de granizo
#> grande. En los archipiélagos predominarán nubes medias y altas
#> con posibilidad de alguna tormenta ocasional que podría dejar
#> algún chubasco aislado al suroeste de Canarias. También se
#> darán intervalos de nubes bajas en los nortes de Canarias.
#> 
#> Probables bancos de niebla matinales en montañas del extremo
#> norte peninsular y en Baleares. Calima débil en el este de la
#> Península, Melilla, Baleares y Canarias.
#> 
#> Las temperaturas máximas aumentarán en el extremo noroeste y
#> buena parte del tercio nordeste peninsular, predominando los
#> descensos en el resto, notables en la meseta sur y nordeste de
#> Andalucía. Las mínimas en general descenderán en la mitad sur
#> peninsular permaneciendo con pocos cambios en el resto.
#> 
#> Predominarán vientos de componente oeste en la Península y
#> Baleares, aunque empezarán de componente norte en el norte y de
#> levante en el área mediterránea oriental y tercio nordeste
#> peninsular. Soplarán más intensos en el Guadalquivir y en
#> litorales, con intervalos de poniente fuerte en el Estrecho y
#> Alborán. En Canarias soplará alisio también con intervalos de
#> fuerte y alguna racha muy fuerte en zonas expuestas.
#> 

# An image

image <- get_data_aemet("/api/mapasygraficos/analisis")
#> 
#> Results are MIME type: image/gif
#> Returning raw data

# Write and read
tmp <- tempfile(fileext = ".gif")

writeBin(image, tmp)

gganimate::gif_file(tmp)
#> Error in shell.exec(url): file association for 'C:\Users\RUNNER~1\AppData\Local\Temp\RtmpofM13G\filef641e7c6531.gif' not available or invalid

Site built with pkgdown 2.1.3.

Template rostemplate by dieghernan, based on bootstraptemple.com.