Warming stripes graph

Plot diferent "climate stripes" or "warming stripes" using ggplot2. This graphics are visual representations of the change in temperature as measured in each location over the past 70-100+ years. Each stripe represents the temperature in that station averaged over a year.

ggstripes(
  data,
  plot_type = c("background", "stripes", "trend", "animation"),
  plot_title = "",
  ...
)

Arguments

data
a data.frame with date(year) and temperature(temp) variables.

plot_type
plot type (whith labels, background, stripes with line trend and animation)

plot_title
character string to be used for the graph title.

...
further arguments passed to theme.

Value

a ggplot object

Note

"Warming stripes" charts are a conceptual idea of Professor Ed Hawkins (University of Reading) and are specifically designed to be as simple as possible and alert about risks of climate change. For more details see #ShowYourStripes: https://showyourstripes.info/.

See also

Examples

if (FALSE) {
ggstripes(data, plot_type = "background")
}

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.