Windrose (speed/direction) diagram of a station over a days period

Plot a windrose showing the wind speed and direction for a station over a days period.

windrose_days(
  station,
  apikey,
  start = "2000-12-31",
  end = "2000-12-31",
  n_directions = 8,
  n_speeds = 5,
  speed_cuts = NA,
  col_pal = "GnBu",
  calm_wind = 0,
  legend_title = "Wind Speed (m/s)"
)

Arguments

station
Character string as station identifier code (see aemet_stations).

apikey
Character string as personal API key (see https://opendata.aemet.es/centrodedescargas/obtencionAPIKey).

start
Character string as start date (format: %Y%m%d).

end
Character string as end date (format: %Y%m%d).

n_directions
Numeric value as number of direction bins to plot (default = 8).

n_speeds
Numeric value as number of equally spaced wind speed bins to plot (default = 5).

speed_cuts
Numeric vector containing the cut points for the wind speed intervals, or NA (default).

col_pal
Character string indicating the name of the brewer.pal.info colour palette to be used for plotting.

calm_wind
Numeric value as the upper limit for wind speed that is considered calm (default = 0).

legend_title
Character string to be used for the legend title.

Value

a plot.

See also

Examples

if (FALSE) {
windrose_days(station, apikey, start = "2000-01-01", end = "2000-12-31")
}